Heckman-t Model Fit Function

Fits a sample selection model based on the Student's t-distribution, extending the classical Heckman model to account for heavy-tailed error terms. The estimation is performed via Maximum Likelihood using the BFGS algorithm.

HeckmantS(selection, outcome, data = sys.frame(sys.parent()), df, start = NULL)

Arguments

selection

A formula specifying the selection equation.

outcome

A formula specifying the outcome equation.

data

A data frame containing the variables in the model.

df

Initial value for the degrees of freedom parameter of the t-distribution.

start

Optional numeric vector of initial parameter values.

Value

A list containing:

coefficients: Named vector of estimated model parameters.

value: Negative of the maximum log-likelihood.

loglik: Maximum log-likelihood.

counts: Number of gradient evaluations performed.

hessian: Hessian matrix at the optimum.

fisher_infotS: Approximate Fisher information matrix.

prop_sigmatS: Standard errors for the parameter estimates.

level: Levels of the selection variable.

nObs: Number of observations.

nParam: Number of model parameters.

N0: Number of censored (unobserved) observations.

N1: Number of uncensored (observed) observations.

NXS: Number of parameters in the selection equation.

NXO: Number of parameters in the outcome equation.

df: Degrees of freedom (observations minus parameters).

aic: Akaike Information Criterion.

bic: Bayesian Information Criterion.

initial.value: Initial parameter values used in the optimization.

Details

The function implements the Heckman sample selection model using the Student's t-distribution for the error terms, as proposed by Marchenko and Genton (2012) . This extension allows for robustness against outliers and heavy-tailed distributions. Initial parameter values can be specified by the user or default to standard starting values.

References

Yulia V Marchenko, Marc G Genton (2012). “A Heckman selection-t model.” Journal of the American Statistical Association, 107(497), 304–317.

Examples

data(MEPS2001)
attach(MEPS2001)
#> The following objects are masked from Mroz87:
#> 
#>     age, educ
#> The following objects are masked from MEPS2001 (pos = 4):
#> 
#>     age, age2, agefem, ambexp, blhisp, dambexp, dhospexp, educ,
#>     fairpoor, female, ffs, good, hospexp, income, ins, instype,
#>     instype_s1, lambexp, lnambx, totchr, vgood, year01
#> The following objects are masked from MEPS2001 (pos = 5):
#> 
#>     age, age2, agefem, ambexp, blhisp, dambexp, dhospexp, educ,
#>     fairpoor, female, ffs, good, hospexp, income, ins, instype,
#>     instype_s1, lambexp, lnambx, totchr, vgood, year01
#> The following objects are masked from MEPS2001 (pos = 6):
#> 
#>     age, age2, agefem, ambexp, blhisp, dambexp, dhospexp, educ,
#>     fairpoor, female, ffs, good, hospexp, income, ins, instype,
#>     instype_s1, lambexp, lnambx, totchr, vgood, year01
selectEq <- dambexp ~ age + female + educ + blhisp + totchr + ins + income
outcomeEq <- lnambx ~ age + female + educ + blhisp + totchr + ins
HeckmantS(selectEq, outcomeEq, data = MEPS2001, df = 12)
#> Start not provided using default start values.
#> Error in optim(start, loglik_tS, gradlik_tS, method = "BFGS", hessian = T,     control = list(fnscale = -1)): objective function in optim evaluates to length 18 not 1