U.S. Women's Labor Force Participation

The Mroz87 data frame contains data about 753 married women. These data are collected within the "Panel Study of Income Dynamics" (PSID). Of the 753 observations, the first 428 are for women with positive hours worked in 1975, while the remaining 325 observations are for women who did not work for pay in 1975. A more complete discussion of the data is found in Mroz (1987) . It also includes the following explanatory variables:

lfp: Dummy variable for labor-force participation.

hours: Wife's hours of work in 1975.

kids5: Number of children 5 years old or younger.

kids618: Number of children 6 to 18 years old.

Age: Wife's age.

Educ: Wife's educational attainment, in years.

wage: Wife's average hourly earnings, in 1975 dollars.

repwage: Wife's wage reported at the time of the 1976 interview.

hushrs: Husband's hours worked in 1975.

husage: Husband's age.

huseduc: Husband's educational attainment, in years.

huswage: Husband's wage, in 1975 dollars.

faminc: Family income, in 1975 dollars.

mtr: Marginal tax rate facing the wife.

motheduc: Wife's mother's educational attainment, in years.

fatheduc: Wife's father's educational attainment, in years.

unem: Unemployment rate in county of residence, in percentage points.

city: Dummy variable = 1 if live in large city, else 0.

exper: Actual years of wife's previous labor market experience.

nwifeinc: Non-wife income.

wifecoll: Dummy variable for wife's college attendance.

huscoll: Dummy variable for husband's college attendance.

Usage

Mroz87

Format

An object of class data.frame with 753 rows and 22 columns.

Source

PSID Staff, The Panel Study of Income Dynamics, Institute for Social ResearchPanel Study of Income Dynamics, University of Michigan, https://psidonline.isr.umich.edu/

References

Thomas A Mroz (1987). “The sensitivity of an empirical model of married women's hours of work to economic and statistical assumptions.” Econometrica: Journal of the Econometric Society, 765–799. Mikhail Zhelonkin, Marc G. Genton, Elvezio Ronchetti (2019). ssmrob: Robust Estimation and Inference in Sample Selection Models. R package version 0.7, https://CRAN.R-project.org/package=ssmrob. Ott Toomet, Arne Henningsen (2008). “Sample Selection Models in R: Package sampleSelection.” Journal of Statistical Software, 27(7). https://www.jstatsoft.org/article/view/v027i07. Jeffrey M Wooldridge (2016). Introductory econometrics: A modern approach. Nelson Education.

Examples

# Wooldridge(2016): page 247
data(Mroz87)
attach(Mroz87)
#> The following objects are masked from MEPS2001 (pos = 3):
#> 
#>     age, educ
#> The following objects are masked from MEPS2001 (pos = 4):
#> 
#>     age, educ
#> The following objects are masked from MEPS2001 (pos = 5):
#> 
#>     age, educ
#> The following objects are masked from Mroz87 (pos = 6):
#> 
#>     age, city, educ, exper, faminc, fatheduc, hours, husage, huscoll,
#>     huseduc, hushrs, huswage, kids5, kids618, lfp, motheduc, mtr,
#>     nwifeinc, repwage, unem, wage, wifecoll
#> The following objects are masked from MEPS2001 (pos = 7):
#> 
#>     age, educ
#> The following objects are masked from MEPS2001 (pos = 8):
#> 
#>     age, educ
#> The following objects are masked from MEPS2001 (pos = 9):
#> 
#>     age, educ
Mroz87$lwage <- ifelse(Mroz87$wage>0,log(Mroz87$wage), NA)
selectEq <- lfp ~ nwifeinc + educ + exper + I(exper^2) + age + kids5 + kids618
outcomeEq <- lwage ~ educ + exper + I(exper^2)
outcomeS <- cbind(educ, exper)
outcomeC <- 1
outcomeBS <- wage ~ educ + exper + I(exper^2)
outcomeBS <- wage ~ educ + exper + I(exper^2)
HeckmanCL(selectEq, outcomeEq, data = Mroz87)
#> Start not provided using default start values.
#> $coefficients
#>   (Intercept)      nwifeinc          educ         exper    I(exper^2) 
#>  0.2664106321 -0.0121314424  0.1313410931  0.1232775040 -0.0018861283 
#>           age         kids5       kids618   (Intercept)          educ 
#> -0.0528275356 -0.8673902142  0.0358743373 -0.5526897684  0.1083486083 
#>         exper    I(exper^2)         sigma           rho 
#>  0.0428375381 -0.0008374191  0.6633967328  0.0265969196 
#> 
#> $value
#> [1] -832.8851
#> 
#> $loglik
#> [1] 832.8851
#> 
#> $counts
#> gradient 
#>       24 
#> 
#> $hessian
#>                [,1]          [,2]          [,3]          [,4]          [,5]
#>  [1,] -3.839442e+02 -7.799208e+03 -4.687746e+03 -3.928679e+03 -6.317489e+04
#>  [2,] -7.799208e+03 -2.090244e+05 -9.831732e+04 -7.473165e+04 -1.123747e+06
#>  [3,] -4.687746e+03 -9.831732e+04 -5.908401e+04 -4.777595e+04 -7.626042e+05
#>  [4,] -3.928679e+03 -7.473165e+04 -4.777595e+04 -6.330539e+04 -1.316490e+06
#>  [5,] -6.317489e+04 -1.123747e+06 -7.626042e+05 -1.316490e+06 -3.249873e+07
#>  [6,] -1.638828e+04 -3.354115e+05 -1.994465e+05 -1.773593e+05 -3.001807e+06
#>  [7,] -9.096101e+01 -1.754755e+03 -1.183147e+03 -7.209243e+02 -8.118441e+03
#>  [8,] -5.333532e+02 -1.093641e+04 -6.438312e+03 -4.127609e+03 -4.833268e+04
#>  [9,]  8.208764e+00  1.618208e+02  1.015762e+02  9.432875e+01  1.601295e+03
#> [10,]  1.015763e+02  2.065593e+03  1.299801e+03  1.160877e+03  1.952808e+04
#> [11,]  9.432876e+01  1.745336e+03  1.160877e+03  1.602000e+03  3.424783e+04
#> [12,]  1.601998e+03  2.778802e+04  1.952177e+04  3.428438e+04  8.543210e+05
#> [13,]  1.585076e-02  6.617950e+00 -7.235105e-01  1.114984e-01  2.480667e-01
#> [14,]  3.776258e+00 -6.913255e+01  6.578445e+01  1.997464e+01  2.561709e+02
#>                [,6]         [,7]          [,8]          [,9]         [,10]
#>  [1,] -1.638828e+04   -90.961008 -5.333532e+02  8.208764e+00  1.015763e+02
#>  [2,] -3.354115e+05 -1754.755116 -1.093641e+04  1.618208e+02  2.065593e+03
#>  [3,] -1.994465e+05 -1183.147449 -6.438312e+03  1.015762e+02  1.299801e+03
#>  [4,] -1.773593e+05  -720.924341 -4.127609e+03  9.432875e+01  1.160877e+03
#>  [5,] -3.001807e+06 -8118.440880 -4.833268e+04  1.601295e+03  1.952808e+04
#>  [6,] -7.242453e+05 -3154.454270 -2.108989e+04  3.485296e+02  4.301366e+03
#>  [7,] -3.154454e+03  -117.336140 -1.449825e+02  1.562092e+00  2.081035e+01
#>  [8,] -2.108989e+04  -144.982547 -1.424248e+03  1.138161e+01  1.388205e+02
#>  [9,]  3.485296e+02     1.562092  1.138161e+01 -9.728459e+02 -1.231504e+04
#> [10,]  4.301366e+03    20.810348  1.388205e+02 -1.231504e+04 -1.609627e+05
#> [11,]  4.240663e+03    13.089725  9.634414e+01 -1.268286e+04 -1.602778e+05
#> [12,]  7.572317e+04   153.732286  1.191437e+03 -2.283330e+05 -2.869251e+06
#> [13,]  2.033597e+00    -0.156839  1.791387e-01 -1.373183e+01 -1.669250e+02
#> [14,]  1.319400e+02     5.604818  1.946302e+00 -3.421950e+02 -4.160733e+03
#>               [,11]         [,12]         [,13]         [,14]
#>  [1,]  9.432876e+01  1.601998e+03  1.585076e-02      3.776258
#>  [2,]  1.745336e+03  2.778802e+04  6.617950e+00    -69.132549
#>  [3,]  1.160877e+03  1.952177e+04 -7.235105e-01     65.784447
#>  [4,]  1.602000e+03  3.428438e+04  1.114984e-01     19.974639
#>  [5,]  3.424783e+04  8.543210e+05  2.480667e-01    256.170936
#>  [6,]  4.240663e+03  7.572317e+04  2.033597e+00    131.940023
#>  [7,]  1.308973e+01  1.537323e+02 -1.568390e-01      5.604818
#>  [8,]  9.634414e+01  1.191437e+03  1.791387e-01      1.946302
#>  [9,] -1.268286e+04 -2.283330e+05 -1.373183e+01   -342.195028
#> [10,] -1.602778e+05 -2.869251e+06 -1.669250e+02  -4160.732930
#> [11,] -2.283330e+05 -4.987945e+06 -1.381121e+02  -3444.955827
#> [12,] -4.987945e+06 -1.243670e+08 -2.191768e+03 -54751.515780
#> [13,] -1.381121e+02 -2.191768e+03 -1.945377e+03     10.352425
#> [14,] -3.444956e+03 -5.475152e+04  1.035243e+01   -196.676482
#> 
#> $fisher_infoHC
#>                [,1]          [,2]          [,3]          [,4]          [,5]
#>  [1,]  2.595002e-01  2.703821e-04 -7.785949e-03 -1.649960e-03  6.368964e-05
#>  [2,]  2.703821e-04  2.378811e-05 -4.327583e-05  3.842743e-06  1.171308e-07
#>  [3,] -7.785949e-03 -4.327583e-05  6.441045e-04 -1.916304e-05  3.187928e-07
#>  [4,] -1.649960e-03  3.842743e-06 -1.916304e-05  3.631673e-04 -1.079661e-05
#>  [5,]  6.368964e-05  1.171308e-07  3.187928e-07 -1.079661e-05  3.797793e-07
#>  [6,] -3.336502e-03 -6.448564e-06  2.056464e-05 -5.296394e-06 -5.027643e-07
#>  [7,] -1.894144e-02  2.122340e-05 -3.975477e-04 -1.555278e-05 -3.261251e-07
#>  [8,] -1.009487e-02 -9.874149e-06  1.161739e-04  6.906179e-05 -3.141298e-07
#>  [9,]  4.696721e-03  9.966198e-05 -4.625535e-04  6.966609e-05 -9.676452e-07
#> [10,] -1.656048e-04 -2.825190e-06  1.602367e-05 -2.630204e-06  4.466903e-08
#> [11,] -1.732155e-04 -4.114259e-06  1.663472e-05 -3.395578e-07 -3.776521e-08
#> [12,]  3.620012e-06  8.504671e-08 -3.418639e-07 -2.119195e-08  1.954783e-09
#> [13,] -2.545333e-05 -6.334061e-07  2.585690e-06 -5.819255e-07  9.526237e-09
#> [14,] -3.322817e-03 -8.619857e-05  3.529856e-04 -7.368420e-05  1.193213e-06
#>                [,6]          [,7]          [,8]          [,9]         [,10]
#>  [1,] -3.336502e-03 -1.894144e-02 -1.009487e-02  4.696721e-03 -1.656048e-04
#>  [2,] -6.448564e-06  2.122340e-05 -9.874149e-06  9.966198e-05 -2.825190e-06
#>  [3,]  2.056464e-05 -3.975477e-04  1.161739e-04 -4.625535e-04  1.602367e-05
#>  [4,] -5.296394e-06 -1.555278e-05  6.906179e-05  6.966609e-05 -2.630204e-06
#>  [5,] -5.027643e-07 -3.261251e-07 -3.141298e-07 -9.676452e-07  4.466903e-08
#>  [6,]  7.189041e-05  4.537887e-04  1.276203e-04 -3.587481e-05  1.038414e-06
#>  [7,]  4.537887e-04  1.407740e-02  6.287397e-04 -1.056772e-03  2.954358e-05
#>  [8,]  1.276203e-04  6.287397e-04  1.890095e-03  1.217019e-04 -3.313594e-06
#>  [9,] -3.587481e-05 -1.056772e-03  1.217019e-04  6.781730e-02 -3.257742e-03
#> [10,]  1.038414e-06  2.954358e-05 -3.313594e-06 -3.257742e-03  2.208478e-04
#> [11,]  1.420877e-06  4.302847e-05 -5.417408e-06 -2.144630e-03  2.221175e-05
#> [12,] -2.879819e-08 -8.763392e-07  1.143493e-07  4.764449e-05 -3.126919e-07
#> [13,]  2.480184e-07  7.161773e-06 -7.563166e-07 -2.326629e-04  6.520905e-06
#> [14,]  2.990519e-05  9.100192e-04 -1.012858e-04 -2.493369e-02  6.988269e-04
#>               [,11]         [,12]         [,13]         [,14]
#>  [1,] -1.732155e-04  3.620012e-06 -2.545333e-05 -3.322817e-03
#>  [2,] -4.114259e-06  8.504671e-08 -6.334061e-07 -8.619857e-05
#>  [3,]  1.663472e-05 -3.418639e-07  2.585690e-06  3.529856e-04
#>  [4,] -3.395578e-07 -2.119195e-08 -5.819255e-07 -7.368420e-05
#>  [5,] -3.776521e-08  1.954783e-09  9.526237e-09  1.193213e-06
#>  [6,]  1.420877e-06 -2.879819e-08  2.480184e-07  2.990519e-05
#>  [7,]  4.302847e-05 -8.763392e-07  7.161773e-06  9.100192e-04
#>  [8,] -5.417408e-06  1.143493e-07 -7.563166e-07 -1.012858e-04
#>  [9,] -2.144630e-03  4.764449e-05 -2.326629e-04 -2.493369e-02
#> [10,]  2.221175e-05 -3.126919e-07  6.520905e-06  6.988269e-04
#> [11,]  2.214533e-04 -5.912473e-06  9.653998e-06  1.034749e-03
#> [12,] -5.912473e-06  1.743274e-07 -1.993902e-07 -2.137691e-05
#> [13,]  9.653998e-06 -1.993902e-07  5.156225e-04  1.813210e-04
#> [14,]  1.034749e-03 -2.137691e-05  1.813210e-04  2.164214e-02
#> 
#> $prop_sigmaHC
#>  [1] 0.5094116689 0.0048773055 0.0253792147 0.0190569489 0.0006162624
#>  [6] 0.0084788210 0.1186482157 0.0434752258 0.2604175578 0.0148609485
#> [11] 0.0148813072 0.0004175254 0.0227073216 0.1471126914
#> 
#> $level
#> [1] "0" "1"
#> 
#> $nObs
#> [1] 753
#> 
#> $nParam
#> [1] 14
#> 
#> $N0
#> [1] 325
#> 
#> $N1
#> [1] 428
#> 
#> $NXS
#> [1] 8
#> 
#> $NXO
#> [1] 4
#> 
#> $df
#> [1] 739
#> 
#> $aic
#> [1] 1693.77
#> 
#> $bic
#> [1] 1758.507
#> 
#> $initial.value
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 1 0
#> 
#> attr(,"class")
#> [1] "HeckmanCL" "list"     
HeckmanBS(selectEq, outcomeBS, data = Mroz87)
#> Start not provided using default start values.
#> $coefficients
#>   (Intercept)      nwifeinc          educ         exper    I(exper^2) 
#>  1.0570371397 -0.0147174639  0.1681402251  0.1592805891 -0.0024341431 
#>           age         kids5       kids618   (Intercept)          educ 
#> -0.0684488944 -1.1249881577  0.0458271262 -0.4608723726  0.1091942007 
#>         exper    I(exper^2)         sigma           rho 
#>  0.0469295741 -0.0008817966  3.4287795888  0.1975753909 
#> 
#> $value
#> [1] -1375.13
#> 
#> $loglik
#> [1] 1375.13
#> 
#> $counts
#> gradient 
#>       30 
#> 
#> $hessian
#>                [,1]          [,2]          [,3]          [,4]          [,5]
#>  [1,]   -238.089526   -4865.46113   -2904.94957   -2509.63094 -4.285376e+04
#>  [2,]  -4865.461125 -135058.58619  -61206.53442  -46851.04065 -7.449115e+05
#>  [3,]  -2904.949567  -61206.53442  -36644.36525  -30752.34022 -5.242626e+05
#>  [4,]  -2509.630942  -46851.04065  -30752.34022  -41959.93998 -9.223070e+05
#>  [5,] -42853.759920 -744911.47161 -524262.55345 -922306.99131 -2.441847e+07
#>  [6,] -10165.162650 -209323.36040 -123562.60388 -112845.77639 -2.031354e+06
#>  [7,]    -56.135424   -1130.22532    -725.54469    -420.86838 -4.740164e+03
#>  [8,]   -332.774081   -6893.67154   -4016.82804   -2677.42235 -3.299349e+04
#>  [9,]     45.258081     887.19632     561.04399     522.44087  8.903304e+03
#> [10,]    561.046188   11319.62932    7186.75565    6455.14863  1.090357e+05
#> [11,]    522.447965    9523.72122    6455.20327    8960.36466  1.906964e+05
#> [12,]   9034.566439  154103.22054  110680.98318  194633.20621  4.810557e+06
#> [13,]      2.673421      47.15736      33.83136      32.04741  5.573144e+02
#> [14,]      5.951368      73.49388      73.79439      11.79108 -6.318479e+01
#>                [,6]          [,7]          [,8]          [,9]         [,10]
#>  [1,]   -10165.1626   -56.1354236   -332.774081  4.525808e+01      561.0462
#>  [2,]  -209323.3604 -1130.2253182  -6893.671541  8.871963e+02    11319.6293
#>  [3,]  -123562.6039  -725.5446895  -4016.828036  5.610440e+02     7186.7557
#>  [4,]  -112845.7764  -420.8683772  -2677.422349  5.224409e+02     6455.1486
#>  [5,] -2031353.7370 -4740.1641090 -32993.493305  8.903304e+03   109035.7214
#>  [6,]  -449438.4162 -1957.9630229 -13143.339674  1.920457e+03    23757.7416
#>  [7,]    -1957.9630   -72.3211219    -94.923624  8.620784e+00      113.7337
#>  [8,]   -13143.3397   -94.9236240   -903.079663  6.333145e+01      775.0956
#>  [9,]     1920.4568     8.6207840     63.331451 -8.655404e+02   -10978.2224
#> [10,]    23757.7416   113.7336615    775.095617 -1.097822e+04  -143580.8193
#> [11,]    23412.1455    68.4265495    544.086500 -1.058027e+04  -134080.9436
#> [12,]   425898.9306   791.1402012   6886.038466 -1.932829e+05 -2432636.9026
#> [13,]      111.8644     0.5844856      3.720192 -4.403043e+01     -563.5888
#> [14,]      244.1153     5.1535316      9.289011 -1.567654e+02    -1912.8357
#>               [,11]         [,12]         [,13]         [,14]
#>  [1,]  5.224480e+02  9.034566e+03  2.673421e+00      5.951368
#>  [2,]  9.523721e+03  1.541032e+05  4.715736e+01     73.493883
#>  [3,]  6.455203e+03  1.106810e+05  3.383136e+01     73.794392
#>  [4,]  8.960365e+03  1.946332e+05  3.204741e+01     11.791077
#>  [5,]  1.906964e+05  4.810557e+06  5.573144e+02    -63.184791
#>  [6,]  2.341215e+04  4.258989e+05  1.118644e+02    244.115346
#>  [7,]  6.842655e+01  7.911402e+02  5.844856e-01      5.153532
#>  [8,]  5.440865e+02  6.886038e+03  3.720192e+00      9.289011
#>  [9,] -1.058027e+04 -1.932829e+05 -4.403043e+01   -156.765384
#> [10,] -1.340809e+05 -2.432637e+06 -5.635888e+02  -1912.835713
#> [11,] -1.851500e+05 -4.234620e+06 -5.631112e+02  -1527.257847
#> [12,] -4.234620e+06 -1.146828e+08 -1.029431e+04 -24325.340095
#> [13,] -5.631112e+02 -1.029431e+04 -2.042694e+01    -14.250516
#> [14,] -1.527258e+03 -2.432534e+04 -1.425052e+01    -54.211140
#> 
#> $fisher_infoBS
#>                [,1]          [,2]          [,3]          [,4]          [,5]
#>  [1,]  4.163441e-01  3.724153e-04 -1.231624e-02 -2.476431e-03  9.399987e-05
#>  [2,]  3.724153e-04  3.341993e-05 -5.809281e-05  1.404664e-05 -1.155363e-08
#>  [3,] -1.231624e-02 -5.809281e-05  9.773919e-04 -6.113494e-05  7.231201e-07
#>  [4,] -2.476431e-03  1.404664e-05 -6.113494e-05  4.630365e-04 -1.273893e-05
#>  [5,]  9.399987e-05 -1.155363e-08  7.231201e-07 -1.273893e-05  4.289573e-07
#>  [6,] -5.419949e-03 -1.107697e-05  4.610984e-05  4.319014e-06 -1.038234e-06
#>  [7,] -3.061508e-02  3.477420e-06 -4.793417e-04  2.024706e-04 -5.488238e-06
#>  [8,] -1.538770e-02 -1.389126e-05  1.668103e-04  8.768746e-05 -3.381345e-08
#>  [9,]  1.278632e-02  5.691313e-05 -6.896354e-04  1.341887e-04 -2.004105e-06
#> [10,] -7.473870e-04 -2.286205e-06  5.064054e-05 -1.055001e-05  1.813434e-07
#> [11,] -2.061721e-04 -1.281190e-06  8.002825e-07  1.455185e-05 -4.018275e-07
#> [12,]  3.274495e-06  1.936042e-08  5.282465e-09 -4.202813e-07  1.343703e-08
#> [13,]  3.110742e-03  2.794015e-05 -6.134281e-05  1.538620e-04 -2.779568e-06
#> [14,] -8.237325e-03 -1.055960e-04  2.747453e-04 -4.443656e-04  8.233052e-06
#>                [,6]          [,7]          [,8]          [,9]         [,10]
#>  [1,] -5.419949e-03 -3.061508e-02 -1.538770e-02  1.278632e-02 -7.473870e-04
#>  [2,] -1.107697e-05  3.477420e-06 -1.389126e-05  5.691313e-05 -2.286205e-06
#>  [3,]  4.610984e-05 -4.793417e-04  1.668103e-04 -6.896354e-04  5.064054e-05
#>  [4,]  4.319014e-06  2.024706e-04  8.768746e-05  1.341887e-04 -1.055001e-05
#>  [5,] -1.038234e-06 -5.488238e-06 -3.381345e-08 -2.004105e-06  1.813434e-07
#>  [6,]  1.121040e-04  6.743743e-04  1.980055e-04 -1.254388e-04  4.965812e-06
#>  [7,]  6.743743e-04  2.243422e-02  7.860265e-04 -2.436785e-03  8.254492e-05
#>  [8,]  1.980055e-04  7.860265e-04  2.919264e-03  6.289278e-05 -9.125213e-07
#>  [9,] -1.254388e-04 -2.436785e-03  6.289278e-05  5.858705e-02 -3.398269e-03
#> [10,]  4.965812e-06  8.254492e-05 -9.125213e-07 -3.398269e-03  2.512164e-04
#> [11,]  2.870692e-06  6.602843e-05 -5.564217e-06 -1.213572e-03  1.714269e-06
#> [12,] -3.628120e-08 -9.328776e-07  1.206363e-07  2.345762e-05  1.638954e-07
#> [13,] -7.879084e-05 -1.639481e-03  2.111412e-05  9.015161e-03 -4.066434e-04
#> [14,]  2.013740e-04  4.469185e-03 -4.796675e-05 -2.843031e-02  9.592015e-04
#>               [,11]         [,12]         [,13]         [,14]
#>  [1,] -2.061721e-04  3.274495e-06  3.110742e-03 -8.237325e-03
#>  [2,] -1.281190e-06  1.936042e-08  2.794015e-05 -1.055960e-04
#>  [3,]  8.002825e-07  5.282465e-09 -6.134281e-05  2.747453e-04
#>  [4,]  1.455185e-05 -4.202813e-07  1.538620e-04 -4.443656e-04
#>  [5,] -4.018275e-07  1.343703e-08 -2.779568e-06  8.233052e-06
#>  [6,]  2.870692e-06 -3.628120e-08 -7.879084e-05  2.013740e-04
#>  [7,]  6.602843e-05 -9.328776e-07 -1.639481e-03  4.469185e-03
#>  [8,] -5.564217e-06  1.206363e-07  2.111412e-05 -4.796675e-05
#>  [9,] -1.213572e-03  2.345762e-05  9.015161e-03 -2.843031e-02
#> [10,]  1.714269e-06  1.638954e-07 -4.066434e-04  9.592015e-04
#> [11,]  1.615008e-04 -4.095036e-06 -4.075408e-04  8.422363e-04
#> [12,] -4.095036e-06  1.191500e-07  7.196361e-06 -1.355139e-05
#> [13,] -4.075408e-04  7.196361e-06  6.239561e-02 -2.004540e-02
#> [14,]  8.422363e-04 -1.355139e-05 -2.004540e-02  5.498052e-02
#> 
#> $prop_sigmaBS
#>  [1] 0.6452473261 0.0057809970 0.0312632684 0.0215182836 0.0006549483
#>  [6] 0.0105879179 0.1497805749 0.0540302175 0.2420476272 0.0158498075
#> [11] 0.0127082974 0.0003451812 0.2497911386 0.2253261143
#> 
#> $level
#> [1] "0" "1"
#> 
#> $nObs
#> [1] 753
#> 
#> $nParam
#> [1] 14
#> 
#> $N0
#> [1] 325
#> 
#> $N1
#> [1] 428
#> 
#> $NXS
#> [1] 8
#> 
#> $NXO
#> [1] 4
#> 
#> $df
#> [1] 739
#> 
#> $aic
#> [1] 2778.261
#> 
#> $bic
#> [1] 2842.997
#> 
#> $initial.value
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 1 0
#> 
#> attr(,"class")
#> [1] "HeckmanBS"
HeckmanSK(selectEq, outcomeEq, data = Mroz87, lambda = 1)
#> Start not provided using default start values.
#> $coefficients
#>   (Intercept)      nwifeinc          educ         exper    I(exper^2) 
#>  0.2663350384 -0.0121320482  0.1313398741  0.1232814605 -0.0018862461 
#>           age         kids5       kids618   (Intercept)          educ 
#> -0.0528283462 -0.8673963062  0.0358745816 -0.5553913893  0.1083498518 
#>         exper    I(exper^2)         sigma           rho        lambda 
#>  0.0428353751 -0.0008373951  0.6633956023  0.0265869564  0.0051435224 
#> 
#> $value
#> [1] -832.8851
#> 
#> $loglik
#> [1] 832.8851
#> 
#> $counts
#> gradient 
#>       31 
#> 
#> $hessian
#>               XS(Intercept)    XSnwifeinc        XSeduc       XSexper
#> XS(Intercept) -3.839434e+02 -7.799189e+03 -4.687737e+03 -3.928680e+03
#> XSnwifeinc    -7.799189e+03 -2.090237e+05 -9.831712e+04 -7.473170e+04
#> XSeduc        -4.687737e+03 -9.831712e+04 -5.908392e+04 -4.777597e+04
#> XSexper       -3.928680e+03 -7.473170e+04 -4.777597e+04 -6.330552e+04
#> XSI(exper^2)  -6.317501e+04 -1.123750e+06 -7.626059e+05 -1.316495e+06
#> XSage         -1.638825e+04 -3.354107e+05 -1.994462e+05 -1.773594e+05
#> XSkids5       -9.096065e+01 -1.754743e+03 -1.183143e+03 -7.209231e+02
#> XSkids618     -5.333519e+02 -1.093638e+04 -6.438298e+03 -4.127602e+03
#> XO(Intercept)  8.205715e+00  1.617611e+02  1.015385e+02  9.429380e+01
#> XOeduc         1.015386e+02  2.064831e+03  1.299319e+03  1.160448e+03
#> XOexper        9.429380e+01  1.744692e+03  1.160447e+03  1.601411e+03
#> XOI(exper^2)   1.601409e+03  2.777785e+04  1.951460e+04  3.427190e+04
#> sigma          4.936320e-02  7.275883e+00 -3.085468e-01  4.978053e-01
#> rho            2.202930e+00 -1.010889e+02  4.657854e+01  3.858208e+00
#>               -3.801678e+00 -7.983124e+01 -4.570128e+01 -3.343441e+01
#>                XSI(exper^2)         XSage       XSkids5     XSkids618
#> XS(Intercept) -6.317501e+04 -1.638825e+04   -90.9606464 -5.333519e+02
#> XSnwifeinc    -1.123750e+06 -3.354107e+05 -1754.7434109 -1.093638e+04
#> XSeduc        -7.626059e+05 -1.994462e+05 -1183.1428466 -6.438298e+03
#> XSexper       -1.316495e+06 -1.773594e+05  -720.9230946 -4.127602e+03
#> XSI(exper^2)  -3.249887e+07 -3.001815e+06 -8118.4314413 -4.833259e+04
#> XSage         -3.001815e+06 -7.242440e+05 -3154.4414820 -2.108984e+04
#> XSkids5       -8.118431e+03 -3.154441e+03  -117.3355831 -1.449819e+02
#> XSkids618     -4.833259e+04 -2.108984e+04  -144.9819147 -1.424244e+03
#> XO(Intercept)  1.600706e+03  3.484004e+02     1.5615110  1.137736e+01
#> XOeduc         1.952090e+04  4.299773e+03    20.8026166  1.387688e+02
#> XOexper        3.423534e+04  4.239097e+03    13.0848409  9.630804e+01
#> XOI(exper^2)   8.540126e+05  7.569543e+04   153.6748441  1.190990e+03
#> sigma          6.820745e+00  3.456450e+00    -0.1504175  2.254802e-01
#> rho           -3.287903e+00  6.476978e+01     5.2320221 -2.378987e-01
#>               -4.929680e+02 -1.632514e+02    -1.1030797 -5.292520e+00
#>               XO(Intercept)        XOeduc       XOexper  XOI(exper^2)
#> XS(Intercept)  8.205715e+00  1.015386e+02  9.429380e+01  1.601409e+03
#> XSnwifeinc     1.617611e+02  2.064831e+03  1.744692e+03  2.777785e+04
#> XSeduc         1.015385e+02  1.299319e+03  1.160447e+03  1.951460e+04
#> XSexper        9.429380e+01  1.160448e+03  1.601411e+03  3.427190e+04
#> XSI(exper^2)   1.600706e+03  1.952090e+04  3.423534e+04  8.540126e+05
#> XSage          3.484004e+02  4.299773e+03  4.239097e+03  7.569543e+04
#> XSkids5        1.561511e+00  2.080262e+01  1.308484e+01  1.536748e+02
#> XSkids618      1.137736e+01  1.387688e+02  9.630804e+01  1.190990e+03
#> XO(Intercept) -9.728654e+02 -1.231529e+04 -1.268311e+04 -2.283376e+05
#> XOeduc        -1.231529e+04 -1.609660e+05 -1.602810e+05 -2.869308e+06
#> XOexper       -1.268311e+04 -1.602810e+05 -2.283376e+05 -4.988045e+06
#> XOI(exper^2)  -2.283376e+05 -2.869308e+06 -4.988045e+06 -1.243695e+08
#> sigma         -1.768870e+01 -2.169920e+02 -1.899467e+02 -3.128105e+03
#> rho           -3.421641e+02 -4.160352e+03 -3.444595e+03 -5.474556e+04
#>               -5.146964e+02 -6.515510e+03 -6.710456e+03 -1.208136e+05
#>                       sigma           rho              
#> XS(Intercept)     0.0493632  2.202930e+00 -3.801678e+00
#> XSnwifeinc        7.2758833 -1.010889e+02 -7.983124e+01
#> XSeduc           -0.3085468  4.657854e+01 -4.570128e+01
#> XSexper           0.4978053  3.858208e+00 -3.343441e+01
#> XSI(exper^2)      6.8207446 -3.287903e+00 -4.929680e+02
#> XSage             3.4564504  6.476978e+01 -1.632514e+02
#> XSkids5          -0.1504175  5.232022e+00 -1.103080e+00
#> XSkids618         0.2254802 -2.378987e-01 -5.292520e+00
#> XO(Intercept)   -17.6887048 -3.421641e+02 -5.146964e+02
#> XOeduc         -216.9920471 -4.160352e+03 -6.515510e+03
#> XOexper        -189.9466569 -3.444595e+03 -6.710456e+03
#> XOI(exper^2)  -3128.1051293 -5.474556e+04 -1.208136e+05
#> sigma         -1945.5691531  8.950718e+00 -5.135650e+00
#> rho               8.9507181 -1.966555e+02 -1.810612e+02
#>                  -5.1356500 -1.810612e+02 -2.728446e+02
#> 
#> $fisher_infoSK
#>               XS(Intercept)    XSnwifeinc        XSeduc       XSexper
#> XS(Intercept)  2.607742e-01  2.707460e-04 -7.787429e-03 -1.649631e-03
#> XSnwifeinc     2.707460e-04  2.378824e-05 -4.327614e-05  3.842568e-06
#> XSeduc        -7.787429e-03 -4.327614e-05  6.441054e-04 -1.916256e-05
#> XSexper       -1.649631e-03  3.842568e-06 -1.916256e-05  3.631697e-04
#> XSI(exper^2)   6.368457e-05  1.171386e-07  3.187663e-07 -1.079667e-05
#> XSage         -3.336645e-03 -6.448504e-06  2.056440e-05 -5.296766e-06
#> XSkids5       -1.894548e-02  2.122565e-05 -3.975518e-04 -1.555675e-05
#> XSkids618     -1.009446e-02 -9.874134e-06  1.161743e-04  6.906245e-05
#> XO(Intercept)  3.587688e-02  9.996349e-05 -4.635215e-04  7.042070e-05
#> XOeduc        -1.685233e-04 -2.824948e-06  1.602191e-05 -2.629876e-06
#> XOexper       -1.777005e-04 -4.114119e-06  1.663565e-05 -3.406032e-07
#> XOI(exper^2)   3.713389e-06  8.504490e-08 -3.418884e-07 -2.115870e-08
#> sigma         -1.537409e-04 -6.340256e-07  2.587376e-06 -5.844552e-07
#> rho           -3.419609e-03 -8.619460e-05  3.529992e-04 -7.367988e-05
#>               -5.870822e-02 -5.764861e-07  1.831934e-06 -1.422437e-06
#>                XSI(exper^2)         XSage       XSkids5     XSkids618
#> XS(Intercept)  6.368457e-05 -3.336645e-03 -1.894548e-02 -1.009446e-02
#> XSnwifeinc     1.171386e-07 -6.448504e-06  2.122565e-05 -9.874134e-06
#> XSeduc         3.187663e-07  2.056440e-05 -3.975518e-04  1.161743e-04
#> XSexper       -1.079667e-05 -5.296766e-06 -1.555675e-05  6.906245e-05
#> XSI(exper^2)   3.797807e-07 -5.027593e-07 -3.260366e-07 -3.141368e-07
#> XSage         -5.027593e-07  7.189071e-05  4.537903e-04  1.276206e-04
#> XSkids5       -3.260366e-07  4.537903e-04  1.407745e-02  6.287437e-04
#> XSkids618     -3.141368e-07  1.276206e-04  6.287437e-04  1.890101e-03
#> XO(Intercept) -9.800692e-07 -3.615505e-05 -1.063092e-03  1.222271e-04
#> XOeduc         4.464710e-08  1.037913e-06  2.953418e-05 -3.313035e-06
#> XOexper       -3.775544e-08  1.420247e-06  4.301682e-05 -5.416606e-06
#> XOI(exper^2)   1.954121e-09 -2.878579e-08 -8.761149e-07  1.143331e-07
#> sigma          9.565656e-09  2.488932e-07  7.180079e-06 -7.576980e-07
#> rho            1.192653e-06  2.989167e-05  9.097625e-04 -1.012726e-04
#>                2.437355e-08  5.591331e-07  1.248090e-05 -1.022435e-06
#>               XO(Intercept)        XOeduc       XOexper  XOI(exper^2)
#> XS(Intercept)  3.587688e-02 -1.685233e-04 -1.777005e-04  3.713389e-06
#> XSnwifeinc     9.996349e-05 -2.824948e-06 -4.114119e-06  8.504490e-08
#> XSeduc        -4.635215e-04  1.602191e-05  1.663565e-05 -3.418884e-07
#> XSexper        7.042070e-05 -2.629876e-06 -3.406032e-07 -2.115870e-08
#> XSI(exper^2)  -9.800692e-07  4.464710e-08 -3.775544e-08  1.954121e-09
#> XSage         -3.615505e-05  1.037913e-06  1.420247e-06 -2.878579e-08
#> XSkids5       -1.063092e-03  2.953418e-05  4.301682e-05 -8.761149e-07
#> XSkids618      1.222271e-04 -3.313035e-06 -5.416606e-06  1.143331e-07
#> XO(Intercept)  8.430479e-01 -3.260040e-03 -2.151335e-03  4.780054e-05
#> XOeduc        -3.260040e-03  2.208403e-04  2.220832e-05 -3.126330e-07
#> XOexper       -2.151335e-03  2.220832e-05  2.214467e-04 -5.912323e-06
#> XOI(exper^2)   4.780054e-05 -3.126330e-07 -5.912323e-06  1.743236e-07
#> sigma         -3.416738e-03  6.527188e-06  9.670545e-06 -1.998369e-07
#> rho           -2.514657e-02  6.987428e-04  1.034672e-03 -2.137560e-05
#>               -1.464421e+00  4.591159e-06  1.283234e-05 -2.977488e-07
#>                       sigma           rho              
#> XS(Intercept) -1.537409e-04 -3.419609e-03 -5.870822e-02
#> XSnwifeinc    -6.340256e-07 -8.619460e-05 -5.764861e-07
#> XSeduc         2.587376e-06  3.529992e-04  1.831934e-06
#> XSexper       -5.844552e-07 -7.367988e-05 -1.422437e-06
#> XSI(exper^2)   9.565656e-09  1.192653e-06  2.437354e-08
#> XSage          2.488932e-07  2.989167e-05  5.591333e-07
#> XSkids5        7.180079e-06  9.097625e-04  1.248090e-05
#> XSkids618     -7.576980e-07 -1.012726e-04 -1.022435e-06
#> XO(Intercept) -3.416738e-03 -2.514657e-02 -1.464421e+00
#> XOeduc         6.527188e-06  6.987428e-04  4.591159e-06
#> XOexper        9.670545e-06  1.034672e-03  1.283234e-05
#> XOI(exper^2)  -1.998369e-07 -2.137560e-05 -2.977488e-07
#> sigma          5.286648e-04  1.820265e-04  6.011182e-03
#> rho            1.820265e-04  2.164031e-02  4.055325e-04
#>                6.011182e-03  4.055325e-04  2.766304e+00
#> 
#> $prop_sigmaSK
#> XS(Intercept)    XSnwifeinc        XSeduc       XSexper  XSI(exper^2) 
#>  0.5106605347  0.0048773185  0.0253792318  0.0190570106  0.0006162635 
#>         XSage       XSkids5     XSkids618 XO(Intercept)        XOeduc 
#>  0.0084788387  0.1186484409  0.0434752885  0.9181764023  0.0148606968 
#>       XOexper  XOI(exper^2)         sigma           rho               
#>  0.0148810852  0.0004175207  0.0229927113  0.1471064574  1.6632210973 
#> 
#> $level
#> [1] "0" "1"
#> 
#> $nObs
#> [1] 753
#> 
#> $nParam
#> [1] 15
#> 
#> $N0
#> [1] 325
#> 
#> $N1
#> [1] 428
#> 
#> $NXS
#> [1] 8
#> 
#> $NXO
#> [1] 4
#> 
#> $df
#> [1] 738
#> 
#> $aic
#> [1] 1695.77
#> 
#> $bic
#> [1] 1765.131
#> 
#> $initial.value
#> XS(Intercept)    XSnwifeinc        XSeduc       XSexper  XSI(exper^2) 
#>  0.2700735726 -0.0120236370  0.1309039692  0.1233471675 -0.0018870674 
#>         XSage       XSkids5     XSkids618 XO(Intercept)        XOeduc 
#> -0.0528524416 -0.8683246795  0.0360056106 -0.5781023058  0.1090654909 
#>       XOexper  XOI(exper^2)         sigma           rho               
#>  0.0438873010 -0.0008591134  0.6637994802  0.0486011464  1.0000000000 
#> 
#> attr(,"class")
#> [1] "HeckmanSK" "list"     
HeckmantS(selectEq, outcomeEq, data = Mroz87, df=5)
#> Start not provided using default start values.
#> Warning: NaNs produced
#> $coefficients
#>   (Intercept)      nwifeinc          educ         exper    I(exper^2) 
#>  0.4977082452 -0.0119469936  0.1395708592  0.1496697013 -0.0022848685 
#>           age         kids5       kids618   (Intercept)          educ 
#> -0.0642745309 -1.0438375049  0.0344009734 -0.2524096698  0.1084611568 
#>         exper    I(exper^2)         sigma           rho            df 
#>  0.0251433197 -0.0004567786  0.4437003789 -0.3825292972  3.0609579467 
#> 
#> $value
#> [1] -788.0048
#> 
#> $loglik
#> [1] 788.0048
#> 
#> $counts
#> gradient 
#>       32 
#> 
#> $hessian
#>                [,1]          [,2]          [,3]         [,4]          [,5]
#>  [1,]   -276.275818   -5676.81483   -3370.12234   -2745.7402 -4.240889e+04
#>  [2,]  -5676.814832 -152126.54313  -71580.60503  -54075.8089 -7.832581e+05
#>  [3,]  -3370.122343  -71580.60503  -42424.09272  -33187.9611 -5.066893e+05
#>  [4,]  -2745.740158  -54075.80892  -33187.96113  -42777.7992 -8.668430e+05
#>  [5,] -42408.894765 -783258.05367 -506689.31425 -866843.0121 -2.065490e+07
#>  [6,] -11808.275669 -244822.70979 -143595.64960 -124430.4535 -2.021754e+06
#>  [7,]    -64.187666   -1198.62774    -838.67344    -549.1864 -6.419652e+03
#>  [8,]   -377.057009   -7769.29217   -4533.19590   -2790.3522 -3.108295e+04
#>  [9,]   -123.623335   -2455.07694   -1535.27558   -1428.3276 -2.383327e+04
#> [10,]  -1535.218339  -31838.85066  -19620.53270  -17642.9066 -2.910668e+05
#> [11,]  -1428.202109  -27453.02481  -17642.18624  -24241.6777 -5.028777e+05
#> [12,] -22512.440106 -413046.27910 -275631.16502 -464109.9108 -1.062164e+07
#> [13,]     57.582326    1176.12252     661.47446     201.1791  3.729109e+02
#> [14,]    -35.865091    -926.27777    -418.14285    -352.8188 -6.064577e+03
#> [15,]     -5.291928     -90.15292     -66.74702     -86.3366 -1.691985e+03
#>                [,6]          [,7]          [,8]          [,9]         [,10]
#>  [1,]   -11808.2757   -64.1876664   -377.057009    -123.62334    -1535.2183
#>  [2,]  -244822.7098 -1198.6277365  -7769.292169   -2455.07694   -31838.8507
#>  [3,]  -143595.6496  -838.6734374  -4533.195897   -1535.27558   -19620.5327
#>  [4,]  -124430.4535  -549.1864442  -2790.352178   -1428.32758   -17642.9066
#>  [5,] -2021753.8704 -6419.6520368 -31082.953715  -23833.26954  -291066.8319
#>  [6,]  -522081.4231 -2216.2635367 -14984.718402   -5243.63079   -64983.0620
#>  [7,]    -2216.2635   -81.8073338    -91.091611     -22.74629     -298.3396
#>  [8,]   -14984.7184   -91.0916108   -980.859124    -169.12476    -2059.8488
#>  [9,]    -5243.6308   -22.7462901   -169.124758   -1603.03162   -20225.4936
#> [10,]   -64983.0620  -298.3396342  -2059.848755  -20225.49358  -263231.0182
#> [11,]   -64588.1076  -219.2209523  -1387.177673  -21948.33044  -276183.0343
#> [12,] -1058335.9439 -2846.0574366 -16542.302191 -365674.14434 -4578479.8001
#> [13,]     2486.4797    29.5101544     95.956523      50.48243      444.7916
#> [14,]    -1582.2799    -8.0900919    -51.787239    -402.76842    -4890.9141
#> [15,]     -223.1512     0.1009162     -7.401103      -8.90269     -118.6399
#>               [,11]         [,12]       [,13]         [,14]         [,15]
#>  [1,]    -1428.2021 -2.251244e+04    57.58233    -35.865091    -5.2919281
#>  [2,]   -27453.0248 -4.130463e+05  1176.12252   -926.277771   -90.1529231
#>  [3,]   -17642.1862 -2.756312e+05   661.47446   -418.142853   -66.7470230
#>  [4,]   -24241.6777 -4.641099e+05   201.17908   -352.818752   -86.3366006
#>  [5,]  -502877.6878 -1.062164e+07   372.91094  -6064.577277 -1691.9845110
#>  [6,]   -64588.1076 -1.058336e+06  2486.47972  -1582.279877  -223.1512438
#>  [7,]     -219.2210 -2.846057e+03    29.51015     -8.090092     0.1009162
#>  [8,]    -1387.1777 -1.654230e+04    95.95652    -51.787239    -7.4011026
#>  [9,]   -21948.3304 -3.656741e+05    50.48243   -402.768424    -8.9026903
#> [10,]  -276183.0343 -4.578480e+06   444.79164  -4890.914117  -118.6399116
#> [11,]  -399814.3522 -7.711897e+06  -242.54553  -4261.636590  -124.2787395
#> [12,] -7711897.1227 -1.549925e+08 -2234.56404 -68387.353472 -1971.2001493
#> [13,]     -242.5455 -2.234564e+03 -2308.28882   -101.065642    72.9783851
#> [14,]    -4261.6366 -6.838735e+04  -101.06564   -188.244428     5.1600284
#> [15,]     -124.2787 -1.971200e+03    72.97839      5.160028    -7.0578849
#> 
#> $fisher_infotS
#>                [,1]          [,2]          [,3]          [,4]          [,5]
#>  [1,]  3.786996e-01  4.322972e-04 -1.128645e-02 -2.474808e-03  9.661468e-05
#>  [2,]  4.322972e-04  3.462376e-05 -6.636531e-05 -1.367311e-05  6.974314e-07
#>  [3,] -1.128645e-02 -6.636531e-05  9.574880e-04  4.603178e-05 -1.149198e-06
#>  [4,] -2.474808e-03 -1.367311e-05  4.603178e-05  7.428990e-04 -2.257863e-05
#>  [5,]  9.661468e-05  6.974314e-07 -1.149198e-06 -2.257863e-05  7.843852e-07
#>  [6,] -4.868780e-03 -7.045165e-06  1.428861e-05 -4.092801e-05  1.261958e-07
#>  [7,] -2.954930e-02  8.452811e-05 -8.346963e-04 -6.392959e-04  1.503247e-05
#>  [8,] -1.557842e-02 -1.681645e-05  2.022282e-04  1.388290e-04 -1.469340e-06
#>  [9,] -6.695159e-03  2.900864e-05  6.168061e-04 -3.173096e-04  1.331534e-05
#> [10,]  6.637063e-04 -4.410778e-06 -4.047478e-05  1.180077e-05 -3.855561e-07
#> [11,] -3.909250e-04  1.095748e-05 -3.996310e-05  1.751092e-07 -2.517624e-07
#> [12,]  1.655182e-05 -4.462522e-07  1.594525e-06  1.322516e-07 -1.563354e-09
#> [13,] -2.942382e-04  6.011876e-06 -5.583166e-05 -1.180870e-04  2.619736e-06
#> [14,] -1.408777e-03 -7.643959e-05  2.218520e-04  4.426163e-04 -1.646557e-05
#> [15,] -1.028974e-02  7.840289e-05 -9.626445e-04 -1.676187e-03  2.504566e-05
#>                [,6]          [,7]          [,8]          [,9]         [,10]
#>  [1,] -4.868780e-03 -2.954930e-02 -1.557842e-02 -6.695159e-03  6.637063e-04
#>  [2,] -7.045165e-06  8.452811e-05 -1.681645e-05  2.900864e-05 -4.410778e-06
#>  [3,]  1.428861e-05 -8.346963e-04  2.022282e-04  6.168061e-04 -4.047478e-05
#>  [4,] -4.092801e-05 -6.392959e-04  1.388290e-04 -3.173096e-04  1.180077e-05
#>  [5,]  1.261958e-07  1.503247e-05 -1.469340e-06  1.331534e-05 -3.855561e-07
#>  [6,]  1.115865e-04  8.185775e-04  1.909400e-04 -1.064042e-05 -2.975964e-06
#>  [7,]  8.185775e-04  2.248188e-02  1.260752e-03 -8.846353e-05 -4.720901e-05
#>  [8,]  1.909400e-04  1.260752e-03  2.830430e-03 -2.889683e-05  3.245478e-06
#>  [9,] -1.064042e-05 -8.846353e-05 -2.889683e-05  1.690201e-02 -1.781778e-03
#> [10,] -2.975964e-06 -4.720901e-05  3.245478e-06 -1.781778e-03  1.380475e-04
#> [11,]  1.475032e-05  2.244353e-04 -6.663290e-07  1.532448e-03 -1.602321e-05
#> [12,] -5.946658e-07 -9.070731e-06  5.354065e-08 -6.275374e-05  7.835921e-07
#> [13,]  3.502328e-05  6.071743e-04  3.248429e-06 -1.224943e-04 -2.817531e-05
#> [14,] -4.093737e-05 -7.924534e-04 -8.869443e-05 -1.765807e-03  3.179617e-04
#> [15,]  6.612516e-04  1.148993e-02 -4.027869e-04 -3.512999e-03 -8.902616e-05
#>               [,11]         [,12]         [,13]         [,14]         [,15]
#>  [1,] -3.909250e-04  1.655182e-05 -2.942382e-04 -1.408777e-03 -1.028974e-02
#>  [2,]  1.095748e-05 -4.462522e-07  6.011876e-06 -7.643959e-05  7.840289e-05
#>  [3,] -3.996310e-05  1.594525e-06 -5.583166e-05  2.218520e-04 -9.626445e-04
#>  [4,]  1.751092e-07  1.322516e-07 -1.180870e-04  4.426163e-04 -1.676187e-03
#>  [5,] -2.517624e-07 -1.563354e-09  2.619736e-06 -1.646557e-05  2.504566e-05
#>  [6,]  1.475032e-05 -5.946658e-07  3.502328e-05 -4.093737e-05  6.612516e-04
#>  [7,]  2.244353e-04 -9.070731e-06  6.071743e-04 -7.924534e-04  1.148993e-02
#>  [8,] -6.663290e-07  5.354065e-08  3.248429e-06 -8.869443e-05 -4.027869e-04
#>  [9,]  1.532448e-03 -6.275374e-05 -1.224943e-04 -1.765807e-03 -3.512999e-03
#> [10,] -1.602321e-05  7.835921e-07 -2.817531e-05  3.179617e-04 -8.902616e-05
#> [11,] -1.890077e-04  6.757247e-06  9.899046e-05 -1.104632e-03  1.200455e-04
#> [12,]  6.757247e-06 -2.249153e-07 -3.834389e-06  4.534739e-05 -1.094519e-06
#> [13,]  9.899046e-05 -3.834389e-06  6.621652e-04 -4.724827e-05  7.154445e-03
#> [14,] -1.104632e-03  4.534739e-05 -4.724827e-05  9.924834e-03  1.027942e-02
#> [15,]  1.200455e-04 -1.094519e-06  7.154445e-03  1.027942e-02  2.372944e-01
#> 
#> $prop_sigmatS
#>  [1] 0.6153857199 0.0058841961 0.0309433026 0.0272561743 0.0008856552
#>  [6] 0.0105634530 0.1499395794 0.0532017873 0.1300077180 0.0117493600
#> [11]          NaN          NaN 0.0257325713 0.0996234624 0.4871287079
#> 
#> $level
#> [1] "0" "1"
#> 
#> $nObs
#> [1] 753
#> 
#> $nParam
#> [1] 15
#> 
#> $N0
#> [1] 325
#> 
#> $N1
#> [1] 428
#> 
#> $NXS
#> [1] 8
#> 
#> $NXO
#> [1] 4
#> 
#> $df
#> [1] 738
#> 
#> $aic
#> [1] 1606.01
#> 
#> $bic
#> [1] 1675.371
#> 
#> $initial.value
#>  [1] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [9] 0.000000 0.000000 0.000000 0.000000 1.000000 0.000000 1.609438
#> 
#> attr(,"class")
#> [1] "HeckmantS" "list"     
HeckmanGe(selectEq, outcomeEq, outcomeS, outcomeC, data = Mroz87)
#> Start not provided using default start values.
#> $coefficients
#>   (Intercept)      nwifeinc          educ         exper    I(exper^2) 
#>  0.2739173767 -0.0118990707  0.1304494057  0.1235010297 -0.0018908371 
#>           age         kids5       kids618   (Intercept)          educ 
#> -0.0528867949 -0.8691947007  0.0360640717 -0.4219353285  0.1067852531 
#>         exper    I(exper^2)    interceptS          educ         exper 
#>  0.0291210431 -0.0004640649 -0.3926769597  0.0164479294 -0.0194760692 
#>   correlation 
#> -0.0269224655 
#> 
#> $value
#> [1] -820.9621
#> 
#> $loglik
#> [1] 820.9621
#> 
#> $counts
#> gradient 
#>       47 
#> 
#> $hessian
#>               (Intercept)      nwifeinc          educ         exper
#> (Intercept) -3.839654e+02 -7.804754e+03 -4.688334e+03 -3.929132e+03
#> nwifeinc    -7.804754e+03 -2.093093e+05 -9.839640e+04 -7.481854e+04
#> educ        -4.688334e+03 -9.839640e+04 -5.910026e+04 -4.778848e+04
#> exper       -3.929132e+03 -7.481854e+04 -4.778848e+04 -6.331507e+04
#> I(exper^2)  -6.318025e+04 -1.125434e+06 -7.628105e+05 -1.317036e+06
#> age         -1.638861e+04 -3.356385e+05 -1.994564e+05 -1.773616e+05
#> kids5       -9.091108e+01 -1.754442e+03 -1.182622e+03 -7.204670e+02
#> kids618     -5.331050e+02 -1.093384e+04 -6.434387e+03 -4.123393e+03
#> (Intercept) -8.444849e+00 -1.630552e+02 -1.036297e+02 -1.083922e+02
#> educ        -1.036298e+02 -2.063743e+03 -1.316122e+03 -1.320241e+03
#> exper       -1.083922e+02 -1.952527e+03 -1.320241e+03 -2.010730e+03
#> I(exper^2)  -2.010710e+03 -3.387081e+04 -2.424479e+04 -4.615355e+04
#> interceptS   3.882897e-02 -3.503142e+00  9.175096e-01 -4.892159e-01
#> educ         9.174939e-01 -3.071382e+01  1.465422e+01 -3.416127e+00
#> exper       -4.894090e-01 -7.100699e+01 -3.417977e+00 -6.423045e+00
#> correlation -2.773793e+00 -2.270862e+02 -1.412066e+01 -4.116075e+01
#>                I(exper^2)           age         kids5       kids618
#> (Intercept) -6.318025e+04 -1.638861e+04   -90.9110825 -5.331050e+02
#> nwifeinc    -1.125434e+06 -3.356385e+05 -1754.4419570 -1.093384e+04
#> educ        -7.628105e+05 -1.994564e+05 -1182.6221850 -6.434387e+03
#> exper       -1.317036e+06 -1.773616e+05  -720.4669865 -4.123393e+03
#> I(exper^2)  -3.252652e+07 -3.001852e+06 -8112.4367438 -4.824409e+04
#> age         -3.001852e+06 -7.242136e+05 -3152.4822063 -2.108314e+04
#> kids5       -8.112437e+03 -3.152482e+03  -117.2294798 -1.448985e+02
#> kids618     -4.824409e+04 -2.108314e+04  -144.8985205 -1.423098e+03
#> (Intercept) -2.008579e+03 -3.639195e+02    -1.4777498 -1.098331e+01
#> educ        -2.424077e+04 -4.451919e+03   -19.5787947 -1.333025e+02
#> exper       -4.606261e+04 -4.963179e+03   -13.1943307 -1.001206e+02
#> I(exper^2)  -1.211294e+06 -9.681701e+04  -161.0898507 -1.308656e+03
#> interceptS  -5.843408e+00  1.037079e+00     0.1174208  8.036606e-03
#> educ        -7.746397e+01  3.679801e+01     1.5175161  1.852249e+00
#> exper       -3.996325e+02  3.399326e+00     0.5465556 -7.906283e-01
#> correlation -4.820515e+02 -1.461506e+02     2.6533345 -6.203897e+00
#>               (Intercept)          educ         exper    I(exper^2)
#> (Intercept) -8.444849e+00 -1.036298e+02 -1.083922e+02 -2.010710e+03
#> nwifeinc    -1.630552e+02 -2.063743e+03 -1.952527e+03 -3.387081e+04
#> educ        -1.036297e+02 -1.316122e+03 -1.320241e+03 -2.424479e+04
#> exper       -1.083922e+02 -1.320241e+03 -2.010730e+03 -4.615355e+04
#> I(exper^2)  -2.008579e+03 -2.424077e+04 -4.606261e+04 -1.211294e+06
#> age         -3.639195e+02 -4.451919e+03 -4.963179e+03 -9.681701e+04
#> kids5       -1.477750e+00 -1.957879e+01 -1.319433e+01 -1.610899e+02
#> kids618     -1.098331e+01 -1.333025e+02 -1.001206e+02 -1.308656e+03
#> (Intercept) -1.088566e+03 -1.356933e+04 -1.727951e+04 -3.610377e+05
#> educ        -1.356933e+04 -1.748077e+05 -2.144237e+05 -4.452771e+06
#> exper       -1.727951e+04 -2.144237e+05 -3.610377e+05 -8.854708e+06
#> I(exper^2)  -3.610377e+05 -4.452771e+06 -8.854708e+06 -2.412674e+08
#> interceptS   9.048302e+00  1.088678e+02  1.032518e+02  1.819865e+03
#> educ         1.088676e+02  5.572622e+02  6.088697e+02  2.926871e+03
#> exper        1.032283e+02  6.084789e+02  1.817611e+03 -1.009733e+05
#> correlation -3.389143e+02 -4.081688e+03 -3.842994e+03 -6.734960e+04
#>                interceptS          educ         exper   correlation
#> (Intercept)  3.882897e-02  9.174939e-01 -4.894090e-01     -2.773793
#> nwifeinc    -3.503142e+00 -3.071382e+01 -7.100699e+01   -227.086163
#> educ         9.175096e-01  1.465422e+01 -3.417977e+00    -14.120655
#> exper       -4.892159e-01 -3.416127e+00 -6.423045e+00    -41.160754
#> I(exper^2)  -5.843408e+00 -7.746397e+01 -3.996325e+02   -482.051494
#> age          1.037079e+00  3.679801e+01  3.399326e+00   -146.150558
#> kids5        1.174208e-01  1.517516e+00  5.465556e-01      2.653335
#> kids618      8.036606e-03  1.852249e+00 -7.906283e-01     -6.203897
#> (Intercept)  9.048302e+00  1.088676e+02  1.032283e+02   -338.914258
#> educ         1.088678e+02  5.572622e+02  6.084789e+02  -4081.688356
#> exper        1.032518e+02  6.088697e+02  1.817611e+03  -3842.993733
#> I(exper^2)   1.819865e+03  2.926871e+03 -1.009733e+05 -67349.596512
#> interceptS  -8.563577e+02 -1.084228e+04 -1.116296e+04     -6.436524
#> educ        -1.084228e+04 -1.411000e+05 -1.411161e+05    -49.942494
#> exper       -1.116296e+04 -1.411161e+05 -2.135174e+05   -109.084810
#> correlation -6.436524e+00 -4.994249e+01 -1.090848e+02   -177.592788
#> 
#> $fisher_infoHG
#>               (Intercept)      nwifeinc          educ         exper
#> (Intercept)  2.596852e-01  2.740501e-04 -7.799149e-03 -1.647124e-03
#> nwifeinc     2.740501e-04  2.417404e-05 -4.469009e-05  4.506088e-06
#> educ        -7.799149e-03 -4.469009e-05  6.477421e-04 -2.165409e-05
#> exper       -1.647124e-03  4.506088e-06 -2.165409e-05  3.642086e-04
#> I(exper^2)   6.357426e-05  9.837641e-08  3.840085e-07 -1.081754e-05
#> age         -3.339035e-03 -6.417016e-06  2.081667e-05 -5.184621e-06
#> kids5       -1.890489e-02  2.196893e-05 -3.985269e-04 -1.461155e-05
#> kids618     -1.010988e-02 -9.779847e-06  1.177635e-04  7.018070e-05
#> (Intercept)  3.711737e-03  1.975527e-04 -6.916745e-04  2.621638e-04
#> educ        -4.153239e-05 -4.197386e-06  1.155943e-05 -5.110089e-06
#> exper       -2.120698e-04 -9.836090e-06  3.717791e-05 -1.509960e-05
#> I(exper^2)   4.057940e-06  1.896040e-07 -7.188404e-07  3.160497e-07
#> interceptS   3.468362e-04  5.772876e-06 -3.008662e-05  7.227938e-06
#> educ        -1.442291e-05  5.227828e-08  4.358240e-07  1.483645e-07
#> exper       -9.382194e-06 -3.493784e-07  1.340851e-06 -5.240987e-07
#> correlation -3.839960e-03 -1.666011e-04  6.267829e-04 -2.077212e-04
#>                I(exper^2)           age         kids5       kids618
#> (Intercept)  6.357426e-05 -3.339035e-03 -1.890489e-02 -1.010988e-02
#> nwifeinc     9.837641e-08 -6.417016e-06  2.196893e-05 -9.779847e-06
#> educ         3.840085e-07  2.081667e-05 -3.985269e-04  1.177635e-04
#> exper       -1.081754e-05 -5.184621e-06 -1.461155e-05  7.018070e-05
#> I(exper^2)   3.800232e-07 -5.046748e-07 -3.350029e-07 -3.395072e-07
#> age         -5.046748e-07  7.186518e-05  4.528762e-04  1.273351e-04
#> kids5       -3.350029e-07  4.528762e-04  1.405940e-02  6.273789e-04
#> kids618     -3.395072e-07  1.273351e-04  6.273789e-04  1.890514e-03
#> (Intercept) -6.177557e-06 -2.378112e-05 -7.485446e-04  9.769039e-05
#> educ         1.204711e-07  4.591793e-07  1.633803e-05 -2.357439e-06
#> exper        3.672134e-07  1.269801e-06  3.767210e-05 -4.317841e-06
#> I(exper^2)  -8.038537e-09 -2.517979e-08 -7.380444e-07  7.972334e-08
#> interceptS  -1.633435e-07 -3.019617e-06 -3.515971e-05 -6.725053e-06
#> educ        -4.301803e-09  1.431707e-07  7.895676e-07  6.697172e-07
#> exper        1.260065e-08  7.724196e-08  1.429916e-06 -6.200017e-08
#> correlation  4.870474e-06  2.137289e-05  6.364769e-04 -8.471048e-05
#>               (Intercept)          educ         exper    I(exper^2)
#> (Intercept)  3.711737e-03 -4.153239e-05 -2.120698e-04  4.057940e-06
#> nwifeinc     1.975527e-04 -4.197386e-06 -9.836090e-06  1.896040e-07
#> educ        -6.916745e-04  1.155943e-05  3.717791e-05 -7.188404e-07
#> exper        2.621638e-04 -5.110089e-06 -1.509960e-05  3.160497e-07
#> I(exper^2)  -6.177557e-06  1.204711e-07  3.672134e-07 -8.038537e-09
#> age         -2.378112e-05  4.591793e-07  1.269801e-06 -2.517979e-08
#> kids5       -7.485446e-04  1.633803e-05  3.767210e-05 -7.380444e-07
#> kids618      9.769039e-05 -2.357439e-06 -4.317841e-06  7.972334e-08
#> (Intercept)  8.897362e-02 -3.339373e-03 -3.604509e-03  7.276434e-05
#> educ        -3.339373e-03  2.061887e-04  4.639392e-05 -7.646015e-07
#> exper       -3.604509e-03  4.639392e-05  2.858810e-04 -6.554602e-06
#> I(exper^2)   7.276434e-05 -7.646015e-07 -6.554602e-06  1.615399e-07
#> interceptS  -4.207052e-03  4.521794e-04 -1.510892e-04  4.275986e-06
#> educ         5.001052e-04 -3.617843e-05 -4.063601e-06  4.695353e-08
#> exper       -1.203412e-04 -4.687997e-07  1.342489e-05 -3.329028e-07
#> correlation -4.284959e-02  9.198748e-04  2.125416e-03 -4.097002e-05
#>                interceptS          educ         exper   correlation
#> (Intercept)  3.468362e-04 -1.442291e-05 -9.382194e-06 -3.839960e-03
#> nwifeinc     5.772876e-06  5.227828e-08 -3.493784e-07 -1.666011e-04
#> educ        -3.008662e-05  4.358240e-07  1.340851e-06  6.267829e-04
#> exper        7.227938e-06  1.483645e-07 -5.240987e-07 -2.077212e-04
#> I(exper^2)  -1.633435e-07 -4.301803e-09  1.260065e-08  4.870474e-06
#> age         -3.019617e-06  1.431707e-07  7.724196e-08  2.137289e-05
#> kids5       -3.515971e-05  7.895676e-07  1.429916e-06  6.364769e-04
#> kids618     -6.725053e-06  6.697172e-07 -6.200017e-08 -8.471048e-05
#> (Intercept) -4.207052e-03  5.001052e-04 -1.203412e-04 -4.284959e-02
#> educ         4.521794e-04 -3.617843e-05 -4.687997e-07  9.198748e-04
#> exper       -1.510892e-04 -4.063601e-06  1.342489e-05  2.125416e-03
#> I(exper^2)   4.275986e-06  4.695353e-08 -3.329028e-07 -4.097002e-05
#> interceptS   4.726022e-02 -3.417580e-03 -2.154750e-04 -1.344976e-03
#> educ        -3.417580e-03  2.688562e-04  1.069814e-06 -5.188102e-06
#> exper       -2.154750e-04  1.069814e-06  1.541583e-05  7.474757e-05
#> correlation -1.344976e-03 -5.188102e-06  7.474757e-05  3.606419e-02
#> 
#> $prop_sigmaHG
#>  (Intercept)     nwifeinc         educ        exper   I(exper^2)          age 
#> 0.5095931433 0.0049167100 0.0254507787 0.0190842497 0.0006164602 0.0084773335 
#>        kids5      kids618  (Intercept)         educ        exper   I(exper^2) 
#> 0.1185723494 0.0434800365 0.2982844532 0.0143592716 0.0169080160 0.0004019202 
#>   interceptS         educ        exper  correlation 
#> 0.2173941493 0.0163968345 0.0039263003 0.1899057290 
#> 
#> $level
#> [1] "0" "1"
#> 
#> $nObs
#> [1] 753
#> 
#> $nParam
#> [1] 16
#> 
#> $N0
#> [1] 325
#> 
#> $N1
#> [1] 428
#> 
#> $NXS
#> [1] 8
#> 
#> $NXO
#> [1] 4
#> 
#> $df
#> [1] 737
#> 
#> $aic
#> [1] 1673.924
#> 
#> $bic
#> [1] 1747.909
#> 
#> $initial.value
#> (Intercept)    nwifeinc        educ       exper  I(exper^2)         age 
#>           0           0           0           0           0           0 
#>       kids5     kids618 (Intercept)        educ       exper  I(exper^2) 
#>           0           0           0           0           0           0 
#>  interceptS        educ       exper correlation 
#>           1           1           1           0 
#> 
#> $NE
#> [1] 3
#> 
#> $NV
#> [1] 1
#> 
#> attr(,"class")
#> [1] "HeckmanGe" "list"