Two-Step Method for Parameter Estimation of the Heckman Model

Estimates the parameters of the classic Heckman model via the two-step method.

HCinitial(selection, outcome, data = sys.frame(sys.parent()))

Arguments

selection
Selection equation.

outcome
Primary Regression Equation.

data
Database.

Examples

data(MEPS2001)
attach(MEPS2001)
selectEq <- dambexp ~ age + female + educ + blhisp + totchr + ins + income
outcomeEq <- lnambx ~ age + female + educ + blhisp + totchr + ins
HCinitial(selectEq,outcomeEq, data = MEPS2001)

#> xs(Intercept)         xsage      xsfemale        xseduc      xsblhisp 
#>  -0.668643899   0.086814848   0.663505390   0.061883892  -0.365784312 
#>      xstotchr         xsins      xsincome xo(Intercept)         xoage 
#>   0.795747277   0.169106526   0.002677301   5.288927373   0.202466773 
#>      xofemale        xoeduc      xoblhisp      xototchr         xoins 
#>   0.292133967   0.012388871  -0.182865733   0.500633176  -0.046509658 
#>         sigma           rho 
#>   1.290541875  -0.359316986