Function for fit of the Generalized Heckman Model

Estimates the parameters of the Generalized Heckman model

#' @return

Returns a list with the following components.

Coefficients: Returns a numerical vector with the best estimated values of the model parameters;

Value: The value of function to be minimized (or maximized) corresponding to par.

loglik: Negative of value. Minimum (or maximum) of the likelihood function calculated from the estimated coefficients.

counts: Component of the Optim function. A two-element integer vector giving the number of calls to fn and gr respectively. This excludes those calls needed to compute the Hessian, if requested, and any calls to fn to compute a finite-difference approximation to the gradient.

hessian: Component of the Optim function, with pre-defined option hessian=TRUE. A symmetric matrix giving an estimate of the Hessian at the solution found. Note that this is the Hessian of the unconstrained problem even if the box constraints are active.

fisher_infoHG: Fisher information matrix

prop_sigmaHG: Square root of the Fisher information matrix diagonal

level: Selection variable levels

nObs: Numeric value representing the size of the database

nParam: Numerical value representing the number of model parameters

N0: Numerical value representing the number of unobserved entries

N1: Numerical value representing the number of complete entries

NXS: Numerical value representing the number of parameters of the selection model

NXO: Numerical value representing the number of parameters of the regression model

df: Numerical value that represents the difference between the size of the response vector of the selection equation and the number of model parameters

aic: Numerical value representing Akaike's information criterion.

bic: Numerical value representing Schwarz's Bayesian Criterion

initial.value: Numerical vector that represents the input values (Initial Values) used in the parameter estimation.

NE: Numerical value that represents the number of parameters related to the covariates fitted to the dispersion parameter considering the constant parameter.

NV: Numerical value that represents the number of parameters related to the covariates fitted to the correlation parameter considering the constant parameter.

HeckmanGe(
  selection,
  outcome,
  outcomeS,
  outcomeC,
  data = sys.frame(sys.parent()),
  start = NULL
)

Arguments

selection

Selection equation.

outcome

Primary Regression Equation.

outcomeS

Matrix with Covariates for fit of the Dispersion Parameter.

outcomeC

Matrix with Covariates for fit of the Correlation Parameter.

data

Database.

start

initial values.

Details

The HeckmanGe() function fits a generalization of the Heckman sample selection model, allowing sample selection bias and dispersion parameters to depend on covariates. For more information, see Bastos et al. (2022)

References

Fernando de Souza Bastos, Wagner Barreto-Souza, Marc G Genton (2022). “A Generalized Heckman Model With Varying Sample Selection Bias and Dispersion Parameters.” Statistica Sinica.

Examples

data(MEPS2001)
attach(MEPS2001)
#> The following objects are masked from MEPS2001 (pos = 3):
#> 
#>     age, age2, agefem, ambexp, blhisp, dambexp, dhospexp, educ,
#>     fairpoor, female, ffs, good, hospexp, income, ins, instype,
#>     instype_s1, lambexp, lnambx, totchr, vgood, year01
#> The following objects are masked from MEPS2001 (pos = 4):
#> 
#>     age, age2, agefem, ambexp, blhisp, dambexp, dhospexp, educ,
#>     fairpoor, female, ffs, good, hospexp, income, ins, instype,
#>     instype_s1, lambexp, lnambx, totchr, vgood, year01
#> The following objects are masked from MEPS2001 (pos = 5):
#> 
#>     age, age2, agefem, ambexp, blhisp, dambexp, dhospexp, educ,
#>     fairpoor, female, ffs, good, hospexp, income, ins, instype,
#>     instype_s1, lambexp, lnambx, totchr, vgood, year01
selectEq <- dambexp ~ age + female + educ + blhisp + totchr + ins + income
outcomeEq <- lnambx ~ age + female + educ + blhisp + totchr + ins
outcomeS <- cbind(age,female,totchr,ins)
outcomeC <- 1
HeckmanGe(selectEq, outcomeEq,outcomeS, outcomeC, data = MEPS2001)
#> $coefficients
#>   (Intercept)           age        female          educ        blhisp 
#> -0.5898951225  0.0938277500  0.6316215135  0.0551870772 -0.3299398157 
#>        totchr           ins        income   (Intercept)           age 
#>  0.7220902788  0.1649678564  0.0023791968  5.8622174739  0.1721671983 
#>        female          educ        blhisp        totchr           ins 
#>  0.1498487808  0.0009664388 -0.1346591903  0.4131454649 -0.1090664186 
#>    interceptS           age        female        totchr           ins 
#>  0.6132211459 -0.0286317500 -0.1297362447 -0.1185326958 -0.1122382329 
#>           rho 
#> -0.7423300084 
#> 
#> $value
#> [1] -5804.973
#> 
#> $loglik
#> [1] 5804.973
#> 
#> $counts
#> gradient 
#>       68 
#> 
#> $hessian
#>             (Intercept)          age       female        educ       blhisp
#> (Intercept)  -1503.8321   -5768.7018   -548.07918  -19505.227   -598.71849
#> age          -5768.7018  -24017.0022  -2063.43590  -74733.451  -2215.73400
#> female        -548.0792   -2063.4359   -548.07918   -7155.367   -244.31155
#> educ        -19505.2270  -74733.4505  -7155.36697 -263526.031  -7335.13335
#> blhisp        -598.7185   -2215.7340   -244.31155   -7335.133   -598.71849
#> totchr        -245.7402   -1022.7993    -87.52826   -3108.367    -93.92917
#> ins           -503.9784   -2074.0131   -166.68303   -6631.956   -184.28535
#> income      -50796.7973 -202912.7679 -16179.02525 -689148.044 -17652.25745
#> (Intercept)   -561.6397   -2199.9853   -234.53705   -7347.266   -212.40639
#> age          -2199.9829   -9336.7914   -884.92661  -28735.647   -806.15891
#> female        -234.5370    -884.9281   -234.53705   -3060.170   -101.53904
#> educ         -7347.1808  -28735.3252  -3060.11965 -100082.244  -2635.68481
#> blhisp        -212.4064    -806.1594   -101.53904   -2635.700   -212.40639
#> totchr        -116.5326    -493.4530    -43.83386   -1471.962    -44.45366
#> ins           -207.4637    -856.9628    -79.88931   -2756.548    -74.98259
#> interceptS    -341.6725   -1423.4934   -204.61841   -4652.111    -89.30248
#> age          -1423.5041   -6350.7808   -783.35768  -19333.859   -369.77722
#> female        -204.6184    -783.3528   -204.61841   -2743.927    -72.10447
#> totchr        -125.1580    -537.2119    -58.27067   -1621.238    -39.21385
#> ins           -142.4253    -607.9887    -77.03432   -1943.558    -34.33872
#> correlation   -220.3325    -824.7952    -57.48614   -2642.320   -117.42626
#>                   totchr          ins       income  (Intercept)          age
#> (Intercept)  -245.740179   -503.97836   -50796.797   -561.63966   -2199.9829
#> age         -1022.799345  -2074.01307  -202912.768  -2199.98528   -9336.7914
#> female        -87.528256   -166.68303   -16179.025   -234.53705    -884.9266
#> educ        -3108.367261  -6631.95635  -689148.044  -7347.26609  -28735.6472
#> blhisp        -93.929173   -184.28535   -17652.257   -212.40639    -806.1589
#> totchr       -287.255447    -72.37817    -8053.980   -116.53266    -493.4517
#> ins           -72.378167   -503.97836   -19383.173   -207.46367    -856.9619
#> income      -8053.979788 -19383.17315 -2551774.372 -19340.47419  -79001.3343
#> (Intercept)  -116.532661   -207.46367   -19340.474  -1840.21274   -7744.8229
#> age          -493.451699   -856.96186   -79001.334  -7744.82287  -34881.3022
#> female        -43.833862    -79.88931    -7113.604  -1091.74364   -4565.2780
#> educ        -1471.920115  -2756.51344  -264136.524 -24837.04372 -104562.5235
#> blhisp        -44.453670    -74.98259    -6429.910   -528.17077   -2122.5724
#> totchr       -139.079889    -39.13086    -3905.382  -1248.85608   -5699.7909
#> ins           -39.130859   -207.46367    -7981.754   -770.75985   -3326.3258
#> interceptS   -125.157938   -142.42526   -12807.227    147.64063     510.4387
#> age          -537.216509   -607.99305   -55388.258    510.43737    1715.4332
#> female        -58.270651    -77.03432    -6578.169     11.14747     109.4034
#> totchr       -154.693293    -40.02222    -4281.340    -20.81837    -147.6882
#> ins           -40.022211   -142.42526    -5670.425     36.75503     123.2460
#> correlation     3.565875    -65.73110    -6352.247   -306.00562   -1171.0634
#>                   female          educ      blhisp       totchr          ins
#> (Intercept)   -234.53705   -7347.18084  -212.40639   -116.53264   -207.46367
#> age           -884.92807  -28735.32515  -806.15937   -493.45301   -856.96284
#> female        -234.53705   -3060.11965  -101.53904    -43.83386    -79.88931
#> educ         -3060.16995 -100082.24416 -2635.70028  -1471.96153  -2756.54839
#> blhisp        -101.53904   -2635.68481  -212.40639    -44.45366    -74.98259
#> totchr         -43.83386   -1471.92012   -44.45367   -139.07989    -39.13086
#> ins            -79.88931   -2756.51344   -74.98259    -39.13086   -207.46367
#> income       -7113.60398 -264136.52379 -6429.90995  -3905.38189  -7981.75441
#> (Intercept)  -1091.74364  -24837.04372  -528.17077  -1248.85608   -770.75985
#> age          -4565.27796 -104562.52350 -2122.57235  -5699.79091  -3326.32579
#> female       -1091.74364  -14729.79698  -328.50041   -826.24715   -426.50902
#> educ        -14729.79698 -347169.03080 -6689.65977 -16793.90142 -10509.20479
#> blhisp        -328.50041   -6689.65977  -528.17077   -324.92495   -199.49091
#> totchr        -826.24715  -16793.90142  -324.92495  -2450.10398   -478.46064
#> ins           -426.50902  -10509.20479  -199.49091   -478.46064   -770.75985
#> interceptS      11.14747    1771.85437    87.46866    -20.81835     36.75503
#> age            109.40182    6050.45777   343.51814   -147.69404    123.24481
#> female          11.14747     -76.20079    51.11119    -55.74000    -27.73100
#> totchr         -55.73985    -576.61871    66.00214    -16.27583     12.36684
#> ins            -27.73100     465.33910    21.34999     12.36681     36.75503
#> correlation   -104.74185   -3842.03581  -140.11927    -28.69378   -105.18105
#>               interceptS          age       female       totchr         ins
#> (Intercept)   -341.67255   -1423.5041   -204.61841   -125.15801  -142.42526
#> age          -1423.49335   -6350.7808   -783.35276   -537.21192  -607.98873
#> female        -204.61841    -783.3577   -204.61841    -58.27067   -77.03432
#> educ         -4652.11052  -19333.8591  -2743.92745  -1621.23803 -1943.55793
#> blhisp         -89.30248    -369.7772    -72.10447    -39.21385   -34.33872
#> totchr        -125.15794    -537.2165    -58.27065   -154.69329   -40.02221
#> ins           -142.42526    -607.9931    -77.03432    -40.02222  -142.42526
#> income      -12807.22693  -55388.2579  -6578.16873  -4281.33961 -5670.42467
#> (Intercept)    147.64063     510.4374     11.14747    -20.81837    36.75503
#> age            510.43868    1715.4332    109.40336   -147.68820   123.24604
#> female          11.14747     109.4018     11.14747    -55.73985   -27.73100
#> educ          1771.85437    6050.4578    -76.20079   -576.61871   465.33910
#> blhisp          87.46866     343.5181     51.11119     66.00214    21.34999
#> totchr         -20.81835    -147.6940    -55.74000    -16.27583    12.36681
#> ins             36.75503     123.2448    -27.73100     12.36684    36.75503
#> interceptS   -5750.05697  -23676.8271  -3175.53782  -3180.92776 -2165.45053
#> age         -23676.82712 -104444.7087 -12867.98022 -14114.68871 -9235.30573
#> female       -3175.53782  -12867.9802  -3175.53782  -1811.26624 -1103.03260
#> totchr       -3180.92776  -14114.6887  -1811.26624  -5433.07711 -1098.86382
#> ins          -2165.45053   -9235.3057  -1103.03260  -1098.86382 -2165.45053
#> correlation   -663.76358   -2555.0891   -272.47777    -97.65759  -234.30989
#>              correlation
#> (Intercept)  -220.332545
#> age          -824.795238
#> female        -57.486143
#> educ        -2642.319751
#> blhisp       -117.426260
#> totchr          3.565875
#> ins           -65.731101
#> income      -6352.247347
#> (Intercept)  -306.005624
#> age         -1171.063388
#> female       -104.741851
#> educ        -3842.035807
#> blhisp       -140.119268
#> totchr        -28.693781
#> ins          -105.181053
#> interceptS   -663.763579
#> age         -2555.089090
#> female       -272.477774
#> totchr        -97.657594
#> ins          -234.309887
#> correlation  -332.668180
#> 
#> $fisher_infoHG
#>               (Intercept)           age        female          educ
#> (Intercept)  3.404667e-02 -2.663191e-03 -8.710766e-04 -1.680577e-03
#> age         -2.663191e-03  6.743981e-04 -1.724763e-05  2.612705e-05
#> female      -8.710766e-04 -1.724763e-05  3.459587e-03 -4.828686e-05
#> educ        -1.680577e-03  2.612705e-05 -4.828686e-05  1.290280e-04
#> blhisp      -3.542046e-03  1.388798e-04 -2.284468e-04  1.249531e-04
#> totchr      -6.839698e-04 -2.337102e-04  3.476923e-05  4.695407e-05
#> ins         -4.421637e-05 -2.503278e-04  1.139025e-04 -8.434861e-06
#> income       2.029671e-05 -5.673905e-06  9.772924e-06 -3.933455e-06
#> (Intercept) -9.433646e-03  8.722436e-04 -1.491933e-04  4.020986e-04
#> age          7.376612e-04 -2.111158e-04  2.926516e-05  5.592983e-06
#> female      -1.383000e-04 -2.183763e-06 -9.193475e-04  4.329440e-05
#> educ         4.832146e-04 -2.121147e-06  2.419808e-05 -3.571097e-05
#> blhisp       1.525645e-03 -5.367674e-05 -6.574146e-05 -6.327103e-05
#> totchr      -3.342743e-04  5.920622e-05  1.536940e-04  1.507605e-05
#> ins         -8.810100e-07  5.606009e-05  5.738360e-06  1.504525e-05
#> interceptS   1.507081e-04  1.698170e-04 -1.189579e-04 -5.435015e-05
#> age          1.090735e-04 -4.360488e-05  1.338565e-05  2.973334e-06
#> female      -1.790534e-04 -6.531310e-06 -1.834293e-04  1.866647e-05
#> totchr      -2.498899e-04  1.441531e-05  7.777825e-05  1.425410e-05
#> ins         -5.697504e-05  1.775463e-05  2.172696e-05  3.530790e-06
#> correlation -2.531815e-03 -5.594776e-05  5.807686e-04  1.584508e-04
#>                    blhisp        totchr           ins        income
#> (Intercept) -3.542046e-03 -6.839698e-04 -4.421637e-05  2.029671e-05
#> age          1.388798e-04 -2.337102e-04 -2.503278e-04 -5.673905e-06
#> female      -2.284468e-04  3.476923e-05  1.139025e-04  9.772924e-06
#> educ         1.249531e-04  4.695407e-05 -8.434861e-06 -3.933455e-06
#> blhisp       3.466291e-03 -2.990865e-05  5.739401e-05  2.735990e-06
#> totchr      -2.990865e-05  4.636480e-03  2.393327e-04  3.465785e-06
#> ins          5.739401e-05  2.393327e-04  3.611532e-03 -5.944945e-06
#> income       2.735990e-06  3.465785e-06 -5.944945e-06  1.460168e-06
#> (Intercept)  1.816117e-03 -1.704642e-03 -2.175980e-05 -7.773133e-06
#> age         -8.581025e-05  1.410340e-04  7.168821e-05  2.687171e-07
#> female      -1.184049e-04  5.193145e-04  1.485348e-05  1.566254e-06
#> educ        -6.048520e-05  4.034947e-05  8.268352e-06  2.179295e-07
#> blhisp      -1.156574e-03 -2.199227e-04 -9.962812e-05 -8.131882e-07
#> totchr      -8.486523e-05 -1.406400e-04 -1.498868e-05  1.022437e-06
#> ins         -1.099071e-04  7.254891e-05 -1.031591e-03  8.645675e-07
#> interceptS   2.872874e-04 -6.785002e-04 -6.232526e-05 -3.218340e-06
#> age         -2.409094e-05  6.612002e-05  2.538089e-05  1.817496e-07
#> female      -8.977562e-05  2.069415e-04  1.919981e-05  8.846878e-07
#> totchr      -6.679883e-05 -3.822607e-06  9.497653e-07  7.761531e-07
#> ins         -1.153324e-05  3.750596e-05 -2.416981e-04  3.926525e-07
#> correlation -6.291713e-04  1.775305e-03  2.000295e-04  9.514458e-06
#>               (Intercept)           age        female          educ
#> (Intercept) -9.433646e-03  7.376612e-04 -1.383000e-04  4.832146e-04
#> age          8.722436e-04 -2.111158e-04 -2.183763e-06 -2.121147e-06
#> female      -1.491933e-04  2.926516e-05 -9.193475e-04  2.419808e-05
#> educ         4.020986e-04  5.592983e-06  4.329440e-05 -3.571097e-05
#> blhisp       1.816117e-03 -8.581025e-05 -1.184049e-04 -6.048520e-05
#> totchr      -1.704642e-03  1.410340e-04  5.193145e-04  4.034947e-05
#> ins         -2.175980e-05  7.168821e-05  1.485348e-05  8.268352e-06
#> income      -7.773133e-06  2.687171e-07  1.566254e-06  2.179295e-07
#> (Intercept)  3.732429e-02 -2.570109e-03 -3.259217e-03 -1.530571e-03
#> age         -2.570109e-03  5.600184e-04  1.332498e-04  9.406302e-06
#> female      -3.259217e-03  1.332498e-04  3.071743e-03  2.734607e-05
#> educ        -1.530571e-03  9.406302e-06  2.734607e-05  1.034614e-04
#> blhisp      -2.448071e-03  9.917021e-05 -2.437918e-04  1.038718e-04
#> totchr      -6.885168e-04 -1.110911e-04  4.370182e-05  2.173698e-05
#> ins         -1.158898e-03 -7.745794e-05  2.783880e-04 -4.416614e-06
#> interceptS   3.291854e-03 -2.305635e-04 -9.326387e-04 -5.869681e-05
#> age         -3.082574e-04  2.780478e-05  9.522248e-05  5.412338e-06
#> female      -8.955818e-04  8.074057e-05  3.290377e-04  8.573168e-06
#> totchr      -4.948810e-04  1.936972e-05  1.103999e-04  7.288909e-06
#> ins         -2.202761e-04  2.826052e-06 -2.770124e-06  5.764341e-06
#> correlation -7.881285e-03  4.050032e-04  2.012913e-03  1.700493e-04
#>                    blhisp        totchr           ins    interceptS
#> (Intercept)  1.525645e-03 -3.342743e-04 -8.810100e-07  1.507081e-04
#> age         -5.367674e-05  5.920622e-05  5.606009e-05  1.698170e-04
#> female      -6.574146e-05  1.536940e-04  5.738360e-06 -1.189579e-04
#> educ        -6.327103e-05  1.507605e-05  1.504525e-05 -5.435015e-05
#> blhisp      -1.156574e-03 -8.486523e-05 -1.099071e-04  2.872874e-04
#> totchr      -2.199227e-04 -1.406400e-04  7.254891e-05 -6.785002e-04
#> ins         -9.962812e-05 -1.498868e-05 -1.031591e-03 -6.232526e-05
#> income      -8.131882e-07  1.022437e-06  8.645675e-07 -3.218340e-06
#> (Intercept) -2.448071e-03 -6.885168e-04 -1.158898e-03  3.291854e-03
#> age          9.917021e-05 -1.110911e-04 -7.745794e-05 -2.305635e-04
#> female      -2.437918e-04  4.370182e-05  2.783880e-04 -9.326387e-04
#> educ         1.038718e-04  2.173698e-05 -4.416614e-06 -5.869681e-05
#> blhisp       3.332859e-03 -4.485026e-05  6.854722e-05  2.170000e-04
#> totchr      -4.485026e-05  8.154044e-04  1.689070e-04 -3.065278e-04
#> ins          6.854722e-05  1.689070e-04  2.637205e-03 -2.543544e-04
#> interceptS   2.170000e-04 -3.065278e-04 -2.543544e-04  4.139431e-03
#> age         -1.558760e-05  1.307147e-05  1.039406e-05 -6.972033e-04
#> female      -1.067869e-05  6.013678e-05  2.844062e-05 -7.942145e-04
#> totchr      -2.008411e-05  9.854479e-05  7.866486e-05 -2.091137e-04
#> ins         -1.439930e-05  4.407145e-05  1.366879e-04 -2.378861e-04
#> correlation -8.525402e-04  9.969204e-04  6.777886e-04 -3.357800e-03
#>                       age        female        totchr           ins
#> (Intercept)  1.090735e-04 -1.790534e-04 -2.498899e-04 -5.697504e-05
#> age         -4.360488e-05 -6.531310e-06  1.441531e-05  1.775463e-05
#> female       1.338565e-05 -1.834293e-04  7.777825e-05  2.172696e-05
#> educ         2.973334e-06  1.866647e-05  1.425410e-05  3.530790e-06
#> blhisp      -2.409094e-05 -8.977562e-05 -6.679883e-05 -1.153324e-05
#> totchr       6.612002e-05  2.069415e-04 -3.822607e-06  3.750596e-05
#> ins          2.538089e-05  1.919981e-05  9.497653e-07 -2.416981e-04
#> income       1.817496e-07  8.846878e-07  7.761531e-07  3.926525e-07
#> (Intercept) -3.082574e-04 -8.955818e-04 -4.948810e-04 -2.202761e-04
#> age          2.780478e-05  8.074057e-05  1.936972e-05  2.826052e-06
#> female       9.522248e-05  3.290377e-04  1.103999e-04 -2.770124e-06
#> educ         5.412338e-06  8.573168e-06  7.288909e-06  5.764341e-06
#> blhisp      -1.558760e-05 -1.067869e-05 -2.008411e-05 -1.439930e-05
#> totchr       1.307147e-05  6.013678e-05  9.854479e-05  4.407145e-05
#> ins          1.039406e-05  2.844062e-05  7.866486e-05  1.366879e-04
#> interceptS  -6.972033e-04 -7.942145e-04 -2.091137e-04 -2.378861e-04
#> age          1.623137e-04  4.473250e-05 -2.839810e-05 -3.477090e-05
#> female       4.473250e-05  8.117343e-04  4.233432e-05  5.823735e-05
#> totchr      -2.839810e-05  4.233432e-05  3.418916e-04  4.749923e-05
#> ins         -3.477090e-05  5.823735e-05  4.749923e-05  7.792912e-04
#> correlation  2.466345e-04  8.815139e-04  7.216753e-04  2.238502e-04
#>               correlation
#> (Intercept) -2.531815e-03
#> age         -5.594776e-05
#> female       5.807686e-04
#> educ         1.584508e-04
#> blhisp      -6.291713e-04
#> totchr       1.775305e-03
#> ins          2.000295e-04
#> income       9.514458e-06
#> (Intercept) -7.881285e-03
#> age          4.050032e-04
#> female       2.012913e-03
#> educ         1.700493e-04
#> blhisp      -8.525402e-04
#> totchr       9.969204e-04
#> ins          6.777886e-04
#> interceptS  -3.357800e-03
#> age          2.466345e-04
#> female       8.815139e-04
#> totchr       7.216753e-04
#> ins          2.238502e-04
#> correlation  1.048150e-02
#> 
#> $prop_sigmaHG
#> (Intercept)         age      female        educ      blhisp      totchr 
#> 0.184517390 0.025969176 0.058818251 0.011359049 0.058875219 0.068091705 
#>         ins      income (Intercept)         age      female        educ 
#> 0.060096024 0.001208374 0.193194942 0.023664708 0.055423304 0.010171598 
#>      blhisp      totchr         ins  interceptS         age      female 
#> 0.057730917 0.028555287 0.051353725 0.064338410 0.012740240 0.028490952 
#>      totchr         ins correlation 
#> 0.018490312 0.027915788 0.102379210 
#> 
#> $level
#> [1] "0" "1"
#> 
#> $nObs
#> [1] 3328
#> 
#> $nParam
#> [1] 21
#> 
#> $N0
#> [1] 526
#> 
#> $N1
#> [1] 2802
#> 
#> $NXS
#> [1] 8
#> 
#> $NXO
#> [1] 7
#> 
#> $df
#> [1] 3307
#> 
#> $aic
#> [1] 11651.95
#> 
#> $bic
#> [1] 11780.26
#> 
#> $initial.value
#>  (Intercept)          age       female         educ       blhisp       totchr 
#> -0.668643899  0.086814848  0.663505390  0.061883892 -0.365784312  0.795747277 
#>          ins       income  (Intercept)          age       female         educ 
#>  0.169106526  0.002677301  5.288927373  0.202466773  0.292133967  0.012388871 
#>       blhisp       totchr          ins   interceptS          age       female 
#> -0.182865733  0.500633176 -0.046509658  1.290541875  1.000000000  1.000000000 
#>       totchr          ins  correlation 
#>  1.000000000  1.000000000 -0.359316986 
#> 
#> $NE
#> [1] 5
#> 
#> $NV
#> [1] 1
#> 
#> attr(,"class")
#> [1] "HeckmanGe" "list"