Skew-Normal Sample Selection Model Fit Function

Fits a sample selection model based on the Skew-Normal distribution using Maximum Likelihood Estimation (MLE). This model allows for asymmetry in the distribution of the outcome variable's error term, addressing potential skewness.

Usage

HeckmanSK(
  selection,
  outcome,
  data = sys.frame(sys.parent()),
  lambda,
  start = NULL
)

Arguments

selection

A formula specifying the selection equation.

outcome

A formula specifying the outcome equation.

data

A data frame containing the variables.

lambda

Initial start value for the skewness parameter (lambda).

start

Optional numeric vector of initial parameter values.

Value

A list containing:

coefficients: Named vector of estimated model parameters.

value: The (negative) log-likelihood at convergence.

loglik: The maximum log-likelihood.

counts: Number of gradient evaluations.

hessian: Hessian matrix at the optimum.

fisher_infoSK: Approximate Fisher information matrix.

prop_sigmaSK: Standard errors for the estimates.

level: Levels of the selection variable.

nObs: Number of observations.

nParam: Number of model parameters.

N0: Number of censored (unobserved) observations.

N1: Number of observed (uncensored) observations.

NXS: Number of covariates in the selection equation.

NXO: Number of covariates in the outcome equation.

df: Degrees of freedom (observations minus parameters).

aic: Akaike Information Criterion.

bic: Bayesian Information Criterion.

initial.value: Initial parameter values used.

Details

The function implements MLE for a sample selection model where the outcome equation's errors follow a Skew-Normal distribution, as proposed in Ogundimu and Hutton (2016) . The optimization is performed via the BFGS algorithm.

The results include estimates for:

Selection equation coefficients.

Outcome equation coefficients.

Standard deviation of the error term (sigma).

Correlation between the selection and outcome errors (rho).

Skewness parameter (lambda).

Robust standard errors from the Fisher information matrix.

References

Emmanuel O Ogundimu, Jane L Hutton (2016). “A Sample Selection Model with Skew-normal Distribution.” Scandinavian Journal of Statistics, 43(1), 172–190.

Examples

data("Mroz87")
attach(Mroz87)
#> The following objects are masked from MEPS2001 (pos = 3):
#> 
#>     age, educ
#> The following objects are masked from MEPS2001 (pos = 4):
#> 
#>     age, educ
#> The following objects are masked from MEPS2001 (pos = 5):
#> 
#>     age, educ
selectEq <- lfp ~ huswage + kids5 + mtr + fatheduc + educ + city
outcomeEq <- log(wage) ~ educ + city
HeckmanSK(selectEq, outcomeEq, data = Mroz87, lambda = -1.5)
#> Start not provided using default start values.
#> $coefficients
#> (Intercept)     huswage       kids5         mtr    fatheduc        educ 
#>  2.91663399 -0.09441332 -0.38394348 -5.13419159 -0.01560548  0.09497407 
#>        city (Intercept)        educ        city       sigma         rho 
#> -0.04294888  1.10265601  0.07462093  0.11951262  0.96402168 -0.79176540 
#>      lambda 
#> -1.58721959 
#> 
#> $value
#> [1] -874.6076
#> 
#> $loglik
#> [1] 874.6076
#> 
#> $counts
#> gradient 
#>       24 
#> 
#> $hessian
#>               XS(Intercept)   XShuswage     XSkids5       XSmtr  XSfatheduc
#> XS(Intercept)    -761.77574  -5697.8421  -174.90832  -518.26756  -6631.4515
#> XShuswage       -5697.84214 -54422.3225 -1333.97862 -3704.51609 -51488.3973
#> XSkids5          -174.90832  -1333.9786  -233.90702  -120.46979  -1668.4691
#> XSmtr            -518.26756  -3704.5161  -120.46979  -357.15300  -4464.9524
#> XSfatheduc      -6631.45155 -51488.3973 -1668.46915 -4464.95244 -67198.5411
#> XSeduc          -9260.18381 -71515.3530 -2284.77789 -6250.15361 -83189.7675
#> XScity           -486.77999  -4141.8544  -107.87602  -323.37060  -4444.6434
#> XO(Intercept)    -382.28475  -2819.8358   -72.27350  -258.60589  -3357.3094
#> XOeduc          -4694.37576 -35737.0251  -969.75186 -3151.07486 -42433.1353
#> XOcity           -242.68166  -2039.8719   -46.11674  -159.98226  -2255.1997
#> sigma             277.36072   1967.0547    58.75192   192.72719   2449.6704
#> rho               292.76673   1745.7145   -31.67112   196.12375   2831.9040
#>                    51.74583    391.4511    10.60955    34.92137    447.1092
#>                     XSeduc      XScity XO(Intercept)       XOeduc      XOcity
#> XS(Intercept)   -9260.1838  -486.77999    -382.28475   -4694.3758  -242.68166
#> XShuswage      -71515.3530 -4141.85437   -2819.83579  -35737.0251 -2039.87191
#> XSkids5         -2284.7779  -107.87602     -72.27350    -969.7519   -46.11674
#> XSmtr           -6250.1536  -323.37060    -258.60589   -3151.0749  -159.98226
#> XSfatheduc     -83189.7675 -4444.64340   -3357.30943  -42433.1353 -2255.19966
#> XSeduc        -116458.0927 -6066.00884   -4694.37753  -59709.0045 -3059.12322
#> XScity          -6066.0088  -486.77999    -242.68166   -3059.1222  -242.68166
#> XO(Intercept)   -4694.3775  -242.68166   -1088.86718  -13745.8008  -694.14923
#> XOeduc         -59709.0045 -3059.12216  -13745.80081 -179515.0804 -8967.53326
#> XOcity          -3059.1232  -242.68166    -694.14923   -8967.5333  -694.14923
#> sigma            3308.4967   174.48097     831.20519   10269.1431   526.05208
#> rho              3793.9617   176.32225    -380.92460   -4459.1313  -253.76078
#>                   634.4798    33.12212     -14.90175    -235.9640   -15.25712
#>                     sigma         rho           
#> XS(Intercept)   277.36072   292.76673   51.74583
#> XShuswage      1967.05468  1745.71451  391.45105
#> XSkids5          58.75192   -31.67112   10.60955
#> XSmtr           192.72719   196.12375   34.92137
#> XSfatheduc     2449.67038  2831.90396  447.10915
#> XSeduc         3308.49673  3793.96166  634.47977
#> XScity          174.48097   176.32225   33.12212
#> XO(Intercept)   831.20519  -380.92460  -14.90175
#> XOeduc        10269.14306 -4459.13132 -235.96401
#> XOcity          526.05208  -253.76078  -15.25712
#> sigma         -1554.50584  -136.12327  -66.88583
#> rho            -136.12327 -1085.90377  -86.96368
#>                 -66.88583   -86.96368  -25.84460
#> 
#> $fisher_infoSK
#>               XS(Intercept)     XShuswage       XSkids5         XSmtr
#> XS(Intercept)  0.5061297229 -8.326701e-03 -4.011100e-03 -0.5370902267
#> XShuswage     -0.0083267014  2.709059e-04  2.743346e-04  0.0097742363
#> XSkids5       -0.0040110996  2.743346e-04  7.225470e-03  0.0060220196
#> XSmtr         -0.5370902267  9.774236e-03  6.022020e-03  0.6309175200
#> XSfatheduc    -0.0003429192 -1.267632e-06 -8.050383e-05  0.0003487119
#> XSeduc        -0.0046992423 -5.474413e-05 -4.439382e-04  0.0010798064
#> XScity        -0.0008805195 -2.612733e-04  2.981502e-04 -0.0004364041
#> XO(Intercept) -0.0590692526  1.185193e-03  5.863147e-03  0.0482304740
#> XOeduc         0.0029627430 -4.714870e-05 -2.337244e-04 -0.0019002557
#> XOcity        -0.0020414493  6.423417e-05  1.887134e-04  0.0030483704
#> sigma         -0.0126676273  3.348246e-04  1.542273e-03  0.0147800159
#> rho            0.0178839595 -5.468719e-04 -2.747504e-03 -0.0206816802
#>                0.0182775613 -2.667198e-04 -7.864052e-04 -0.0212784754
#>                  XSfatheduc        XSeduc        XScity XO(Intercept)
#> XS(Intercept) -3.429192e-04 -4.699242e-03 -8.805195e-04 -0.0590692526
#> XShuswage     -1.267632e-06 -5.474413e-05 -2.612733e-04  0.0011851934
#> XSkids5       -8.050383e-05 -4.439382e-04  2.981502e-04  0.0058631469
#> XSmtr          3.487119e-04  1.079806e-03 -4.364041e-04  0.0482304740
#> XSfatheduc     1.324545e-04 -7.721167e-05 -7.500725e-05 -0.0001610668
#> XSeduc        -7.721167e-05  4.486120e-04 -1.006711e-04  0.0007570517
#> XScity        -7.500725e-05 -1.006711e-04  7.830044e-03  0.0006403469
#> XO(Intercept) -1.610668e-04  7.570517e-04  6.403469e-04  0.0542898258
#> XOeduc         1.038411e-05 -9.219724e-05  7.180994e-05 -0.0033538995
#> XOcity        -1.476012e-05  6.043411e-05 -2.509809e-03 -0.0002456014
#> sigma         -7.952192e-06 -1.996982e-04 -4.474209e-05  0.0062873272
#> rho            6.551474e-05  2.855431e-04 -5.316432e-07 -0.0089209228
#>               -1.609581e-04  5.129935e-04  6.526070e-04 -0.0029111309
#>                      XOeduc        XOcity         sigma           rho
#> XS(Intercept)  2.962743e-03 -2.041449e-03 -1.266763e-02  1.788396e-02
#> XShuswage     -4.714870e-05  6.423417e-05  3.348246e-04 -5.468719e-04
#> XSkids5       -2.337244e-04  1.887134e-04  1.542273e-03 -2.747504e-03
#> XSmtr         -1.900256e-03  3.048370e-03  1.478002e-02 -2.068168e-02
#> XSfatheduc     1.038411e-05 -1.476012e-05 -7.952192e-06  6.551474e-05
#> XSeduc        -9.219724e-05  6.043411e-05 -1.996982e-04  2.855431e-04
#> XScity         7.180994e-05 -2.509809e-03 -4.474209e-05 -5.316432e-07
#> XO(Intercept) -3.353899e-03 -2.456014e-04  6.287327e-03 -8.920923e-03
#> XOeduc         2.448577e-04 -1.889875e-04 -2.045384e-04  3.721174e-04
#> XOcity        -1.889875e-04  5.034548e-03  3.553576e-04 -3.573576e-04
#> sigma         -2.045384e-04  3.553576e-04  2.866223e-03 -2.260600e-03
#> rho            3.721174e-04 -3.573576e-04 -2.260600e-03  4.368764e-03
#>               -3.504491e-04 -1.728474e-03 -6.563844e-03 -2.990192e-04
#>                            
#> XS(Intercept)  0.0182775613
#> XShuswage     -0.0002667198
#> XSkids5       -0.0007864052
#> XSmtr         -0.0212784754
#> XSfatheduc    -0.0001609581
#> XSeduc         0.0005129935
#> XScity         0.0006526070
#> XO(Intercept) -0.0029111309
#> XOeduc        -0.0003504491
#> XOcity        -0.0017284737
#> sigma         -0.0065638440
#> rho           -0.0002990192
#>                0.0767113669
#> 
#> $prop_sigmaSK
#> XS(Intercept)     XShuswage       XSkids5         XSmtr    XSfatheduc 
#>    0.71142795    0.01645922    0.08500276    0.79430317    0.01150889 
#>        XSeduc        XScity XO(Intercept)        XOeduc        XOcity 
#>    0.02118046    0.08848754    0.23300177    0.01564793    0.07095455 
#>         sigma           rho               
#>    0.05353712    0.06609663    0.27696817 
#> 
#> $level
#> [1] "0" "1"
#> 
#> $nObs
#> [1] 753
#> 
#> $nParam
#> [1] 13
#> 
#> $N0
#> [1] 325
#> 
#> $N1
#> [1] 428
#> 
#> $NXS
#> [1] 7
#> 
#> $NXO
#> [1] 3
#> 
#> $df
#> [1] 740
#> 
#> $aic
#> [1] 1775.215
#> 
#> $bic
#> [1] 1835.328
#> 
#> $initial.value
#> XS(Intercept)     XShuswage       XSkids5         XSmtr    XSfatheduc 
#>   3.373565235  -0.116721911  -0.549776930  -5.328591099  -0.004906663 
#>        XSeduc        XScity XO(Intercept)        XOeduc        XOcity 
#>   0.122317681  -0.030769825   0.641598010   0.067208950   0.100413738 
#>         sigma           rho               
#>   0.851409314  -0.709828606  -1.500000000 
#> 
#> attr(,"class")
#> [1] "HeckmanSK" "list"