Heckman-t Model fit Function

Estimates the parameters of the Heckman-t model

HeckmantS(selection, outcome, data = sys.frame(sys.parent()), df, start = NULL)

Arguments

selection

Selection equation.

outcome

Primary Regression Equation.

data

Database.

df

Initial start to the degree of freedom.

start

initial values.

Value

Returns a list with the following components.

Coefficients: Returns a numerical vector with the best estimated values of the model parameters;

Value: The value of function to be minimized (or maximized) corresponding to par.

loglik: Negative of value. Minimum (or maximum) of the likelihood function calculated from the estimated coefficients.

counts: Component of the Optim function. A two-element integer vector giving the number of calls to fn and gr respectively. This excludes those calls needed to compute the Hessian, if requested, and any calls to fn to compute a finite-difference approximation to the gradient.

hessian: Component of the Optim function, with pre-defined option hessian=TRUE. A symmetric matrix giving an estimate of the Hessian at the solution found. Note that this is the Hessian of the unconstrained problem even if the box constraints are active.

fisher_infotS: Fisher information matrix

prop_sigmatS: Square root of the Fisher information matrix diagonal

level: Selection variable levels

nObs: Numeric value representing the size of the database

nParam: Numerical value representing the number of model parameters

N0: Numerical value representing the number of unobserved entries

N1: Numerical value representing the number of complete entries

NXS: Numerical value representing the number of parameters of the selection model

NXO: Numerical value representing the number of parameters of the regression model

df: Numerical value that represents the difference between the size of the response vector of the selection equation and the number of model parameters

aic: Numerical value representing Akaike's information criterion.

bic: Numerical value representing Schwarz's Bayesian Criterion

initial.value: Numerical vector that represents the input values (Initial Values) used in the parameter estimation.

Details

The HeckmantS() function fits the Sample Selection Model based on the Student's t distribution. For more information see Marchenko and Genton (2012)

References

Yulia V Marchenko, Marc G Genton (2012). “A Heckman selection-t model.” Journal of the American Statistical Association, 107(497), 304--317.

Examples

data(MEPS2001)
attach(MEPS2001)
#> The following objects are masked from Mroz87:
#> 
#>     age, educ
#> The following objects are masked from MEPS2001 (pos = 4):
#> 
#>     age, age2, agefem, ambexp, blhisp, dambexp, dhospexp, educ,
#>     fairpoor, female, ffs, good, hospexp, income, ins, instype,
#>     instype_s1, lambexp, lnambx, totchr, vgood, year01
#> The following objects are masked from MEPS2001 (pos = 5):
#> 
#>     age, age2, agefem, ambexp, blhisp, dambexp, dhospexp, educ,
#>     fairpoor, female, ffs, good, hospexp, income, ins, instype,
#>     instype_s1, lambexp, lnambx, totchr, vgood, year01
#> The following objects are masked from MEPS2001 (pos = 6):
#> 
#>     age, age2, agefem, ambexp, blhisp, dambexp, dhospexp, educ,
#>     fairpoor, female, ffs, good, hospexp, income, ins, instype,
#>     instype_s1, lambexp, lnambx, totchr, vgood, year01
#> The following objects are masked from MEPS2001 (pos = 7):
#> 
#>     age, age2, agefem, ambexp, blhisp, dambexp, dhospexp, educ,
#>     fairpoor, female, ffs, good, hospexp, income, ins, instype,
#>     instype_s1, lambexp, lnambx, totchr, vgood, year01
selectEq <- dambexp ~ age + female + educ + blhisp + totchr + ins + income
outcomeEq <- lnambx ~ age + female + educ + blhisp + totchr + ins
HeckmantS(selectEq, outcomeEq, data = MEPS2001, df=12)
#> $coefficients
#>  (Intercept)          age       female         educ       blhisp       totchr 
#> -0.747972299  0.098549379  0.724861923  0.064839075 -0.393568747  0.890081444 
#>          ins       income  (Intercept)          age       female         educ 
#>  0.180028786  0.002977546  5.205843434  0.206833445  0.306532774  0.017314842 
#>       blhisp       totchr          ins        sigma          rho           df 
#> -0.192966533  0.512711577 -0.052497676  1.194852538 -0.321991620 12.930320957 
#> 
#> $value
#> [1] -5822.076
#> 
#> $loglik
#> [1] 5822.076
#> 
#> $counts
#> gradient 
#>       25 
#> 
#> $hessian
#>               XS(Intercept)         XSage      XSfemale       XSeduc
#> XS(Intercept)  -1045.260400   -3974.90009   -359.382145  -13503.6213
#> XSage          -3974.900088  -16386.94634  -1349.195238  -51285.6828
#> XSfemale        -359.382145   -1349.19524   -359.382145   -4656.3861
#> XSeduc        -13503.621338  -51285.68283  -4656.386108 -181847.3957
#> XSblhisp        -430.062783   -1582.90667   -169.548343   -5232.0707
#> XStotchr        -145.845489    -602.78949    -46.157957   -1834.9704
#> XSins           -346.530157   -1411.30718   -103.891102   -4536.0179
#> XSincome      -35001.958269 -138422.09776 -10348.425492 -473344.6507
#> XO(Intercept)   -190.685144    -739.11549    -76.199025   -2506.1790
#> XOage           -739.115020   -3104.45087   -287.623844   -9698.6907
#> XOfemale         -76.199025    -287.62418    -76.199025    -995.5972
#> XOeduc         -2506.162799   -9698.62786   -995.586110  -34285.2474
#> XOblhisp         -70.736951    -266.03473    -32.914756    -874.7341
#> XOtotchr         -35.548417    -148.71612    -12.320213    -450.4024
#> XOins            -66.914409    -273.86663    -23.852604    -891.5231
#> sigma            -30.247935    -141.09352    -27.356370    -455.2462
#> rho              -84.838569    -323.60240    -16.516751   -1036.6100
#>                   -3.459894     -14.61434     -2.229851     -47.5671
#>                    XSblhisp     XStotchr         XSins      XSincome
#> XS(Intercept) -4.300628e+02  -145.845489   -346.530157   -35001.9583
#> XSage         -1.582907e+03  -602.789486  -1411.307178  -138422.0978
#> XSfemale      -1.695483e+02   -46.157957   -103.891102   -10348.4255
#> XSeduc        -5.232071e+03 -1834.970448  -4536.017855  -473344.6507
#> XSblhisp      -4.300628e+02   -59.288185   -132.943170   -12403.5646
#> XStotchr      -5.928818e+01  -166.272099    -43.194262    -4820.1821
#> XSins         -1.329432e+02   -43.194262   -346.530157   -13209.0794
#> XSincome      -1.240356e+04 -4820.182080 -13209.079436 -1752514.1432
#> XO(Intercept) -7.073695e+01   -35.548420    -66.914409    -6585.1940
#> XOage         -2.660347e+02  -148.715832   -273.866428   -26614.2859
#> XOfemale      -3.291476e+01   -12.320215    -23.852604    -2303.1325
#> XOeduc        -8.747321e+02  -450.393577   -891.515861   -90404.3523
#> XOblhisp      -7.073695e+01   -13.467160    -24.144987    -2086.7792
#> XOtotchr      -1.346716e+01   -41.520766    -11.166676    -1209.0065
#> XOins         -2.414499e+01   -11.166677    -66.914409    -2603.7106
#> sigma         -7.918516e-01   -20.376659    -15.302573    -1480.3096
#> rho           -4.373482e+01     2.012891    -22.489829    -2805.8010
#>               -7.613901e-01    -1.594650     -1.357224     -131.8663
#>               XO(Intercept)         XOage      XOfemale        XOeduc
#> XS(Intercept)   -190.685144    -739.11502    -76.199025   -2506.16280
#> XSage           -739.115490   -3104.45087   -287.624181   -9698.62786
#> XSfemale         -76.199025    -287.62384    -76.199025    -995.58611
#> XSeduc         -2506.178996   -9698.69072   -995.597219  -34285.24739
#> XSblhisp         -70.736951    -266.03467    -32.914756    -874.73208
#> XStotchr         -35.548420    -148.71583    -12.320215    -450.39358
#> XSins            -66.914409    -273.86643    -23.852604    -891.51586
#> XSincome       -6585.193995  -26614.28595  -2303.132538  -90404.35229
#> XO(Intercept)  -1795.394585   -7408.19155   -998.535041  -24387.53348
#> XOage          -7408.191550  -32811.92952  -4103.238261 -100670.18992
#> XOfemale        -998.535041   -4103.23826   -998.535041  -13532.30008
#> XOeduc        -24387.533481 -100670.18992 -13532.300077 -342593.40479
#> XOblhisp        -498.022405   -1963.47560   -294.903126   -6339.24350
#> XOtotchr       -1005.590008   -4506.16855   -620.839678  -13568.40189
#> XOins           -689.113680   -2935.27086   -353.798208   -9447.70462
#> sigma             37.501223     101.26251     -9.506342     375.35512
#> rho             -510.021232   -1946.30659   -175.482274   -6578.36149
#>                   -3.263302     -14.02572     -1.940081     -46.05649
#>                    XOblhisp      XOtotchr        XOins         sigma
#> XS(Intercept)   -70.7369511    -35.548417   -66.914409   -30.2479348
#> XSage          -266.0347342   -148.716123  -273.866632  -141.0935183
#> XSfemale        -32.9147560    -12.320213   -23.852604   -27.3563697
#> XSeduc         -874.7340620   -450.402400  -891.523143  -455.2461888
#> XSblhisp        -70.7369511    -13.467158   -24.144987    -0.7918516
#> XStotchr        -13.4671598    -41.520766   -11.166677   -20.3766593
#> XSins           -24.1449869    -11.166676   -66.914409   -15.3025729
#> XSincome      -2086.7791519  -1209.006464 -2603.710551 -1480.3096361
#> XO(Intercept)  -498.0224047  -1005.590008  -689.113680    37.5012233
#> XOage         -1963.4755989  -4506.168549 -2935.270860   101.2625080
#> XOfemale       -294.9031264   -620.839678  -353.798208    -9.5063422
#> XOeduc        -6339.2435005 -13568.401893 -9447.704621   375.3551179
#> XOblhisp       -498.0224047   -254.226359  -173.624649    30.7049535
#> XOtotchr       -254.2263587  -1757.408787  -349.662081   -35.0513770
#> XOins          -173.6246495   -349.662081  -689.113680    12.7433435
#> sigma            30.7049535    -35.051377    12.743344 -3856.0428696
#> rho            -209.5551923    -68.166024  -166.686242  -328.3871183
#>                  -0.8024373     -2.086292    -1.016658    19.8756062
#>                        rho             
#> XS(Intercept)   -84.838569   -3.4598943
#> XSage          -323.602398  -14.6143372
#> XSfemale        -16.516751   -2.2298513
#> XSeduc        -1036.609956  -47.5671032
#> XSblhisp        -43.734819   -0.7613901
#> XStotchr          2.012891   -1.5946503
#> XSins           -22.489829   -1.3572235
#> XSincome      -2805.801047 -131.8663089
#> XO(Intercept)  -510.021232   -3.2633017
#> XOage         -1946.306587  -14.0257243
#> XOfemale       -175.482274   -1.9400807
#> XOeduc        -6578.361490  -46.0564921
#> XOblhisp       -209.555192   -0.8024373
#> XOtotchr        -68.166024   -2.0862922
#> XOins          -166.686242   -1.0166581
#> sigma          -328.387118   19.8756062
#> rho            -338.718176    1.8040460
#>                   1.804046   -0.2659644
#> 
#> $fisher_infotS
#>               XS(Intercept)         XSage      XSfemale        XSeduc
#> XS(Intercept)  4.313922e-02 -3.500210e-03 -1.216844e-03 -2.126610e-03
#> XSage         -3.500210e-03  8.848053e-04 -1.106393e-05  3.884258e-05
#> XSfemale      -1.216844e-03 -1.106393e-05  4.700278e-03 -5.218862e-05
#> XSeduc        -2.126610e-03  3.884258e-05 -5.218862e-05  1.639796e-04
#> XSblhisp      -4.653816e-03  1.471184e-04 -4.579127e-04  1.706177e-04
#> XStotchr      -9.537949e-04 -2.578311e-04  4.708729e-04  5.855568e-05
#> XSins         -7.050974e-05 -3.081996e-04  2.686799e-04 -9.475320e-06
#> XSincome       2.931276e-05 -7.887692e-06  1.434611e-05 -5.703991e-06
#> XO(Intercept) -4.159339e-03  5.109932e-04 -3.590409e-04  1.614582e-04
#> XOage          3.186456e-04 -9.671472e-05  3.174769e-05  3.119701e-06
#> XOfemale      -4.148472e-05 -3.198028e-05 -3.400934e-04  2.109081e-05
#> XOeduc         2.084709e-04 -3.476959e-06  2.267254e-05 -1.503420e-05
#> XOblhisp       6.713354e-04  1.616945e-08 -4.969208e-05 -3.232422e-05
#> XOtotchr      -1.543929e-04  4.731693e-06  1.083817e-04  1.111920e-05
#> XOins          3.286943e-05  1.047594e-05 -2.338154e-05  6.365650e-06
#> sigma          3.154143e-04 -2.005610e-05 -2.634153e-04 -1.718690e-05
#> rho           -6.062956e-04 -1.720827e-04  3.256463e-04  6.371185e-05
#>                4.349877e-02 -6.408184e-03 -4.043335e-02 -1.884692e-03
#>                    XSblhisp      XStotchr         XSins      XSincome
#> XS(Intercept) -4.653816e-03 -9.537949e-04 -7.050974e-05  2.931276e-05
#> XSage          1.471184e-04 -2.578311e-04 -3.081996e-04 -7.887692e-06
#> XSfemale      -4.579127e-04  4.708729e-04  2.686799e-04  1.434611e-05
#> XSeduc         1.706177e-04  5.855568e-05 -9.475320e-06 -5.703991e-06
#> XSblhisp       4.426094e-03 -1.456496e-04 -9.580534e-06  5.654848e-06
#> XStotchr      -1.456496e-04  7.612860e-03  3.498068e-04  2.733373e-06
#> XSins         -9.580534e-06  3.498068e-04  4.625314e-03 -8.482626e-06
#> XSincome       5.654848e-06  2.733373e-06 -8.482626e-06  2.092589e-06
#> XO(Intercept)  8.093827e-04 -1.410652e-03 -2.948364e-04  2.034895e-06
#> XOage         -3.703410e-05  9.285434e-05  4.336198e-05 -1.221543e-07
#> XOfemale      -3.754290e-05  3.428463e-04  5.360484e-05 -6.493827e-07
#> XOeduc        -3.100959e-05  4.596481e-05  1.238556e-05 -1.606099e-08
#> XOblhisp      -5.009667e-04 -1.926896e-04 -7.706378e-05  3.696687e-07
#> XOtotchr      -4.966990e-05 -5.629332e-06  3.607011e-05 -2.625975e-07
#> XOins         -2.672521e-05  2.994162e-05 -4.425393e-04 -2.786759e-07
#> sigma          1.247369e-04 -4.408124e-04 -7.131725e-05 -7.243034e-07
#> rho           -1.635925e-04  1.036656e-03  3.112422e-04 -2.406242e-06
#>                1.911024e-02 -5.966139e-02 -6.955095e-03 -1.542936e-04
#>               XO(Intercept)         XOage      XOfemale        XOeduc
#> XS(Intercept) -4.159339e-03  3.186456e-04 -4.148472e-05  2.084709e-04
#> XSage          5.109932e-04 -9.671472e-05 -3.198028e-05 -3.476959e-06
#> XSfemale      -3.590409e-04  3.174769e-05 -3.400934e-04  2.267254e-05
#> XSeduc         1.614582e-04  3.119701e-06  2.109081e-05 -1.503420e-05
#> XSblhisp       8.093827e-04 -3.703410e-05 -3.754290e-05 -3.100959e-05
#> XStotchr      -1.410652e-03  9.285434e-05  3.428463e-04  4.596481e-05
#> XSins         -2.948364e-04  4.336198e-05  5.360484e-05  1.238556e-05
#> XSincome       2.034895e-06 -1.221543e-07 -6.493827e-07 -1.606099e-08
#> XO(Intercept)  4.393772e-02 -2.552845e-03 -4.989094e-03 -1.728591e-03
#> XOage         -2.552845e-03  5.107889e-04  1.638307e-04  1.753710e-05
#> XOfemale      -4.989094e-03  1.638307e-04  3.180826e-03  9.372247e-05
#> XOeduc        -1.728591e-03  1.753710e-05  9.372247e-05  1.052740e-04
#> XOblhisp      -5.727071e-04  2.524474e-05 -6.287232e-04  5.447514e-05
#> XOtotchr      -2.436548e-03 -5.498951e-05  4.418245e-04  7.167563e-05
#> XOins         -1.644306e-03 -6.736117e-05  4.224072e-04  1.603966e-05
#> sigma          1.309132e-03 -5.859077e-05 -2.929292e-04 -4.012826e-05
#> rho           -1.519548e-02  5.738413e-04  3.542260e-03  3.912621e-04
#>               -4.006095e-02  1.418476e-05  1.157226e-02 -2.627832e-05
#>                    XOblhisp      XOtotchr         XOins         sigma
#> XS(Intercept)  6.713354e-04 -1.543929e-04  3.286943e-05  3.154143e-04
#> XSage          1.616945e-08  4.731693e-06  1.047594e-05 -2.005610e-05
#> XSfemale      -4.969208e-05  1.083817e-04 -2.338154e-05 -2.634153e-04
#> XSeduc        -3.232422e-05  1.111920e-05  6.365650e-06 -1.718690e-05
#> XSblhisp      -5.009667e-04 -4.966990e-05 -2.672521e-05  1.247369e-04
#> XStotchr      -1.926896e-04 -5.629332e-06  2.994162e-05 -4.408124e-04
#> XSins         -7.706378e-05  3.607011e-05 -4.425393e-04 -7.131725e-05
#> XSincome       3.696687e-07 -2.625975e-07 -2.786759e-07 -7.243034e-07
#> XO(Intercept) -5.727071e-04 -2.436548e-03 -1.644306e-03  1.309132e-03
#> XOage          2.524474e-05 -5.498951e-05 -6.736117e-05 -5.859077e-05
#> XOfemale      -6.287232e-04  4.418245e-04  4.224072e-04 -2.929292e-04
#> XOeduc         5.447514e-05  7.167563e-05  1.603966e-05 -4.012826e-05
#> XOblhisp       3.334561e-03 -3.046175e-04 -9.648881e-05  1.770157e-04
#> XOtotchr      -3.046175e-04  1.283735e-03  2.901597e-04 -1.997668e-04
#> XOins         -9.648881e-05  2.901597e-04  2.548323e-03 -6.198009e-05
#> sigma          1.770157e-04 -1.997668e-04 -6.198009e-05  5.592752e-04
#> rho           -2.183412e-03  2.383726e-03  1.191633e-03 -1.089371e-03
#>               -7.772549e-03  7.114243e-03  9.941469e-03  3.707844e-02
#>                         rho              
#> XS(Intercept) -6.062956e-04  4.349877e-02
#> XSage         -1.720827e-04 -6.408184e-03
#> XSfemale       3.256463e-04 -4.043335e-02
#> XSeduc         6.371185e-05 -1.884692e-03
#> XSblhisp      -1.635925e-04  1.911024e-02
#> XStotchr       1.036656e-03 -5.966139e-02
#> XSins          3.112422e-04 -6.955095e-03
#> XSincome      -2.406242e-06 -1.542936e-04
#> XO(Intercept) -1.519548e-02 -4.006095e-02
#> XOage          5.738413e-04  1.418476e-05
#> XOfemale       3.542260e-03  1.157226e-02
#> XOeduc         3.912621e-04 -2.627832e-05
#> XOblhisp      -2.183412e-03 -7.772549e-03
#> XOtotchr       2.383726e-03  7.114243e-03
#> XOins          1.191633e-03  9.941469e-03
#> sigma         -1.089371e-03  3.707844e-02
#> rho            1.490818e-02  6.271223e-02
#>                6.271223e-02  8.174046e+00
#> 
#> $prop_sigmatS
#> XS(Intercept)         XSage      XSfemale        XSeduc      XSblhisp 
#>   0.207699822   0.029745678   0.068558571   0.012805452   0.066528899 
#>      XStotchr         XSins      XSincome XO(Intercept)         XOage 
#>   0.087251702   0.068009662   0.001446579   0.209613253   0.022600639 
#>      XOfemale        XOeduc      XOblhisp      XOtotchr         XOins 
#>   0.056398811   0.010260313   0.057745661   0.035829254   0.050480914 
#>         sigma           rho               
#>   0.023649001   0.122099068   2.859028916 
#> 
#> $level
#> [1] "0" "1"
#> 
#> $nObs
#> [1] 3328
#> 
#> $nParam
#> [1] 18
#> 
#> $N0
#> [1] 526
#> 
#> $N1
#> [1] 2802
#> 
#> $NXS
#> [1] 8
#> 
#> $NXO
#> [1] 7
#> 
#> $df
#> [1] 3310
#> 
#> $aic
#> [1] 11680.15
#> 
#> $bic
#> [1] 11790.13
#> 
#> $initial.value
#> XS(Intercept)         XSage      XSfemale        XSeduc      XSblhisp 
#>  -0.668643899   0.086814848   0.663505390   0.061883892  -0.365784312 
#>      XStotchr         XSins      XSincome XO(Intercept)         XOage 
#>   0.795747277   0.169106526   0.002677301   5.288927373   0.202466773 
#>      XOfemale        XOeduc      XOblhisp      XOtotchr         XOins 
#>   0.292133967   0.012388871  -0.182865733   0.500633176  -0.046509658 
#>         sigma           rho               
#>   1.290541875  -0.359316986  12.000000000 
#> 
#> attr(,"class")
#> [1] "HeckmantS" "list"