Heckman-t Model Fit Function

Fits a sample selection model based on the Student's t-distribution, extending the classical Heckman model to account for heavy-tailed error terms. The estimation is performed via Maximum Likelihood using the BFGS algorithm.

Usage

HeckmantS(selection, outcome, data = sys.frame(sys.parent()), df, start = NULL)

Arguments

selection

A formula specifying the selection equation.

outcome

A formula specifying the outcome equation.

data

A data frame containing the variables in the model.

df

Initial value for the degrees of freedom parameter of the t-distribution.

start

Optional numeric vector of initial parameter values.

Value

A list containing:

coefficients: Named vector of estimated model parameters.

value: Negative of the maximum log-likelihood.

loglik: Maximum log-likelihood.

counts: Number of gradient evaluations performed.

hessian: Hessian matrix at the optimum.

fisher_infotS: Approximate Fisher information matrix.

prop_sigmatS: Standard errors for the parameter estimates.

level: Levels of the selection variable.

nObs: Number of observations.

nParam: Number of model parameters.

N0: Number of censored (unobserved) observations.

N1: Number of uncensored (observed) observations.

NXS: Number of parameters in the selection equation.

NXO: Number of parameters in the outcome equation.

df: Degrees of freedom (observations minus parameters).

aic: Akaike Information Criterion.

bic: Bayesian Information Criterion.

initial.value: Initial parameter values used in the optimization.

Details

The function implements the Heckman sample selection model using the Student's t-distribution for the error terms, as proposed by Marchenko and Genton (2012) . This extension allows for robustness against outliers and heavy-tailed distributions. Initial parameter values can be specified by the user or default to standard starting values.

References

Yulia V Marchenko, Marc G Genton (2012). “A Heckman selection-t model.” Journal of the American Statistical Association, 107(497), 304–317.

Examples

data(MEPS2001)
attach(MEPS2001)
#> The following objects are masked from Mroz87:
#> 
#>     age, educ
#> The following objects are masked from MEPS2001 (pos = 4):
#> 
#>     age, age2, agefem, ambexp, blhisp, dambexp, dhospexp, educ,
#>     fairpoor, female, ffs, good, hospexp, income, ins, instype,
#>     instype_s1, lambexp, lnambx, totchr, vgood, year01
#> The following objects are masked from MEPS2001 (pos = 5):
#> 
#>     age, age2, agefem, ambexp, blhisp, dambexp, dhospexp, educ,
#>     fairpoor, female, ffs, good, hospexp, income, ins, instype,
#>     instype_s1, lambexp, lnambx, totchr, vgood, year01
#> The following objects are masked from MEPS2001 (pos = 6):
#> 
#>     age, age2, agefem, ambexp, blhisp, dambexp, dhospexp, educ,
#>     fairpoor, female, ffs, good, hospexp, income, ins, instype,
#>     instype_s1, lambexp, lnambx, totchr, vgood, year01
selectEq <- dambexp ~ age + female + educ + blhisp + totchr + ins + income
outcomeEq <- lnambx ~ age + female + educ + blhisp + totchr + ins
HeckmantS(selectEq, outcomeEq, data = MEPS2001, df = 12)
#> Start not provided using default start values.
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> $coefficients
#>  (Intercept)          age       female         educ       blhisp       totchr 
#> -0.748008426  0.098537421  0.724814008  0.064845037 -0.393545833  0.890048600 
#>          ins       income  (Intercept)          age       female         educ 
#>  0.180013566  0.002977689  5.205851389  0.206829557  0.306506946  0.017320814 
#>       blhisp       totchr          ins        sigma          rho           df 
#> -0.192918952  0.512679918 -0.052509481  1.194922359 -0.322182914 12.939700903 
#> 
#> $value
#> [1] -5822.076
#> 
#> $loglik
#> [1] 5822.076
#> 
#> $counts
#> gradient 
#>       69 
#> 
#> $hessian
#>               [,1]          [,2]          [,3]          [,4]          [,5]
#>  [1,]  -1045.39370   -3975.43317   -359.445086  -13505.33713 -4.301082e+02
#>  [2,]  -3975.43317  -16389.25636  -1349.441558  -51292.51684 -1.583080e+03
#>  [3,]   -359.44509   -1349.44156   -359.445086   -4657.20181 -1.695715e+02
#>  [4,] -13505.33713  -51292.51684  -4657.201812 -181870.33952 -5.232626e+03
#>  [5,]   -430.10822   -1583.08044   -169.571465   -5232.62577 -4.301082e+02
#>  [6,]   -145.87272    -602.90254    -46.168278   -1835.30154 -5.929791e+01
#>  [7,]   -346.57866   -1411.51466   -103.912034   -4536.65079 -1.329567e+02
#>  [8,] -35006.37702 -138440.32893 -10350.275366 -473403.83882 -1.240492e+04
#>  [9,]   -190.81419    -739.61770    -76.251135   -2507.86195 -7.078488e+01
#> [10,]   -739.61723   -3106.56847   -287.820215   -9705.23068 -2.662153e+02
#> [11,]    -76.25113    -287.82055    -76.251135    -996.26987 -3.293747e+01
#> [12,]  -2507.84574   -9705.16778   -996.258749  -34308.11094 -8.753234e+02
#> [13,]    -70.78488    -266.21539    -32.937470    -875.32538 -7.078488e+01
#> [14,]    -35.57108    -148.81099    -12.327729    -450.68314 -1.347594e+01
#> [15,]    -66.95956    -274.05210    -23.868850    -892.12039 -2.416112e+01
#> [16,]    -27.60119    -128.72212    -24.944849    -415.34031 -7.328243e-01
#> [17,]    -89.56736    -341.62573    -17.436733   -1094.34522 -4.616679e+01
#> [18,]     -3.45534     -14.59523     -2.227021     -47.50509 -7.603206e-01
#>               [,6]          [,7]          [,8]          [,9]         [,10]
#>  [1,]  -145.872724   -346.578656   -35006.3770   -190.814190    -739.61723
#>  [2,]  -602.902535  -1411.514655  -138440.3289   -739.617702   -3106.56847
#>  [3,]   -46.168278   -103.912034   -10350.2754    -76.251135    -287.82021
#>  [4,] -1835.301539  -4536.650793  -473403.8388  -2507.861949   -9705.23068
#>  [5,]   -59.297907   -132.956718   -12404.9186    -70.784882    -266.21532
#>  [6,]  -166.302392    -43.201562    -4820.9680    -35.571084    -148.81070
#>  [7,]   -43.201562   -346.578656   -13210.8981    -66.959560    -274.05190
#>  [8,] -4820.968011 -13210.898073 -1752725.2576  -6589.610077  -26632.20267
#>  [9,]   -35.571084    -66.959560    -6589.6101  -1795.383139   -7408.12511
#> [10,]  -148.810697   -274.051897   -26632.2027  -7408.125110  -32811.56748
#> [11,]   -12.327730    -23.868850    -2304.6653   -998.513405   -4103.13340
#> [12,]  -450.674308   -892.113099   -90464.4380 -24387.321975 -100669.08015
#> [13,]   -13.475940    -24.161121    -2088.2049   -498.025973   -1963.48482
#> [14,]   -41.546539    -11.173831    -1209.7458  -1005.547415   -4505.96838
#> [15,]   -11.173832    -66.959560    -2605.4551   -689.099546   -2935.20759
#> [16,]   -18.576624    -13.956995    -1350.4036     34.269932      92.68289
#> [17,]     2.124602    -23.739620    -2962.0442   -537.929224   -2052.80441
#> [18,]    -1.592731     -1.355451     -131.6966     -3.260215     -14.01219
#>               [,11]         [,12]         [,13]         [,14]        [,15]
#>  [1,]    -76.251135   -2507.84574   -70.7848816    -35.571081   -66.959560
#>  [2,]   -287.820552   -9705.16778  -266.2153863   -148.810988  -274.052101
#>  [3,]    -76.251135    -996.25875   -32.9374699    -12.327729   -23.868850
#>  [4,]   -996.269866  -34308.11094  -875.3253803   -450.683137  -892.120386
#>  [5,]    -32.937470    -875.32339   -70.7848816    -13.475938   -24.161121
#>  [6,]    -12.327730    -450.67431   -13.4759395    -41.546539   -11.173832
#>  [7,]    -23.868850    -892.11310   -24.1611215    -11.173831   -66.959560
#>  [8,]  -2304.665332  -90464.43798 -2088.2049264  -1209.745830 -2605.455069
#>  [9,]   -998.513405  -24387.32197  -498.0259731  -1005.547415  -689.099546
#> [10,]  -4103.133403 -100669.08015 -1963.4848206  -4505.968379 -2935.207588
#> [11,]   -998.513405  -13531.98702  -294.9019286   -620.804547  -353.785224
#> [12,] -13531.987017 -342589.85324 -6339.2770447 -13567.798161 -9447.500214
#> [13,]   -294.901929   -6339.27704  -498.0259731   -254.218174  -173.622727
#> [14,]   -620.804547  -13567.79816  -254.2181743  -1757.318782  -349.643936
#> [15,]   -353.785224   -9447.50021  -173.6227274   -349.643936  -689.099546
#> [16,]     -8.605080     343.53475    28.0153163    -31.897419    11.642713
#> [17,]   -185.076036   -6938.22010  -221.0280225    -71.882632  -175.802995
#> [18,]     -1.938215     -46.01362    -0.8016963     -2.084141    -1.015717
#>               [,16]        [,17]        [,18]
#>  [1,]   -27.6011882   -89.567364   -3.4553397
#>  [2,]  -128.7221203  -341.625732  -14.5952306
#>  [3,]   -24.9448492   -17.436733   -2.2270210
#>  [4,]  -415.3403117 -1094.345219  -47.5050926
#>  [5,]    -0.7328243   -46.166787   -0.7603206
#>  [6,]   -18.5766237     2.124602   -1.5927314
#>  [7,]   -13.9569947   -23.739620   -1.3554510
#>  [8,] -1350.4035811 -2962.044218 -131.6966174
#>  [9,]    34.2699319  -537.929224   -3.2602148
#> [10,]    92.6828902 -2052.804410  -14.0121909
#> [11,]    -8.6050799  -185.076036   -1.9382150
#> [12,]   343.5347471 -6938.220101  -46.0136155
#> [13,]    28.0153163  -221.028023   -0.8016963
#> [14,]   -31.8974187   -71.882632   -2.0841413
#> [15,]    11.6427126  -175.802995   -1.0157173
#> [16,] -3227.2589475  -314.317386   18.0903754
#> [17,]  -314.3173864  -378.050198    1.9015969
#> [18,]    18.0903754     1.901597   -0.2650978
#> 
#> $fisher_infotS
#>                [,1]          [,2]          [,3]          [,4]          [,5]
#>  [1,]  4.313206e-02 -3.499845e-03 -1.214164e-03 -2.126212e-03 -4.654615e-03
#>  [2,] -3.499845e-03  8.846723e-04 -1.106910e-05  3.884420e-05  1.471107e-04
#>  [3,] -1.214164e-03 -1.106910e-05  4.697590e-03 -5.235051e-05 -4.567964e-04
#>  [4,] -2.126212e-03  3.884420e-05 -5.235051e-05  1.639504e-04  1.706810e-04
#>  [5,] -4.654615e-03  1.471107e-04 -4.567964e-04  1.706810e-04  4.425169e-03
#>  [6,] -9.490630e-04 -2.577233e-04  4.669552e-04  5.823543e-05 -1.437192e-04
#>  [7,] -6.960914e-05 -3.081453e-04  2.678905e-04 -9.537224e-06 -9.163696e-06
#>  [8,]  2.931249e-05 -7.886939e-06  1.434185e-05 -5.703532e-06  5.654900e-06
#>  [9,] -4.189296e-03  5.093232e-04 -3.374066e-04  1.636323e-04  7.988552e-04
#> [10,]  3.199587e-04 -9.669733e-05  3.086792e-05  3.037357e-06 -3.660503e-05
#> [11,] -3.522331e-05 -3.153578e-05 -3.452915e-04  2.062648e-05 -3.504332e-05
#> [12,]  2.093608e-04 -3.430417e-06  2.207745e-05 -1.510006e-05 -3.072246e-05
#> [13,]  6.678964e-04 -2.695695e-07 -4.667333e-05 -3.205206e-05 -5.027988e-04
#> [14,] -1.502201e-04  5.043946e-06  1.050588e-04  1.080325e-05 -4.799420e-05
#> [15,]  3.466588e-05  1.064706e-05 -2.480528e-05  6.229292e-06 -2.602564e-05
#> [16,]  3.392095e-04 -2.202508e-05 -2.834953e-04 -1.841428e-05  1.341832e-04
#> [17,] -5.513595e-04 -1.613629e-04  2.900273e-04  5.861463e-05 -1.456420e-04
#> [18,]  4.330287e-02 -6.403252e-03 -4.028177e-02 -1.872992e-03  1.903963e-02
#>                [,6]          [,7]          [,8]          [,9]         [,10]
#>  [1,] -9.490630e-04 -6.960914e-05  2.931249e-05 -4.189296e-03  3.199587e-04
#>  [2,] -2.577233e-04 -3.081453e-04 -7.886939e-06  5.093232e-04 -9.669733e-05
#>  [3,]  4.669552e-04  2.678905e-04  1.434185e-05 -3.374066e-04  3.086792e-05
#>  [4,]  5.823543e-05 -9.537224e-06 -5.703532e-06  1.636323e-04  3.037357e-06
#>  [5,] -1.437192e-04 -9.163696e-06  5.654900e-06  7.988552e-04 -3.660503e-05
#>  [6,]  7.604288e-03  3.483598e-04  2.729590e-06 -1.365975e-03  9.105737e-05
#>  [7,]  3.483598e-04  4.624525e-03 -8.482264e-06 -2.853521e-04  4.299910e-05
#>  [8,]  2.729590e-06 -8.482264e-06  2.092344e-06  2.038845e-06 -1.224791e-07
#>  [9,] -1.365975e-03 -2.853521e-04  2.038845e-06  4.358948e-02 -2.539097e-03
#> [10,]  9.105737e-05  4.299910e-05 -1.224791e-07 -2.539097e-03  5.102564e-04
#> [11,]  3.325927e-04  5.141202e-05 -6.502499e-07 -4.908491e-03  1.606477e-04
#> [12,]  4.473782e-05  1.212931e-05 -1.621376e-08 -1.719266e-03  1.716575e-05
#> [13,] -1.864294e-04 -7.575242e-05  3.701352e-07 -6.221126e-04  2.719712e-05
#> [14,] -1.267151e-05  3.458834e-05 -2.631389e-07 -2.382220e-03 -5.714144e-05
#> [15,]  2.692595e-05 -4.434557e-04 -2.785408e-07 -1.619414e-03 -6.834498e-05
#> [16,] -4.733481e-04 -7.627660e-05 -7.870119e-07  1.378350e-03 -6.191941e-05
#> [17,]  9.436694e-04  2.867313e-04 -2.282783e-06 -1.409368e-02  5.316979e-04
#> [18,] -5.939423e-02 -6.903804e-03 -1.540497e-04 -4.134406e-02  6.778188e-05
#>               [,11]         [,12]         [,13]         [,14]         [,15]
#>  [1,] -3.522331e-05  2.093608e-04  6.678964e-04 -1.502201e-04  3.466588e-05
#>  [2,] -3.153578e-05 -3.430417e-06 -2.695695e-07  5.043946e-06  1.064706e-05
#>  [3,] -3.452915e-04  2.207745e-05 -4.667333e-05  1.050588e-04 -2.480528e-05
#>  [4,]  2.062648e-05 -1.510006e-05 -3.205206e-05  1.080325e-05  6.229292e-06
#>  [5,] -3.504332e-05 -3.072246e-05 -5.027988e-04 -4.799420e-05 -2.602564e-05
#>  [6,]  3.325927e-04  4.473782e-05 -1.864294e-04 -1.267151e-05  2.692595e-05
#>  [7,]  5.141202e-05  1.212931e-05 -7.575242e-05  3.458834e-05 -4.434557e-04
#>  [8,] -6.502499e-07 -1.621376e-08  3.701352e-07 -2.631389e-07 -2.785408e-07
#>  [9,] -4.908491e-03 -1.719266e-03 -6.221126e-04 -2.382220e-03 -1.619414e-03
#> [10,]  1.606477e-04  1.716575e-05  2.719712e-05 -5.714144e-05 -6.834498e-05
#> [11,]  3.162260e-03  9.156055e-05 -6.173219e-04  4.292596e-04  4.166542e-04
#> [12,]  9.156055e-05  1.050250e-04  5.579825e-05  7.022066e-05  1.537482e-05
#> [13,] -6.173219e-04  5.579825e-05  3.327651e-03 -2.969220e-04 -9.295391e-05
#> [14,]  4.292596e-04  7.022066e-05 -2.969220e-04  1.275311e-03  2.862806e-04
#> [15,]  4.166542e-04  1.537482e-05 -9.295391e-05  2.862806e-04  2.546616e-03
#> [16,] -3.081021e-04 -4.241674e-05  1.861193e-04 -2.102037e-04 -6.420410e-05
#> [17,]  3.285988e-03  3.625831e-04 -2.025726e-03  2.211186e-03  1.107333e-03
#> [18,]  1.186541e-02  1.008059e-05 -7.950500e-03  7.311587e-03  1.002101e-02
#>               [,16]         [,17]         [,18]
#>  [1,]  3.392095e-04 -5.513595e-04  4.330287e-02
#>  [2,] -2.202508e-05 -1.613629e-04 -6.403252e-03
#>  [3,] -2.834953e-04  2.900273e-04 -4.028177e-02
#>  [4,] -1.841428e-05  5.861463e-05 -1.872992e-03
#>  [5,]  1.341832e-04 -1.456420e-04  1.903963e-02
#>  [6,] -4.733481e-04  9.436694e-04 -5.939423e-02
#>  [7,] -7.627660e-05  2.867313e-04 -6.903804e-03
#>  [8,] -7.870119e-07 -2.282783e-06 -1.540497e-04
#>  [9,]  1.378350e-03 -1.409368e-02 -4.134406e-02
#> [10,] -6.191941e-05  5.316979e-04  6.778188e-05
#> [11,] -3.081021e-04  3.285988e-03  1.186541e-02
#> [12,] -4.241674e-05  3.625831e-04  1.008059e-05
#> [13,]  1.861193e-04 -2.025726e-03 -7.950500e-03
#> [14,] -2.102037e-04  2.211186e-03  7.311587e-03
#> [15,] -6.420410e-05  1.107333e-03  1.002101e-02
#> [16,]  6.591790e-04 -1.082752e-03  4.020097e-02
#> [17,] -1.082752e-03  1.311899e-02  6.052688e-02
#> [18,]  4.020097e-02  6.052688e-02  8.170450e+00
#> 
#> $prop_sigmatS
#>  [1] 0.207682592 0.029743441 0.068538966 0.012804314 0.066521946 0.087202571
#>  [7] 0.068003862 0.001446494 0.208780936 0.022588856 0.056233977 0.010248169
#> [13] 0.057685794 0.035711499 0.050464003 0.025674482 0.114538138 2.858399986
#> 
#> $level
#> [1] "0" "1"
#> 
#> $nObs
#> [1] 3328
#> 
#> $nParam
#> [1] 18
#> 
#> $N0
#> [1] 526
#> 
#> $N1
#> [1] 2802
#> 
#> $NXS
#> [1] 8
#> 
#> $NXO
#> [1] 7
#> 
#> $df
#> [1] 3310
#> 
#> $aic
#> [1] 11680.15
#> 
#> $bic
#> [1] 11790.13
#> 
#> $initial.value
#>  [1] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [9] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 1.000000
#> [17] 0.000000 2.484907
#> 
#> attr(,"class")
#> [1] "HeckmantS" "list"