Heckman-t Model Fit Function

Fits a sample selection model based on the Student's t-distribution, extending the classical Heckman model to account for heavy-tailed error terms. The estimation is performed via Maximum Likelihood using the BFGS algorithm.

Usage

HeckmantS(selection, outcome, data = sys.frame(sys.parent()), df, start = NULL)

Arguments

selection

A formula specifying the selection equation.

outcome

A formula specifying the outcome equation.

data

A data frame containing the variables in the model.

df

Initial value for the degrees of freedom parameter of the t-distribution.

start

Optional numeric vector of initial parameter values.

Value

A list containing:

coefficients: Named vector of estimated model parameters.

value: Negative of the maximum log-likelihood.

loglik: Maximum log-likelihood.

counts: Number of gradient evaluations performed.

hessian: Hessian matrix at the optimum.

fisher_infotS: Approximate Fisher information matrix.

prop_sigmatS: Standard errors for the parameter estimates.

level: Levels of the selection variable.

nObs: Number of observations.

nParam: Number of model parameters.

N0: Number of censored (unobserved) observations.

N1: Number of uncensored (observed) observations.

NXS: Number of parameters in the selection equation.

NXO: Number of parameters in the outcome equation.

df: Degrees of freedom (observations minus parameters).

aic: Akaike Information Criterion.

bic: Bayesian Information Criterion.

initial.value: Initial parameter values used in the optimization.

Details

The function implements the Heckman sample selection model using the Student's t-distribution for the error terms, as proposed by Marchenko and Genton (2012) . This extension allows for robustness against outliers and heavy-tailed distributions. Initial parameter values can be specified by the user or default to standard starting values.

References

Yulia V Marchenko, Marc G Genton (2012). “A Heckman selection-t model.” Journal of the American Statistical Association, 107(497), 304–317.

Examples

data(MEPS2001)
attach(MEPS2001)
#> The following objects are masked from Mroz87:
#> 
#>     age, educ
#> The following objects are masked from MEPS2001 (pos = 4):
#> 
#>     age, age2, agefem, ambexp, blhisp, dambexp, dhospexp, educ,
#>     fairpoor, female, ffs, good, hospexp, income, ins, instype,
#>     instype_s1, lambexp, lnambx, totchr, vgood, year01
#> The following objects are masked from MEPS2001 (pos = 5):
#> 
#>     age, age2, agefem, ambexp, blhisp, dambexp, dhospexp, educ,
#>     fairpoor, female, ffs, good, hospexp, income, ins, instype,
#>     instype_s1, lambexp, lnambx, totchr, vgood, year01
#> The following objects are masked from MEPS2001 (pos = 6):
#> 
#>     age, age2, agefem, ambexp, blhisp, dambexp, dhospexp, educ,
#>     fairpoor, female, ffs, good, hospexp, income, ins, instype,
#>     instype_s1, lambexp, lnambx, totchr, vgood, year01
selectEq <- dambexp ~ age + female + educ + blhisp + totchr + ins + income
outcomeEq <- lnambx ~ age + female + educ + blhisp + totchr + ins
HeckmantS(selectEq, outcomeEq, data = MEPS2001, df = 12)
#> Start not provided using default start values.
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> $coefficients
#>  (Intercept)          age       female         educ       blhisp       totchr 
#> -0.748019043  0.098533957  0.724842140  0.064846420 -0.393556230  0.890059436 
#>          ins       income  (Intercept)          age       female         educ 
#>  0.180016364  0.002977867  5.205886164  0.206829239  0.306506438  0.017318512 
#>       blhisp       totchr          ins        sigma          rho           df 
#> -0.192934870  0.512686542 -0.052509944  1.194892986 -0.322168670 12.934344590 
#> 
#> $value
#> [1] -5822.076
#> 
#> $loglik
#> [1] 5822.076
#> 
#> $counts
#> gradient 
#>       69 
#> 
#> $hessian
#>                [,1]          [,2]          [,3]         [,4]          [,5]
#>  [1,]  -1045.335913   -3975.20743   -359.414455  -13504.5652 -4.300910e+02
#>  [2,]  -3975.207432  -16388.30432  -1349.322374  -51289.5058 -1.583016e+03
#>  [3,]   -359.414455   -1349.32237   -359.414455   -4656.7914 -1.695611e+02
#>  [4,] -13504.565168  -51289.50582  -4656.791414 -181859.6871 -5.232408e+03
#>  [5,]   -430.090984   -1583.01619   -169.561150   -5232.4082 -4.300910e+02
#>  [6,]   -145.860600    -602.85462    -46.163000   -1835.1456 -5.929394e+01
#>  [7,]   -346.558168   -1411.42865   -103.902184   -4536.3727 -1.329523e+02
#>  [8,] -35004.285459 -138431.90013 -10349.321495 -473374.4981 -1.240440e+04
#>  [9,]   -190.799948    -739.56329    -76.244731   -2507.6750 -7.078001e+01
#> [10,]   -739.562821   -3106.34398   -287.796589   -9704.5152 -2.661973e+02
#> [11,]    -76.244731    -287.79693    -76.244731    -996.1872 -3.293481e+01
#> [12,]  -2507.658742   -9704.45232   -996.176053  -34305.5602 -8.752624e+02
#> [13,]    -70.780011    -266.19737    -32.934807    -875.2644 -7.078001e+01
#> [14,]    -35.569055    -148.80304    -12.327036    -450.6581 -1.347519e+01
#> [15,]    -66.954815    -274.03304    -23.866914    -892.0566 -2.415969e+01
#> [16,]    -27.596750    -128.70745    -24.945396    -415.2929 -7.303460e-01
#> [17,]    -89.557989    -341.59549    -17.435199   -1094.2383 -4.616415e+01
#> [18,]     -3.457874     -14.60579     -2.228612     -47.5399 -7.609129e-01
#>               [,6]          [,7]          [,8]         [,9]         [,10]
#>  [1,]  -145.860600   -346.558168   -35004.2855   -190.79995    -739.56282
#>  [2,]  -602.854624  -1411.428645  -138431.9001   -739.56329   -3106.34398
#>  [3,]   -46.163000   -103.902184   -10349.3215    -76.24473    -287.79659
#>  [4,] -1835.145604  -4536.372685  -473374.4981  -2507.67495   -9704.51523
#>  [5,]   -59.293944   -132.952337   -12404.3954    -70.78001    -266.19730
#>  [6,]  -166.289194    -43.198594    -4820.5854    -35.56906    -148.80274
#>  [7,]   -43.198594   -346.558168   -13210.0409    -66.95482    -274.03283
#>  [8,] -4820.585397 -13210.040947 -1752607.6357  -6589.10677  -26630.18862
#>  [9,]   -35.569058    -66.954815    -6589.1068  -1795.39082   -7408.16204
#> [10,]  -148.802745   -274.032834   -26630.1886  -7408.16204  -32811.75078
#> [11,]   -12.327037    -23.866914    -2304.4776   -998.52227   -4103.17504
#> [12,]  -450.649269   -892.049315   -90457.5089 -24387.44513 -100669.64445
#> [13,]   -13.475193    -24.159689    -2088.0579   -498.02689   -1963.48937
#> [14,]   -41.544562    -11.173235    -1209.6813  -1005.56255   -4506.03994
#> [15,]   -11.173236    -66.954815    -2605.2613   -689.10720   -2935.24072
#> [16,]   -18.577696    -13.956406    -1350.3027     34.23691      92.53846
#> [17,]     2.124710    -23.738318    -2961.7796   -537.92964   -2052.81383
#> [18,]    -1.593787     -1.356436     -131.7931     -3.26218     -14.02076
#>               [,11]         [,12]         [,13]         [,14]        [,15]
#>  [1,]    -76.244731   -2507.65874   -70.7800111    -35.569055   -66.954815
#>  [2,]   -287.796926   -9704.45232  -266.1973677   -148.803036  -274.033038
#>  [3,]    -76.244731    -996.17605   -32.9348068    -12.327036   -23.866914
#>  [4,]   -996.187168  -34305.56020  -875.2643948   -450.658097  -892.056601
#>  [5,]    -32.934807    -875.26241   -70.7800111    -13.475191   -24.159689
#>  [6,]    -12.327037    -450.64927   -13.4751928    -41.544562   -11.173236
#>  [7,]    -23.866914    -892.04931   -24.1596887    -11.173235   -66.954815
#>  [8,]  -2304.477593  -90457.50893 -2088.0579023  -1209.681264 -2605.261296
#>  [9,]   -998.522272  -24387.44513  -498.0268913  -1005.562550  -689.107202
#> [10,]  -4103.175038 -100669.64445 -1963.4893720  -4506.039945 -2935.240722
#> [11,]   -998.522272  -13532.11319  -294.9030200   -620.817336  -353.790925
#> [12,] -13532.113186 -342591.76954 -6339.2923835 -13568.013755 -9447.606329
#> [13,]   -294.903020   -6339.29238  -498.0268913   -254.221224  -173.624910
#> [14,]   -620.817336  -13568.01375  -254.2212239  -1757.351110  -349.650848
#> [15,]   -353.790925   -9447.60633  -173.6249104   -349.650848  -689.107202
#> [16,]     -8.626206     343.03363    28.0011939    -31.910629    11.632587
#> [17,]   -185.076727   -6938.23634  -221.0280269    -71.886791  -175.804616
#> [18,]     -1.939404     -46.04109    -0.8021633     -2.085468    -1.016321
#>              [,16]        [,17]        [,18]
#>  [1,]   -27.596750   -89.557989   -3.4578743
#>  [2,]  -128.707453  -341.595494  -14.6057940
#>  [3,]   -24.945396   -17.435199   -2.2286117
#>  [4,]  -415.292925 -1094.238253  -47.5399006
#>  [5,]    -0.730346   -46.164151   -0.7609129
#>  [6,]   -18.577696     2.124710   -1.5937867
#>  [7,]   -13.956406   -23.738318   -1.3564356
#>  [8,] -1350.302702 -2961.779624 -131.7930574
#>  [9,]    34.236914  -537.929642   -3.2621802
#> [10,]    92.538456 -2052.813834  -14.0207611
#> [11,]    -8.626206  -185.076727   -1.9394037
#> [12,]   343.033627 -6938.236339  -46.0410879
#> [13,]    28.001194  -221.028027   -0.8021633
#> [14,]   -31.910629   -71.886791   -2.0854683
#> [15,]    11.632587  -175.804616   -1.0163213
#> [16,] -3227.190075  -314.286394   18.1021806
#> [17,]  -314.286394  -378.037383    1.9028018
#> [18,]    18.102181     1.902802   -0.2655836
#> 
#> $fisher_infotS
#>                [,1]          [,2]          [,3]          [,4]          [,5]
#>  [1,]  4.313375e-02 -3.500000e-03 -1.214053e-03 -2.126291e-03 -4.654934e-03
#>  [2,] -3.500000e-03  8.847170e-04 -1.109301e-05  3.884487e-05  1.471318e-04
#>  [3,] -1.214053e-03 -1.109301e-05  4.697741e-03 -5.235787e-05 -4.567688e-04
#>  [4,] -2.126291e-03  3.884487e-05 -5.235787e-05  1.639570e-04  1.706921e-04
#>  [5,] -4.654934e-03  1.471318e-04 -4.567688e-04  1.706921e-04  4.425352e-03
#>  [6,] -9.486884e-04 -2.578003e-04  4.667081e-04  5.822508e-05 -1.436173e-04
#>  [7,] -6.955367e-05 -3.081641e-04  2.678786e-04 -9.540082e-06 -9.178129e-06
#>  [8,]  2.931381e-05 -7.887625e-06  1.434196e-05 -5.703822e-06  5.655671e-06
#>  [9,] -4.189627e-03  5.093448e-04 -3.375106e-04  1.636414e-04  7.990585e-04
#> [10,]  3.199742e-04 -9.669708e-05  3.086669e-05  3.036477e-06 -3.661025e-05
#> [11,] -3.512132e-05 -3.154244e-05 -3.452484e-04  2.062261e-05 -3.509919e-05
#> [12,]  2.093716e-04 -3.432091e-06  2.207445e-05 -1.510017e-05 -3.072496e-05
#> [13,]  6.678148e-04 -2.657410e-07 -4.669232e-05 -3.204896e-05 -5.027497e-04
#> [14,] -1.501501e-04  5.039836e-06  1.050776e-04  1.080070e-05 -4.802927e-05
#> [15,]  3.466940e-05  1.064825e-05 -2.476736e-05  6.229279e-06 -2.605442e-05
#> [16,]  3.389544e-04 -2.199003e-05 -2.833154e-04 -1.840409e-05  1.341094e-04
#> [17,] -5.510202e-04 -1.613765e-04  2.902144e-04  5.860347e-05 -1.458891e-04
#> [18,]  4.323498e-02 -6.392716e-03 -4.021904e-02 -1.870223e-03  1.900890e-02
#>                [,6]          [,7]          [,8]          [,9]         [,10]
#>  [1,] -9.486884e-04 -6.955367e-05  2.931381e-05 -4.189627e-03  3.199742e-04
#>  [2,] -2.578003e-04 -3.081641e-04 -7.887625e-06  5.093448e-04 -9.669708e-05
#>  [3,]  4.667081e-04  2.678786e-04  1.434196e-05 -3.375106e-04  3.086669e-05
#>  [4,]  5.822508e-05 -9.540082e-06 -5.703822e-06  1.636414e-04  3.036477e-06
#>  [5,] -1.436173e-04 -9.178129e-06  5.655671e-06  7.990585e-04 -3.661025e-05
#>  [6,]  7.604425e-03  3.483078e-04  2.728479e-06 -1.366306e-03  9.106068e-05
#>  [7,]  3.483078e-04  4.624768e-03 -8.482870e-06 -2.853336e-04  4.299572e-05
#>  [8,]  2.728479e-06 -8.482870e-06  2.092491e-06  2.038862e-06 -1.224739e-07
#>  [9,] -1.366306e-03 -2.853336e-04  2.038862e-06  4.358989e-02 -2.539082e-03
#> [10,]  9.106068e-05  4.299572e-05 -1.224739e-07 -2.539082e-03  5.102522e-04
#> [11,]  3.326870e-04  5.141118e-05 -6.502066e-07 -4.908609e-03  1.606443e-04
#> [12,]  4.473854e-05  1.212803e-05 -1.624229e-08 -1.719292e-03  1.716642e-05
#> [13,] -1.864809e-04 -7.574758e-05  3.701026e-07 -6.221111e-04  2.720135e-05
#> [14,] -1.262210e-05  3.458621e-05 -2.631148e-07 -2.382221e-03 -5.714482e-05
#> [15,]  2.700121e-05 -4.434396e-04 -2.784404e-07 -1.619448e-03 -6.834446e-05
#> [16,] -4.730626e-04 -7.624058e-05 -7.863843e-07  1.378454e-03 -6.191654e-05
#> [17,]  9.441059e-04  2.867290e-04 -2.282407e-06 -1.409395e-02  5.316861e-04
#> [18,] -5.929615e-02 -6.893023e-03 -1.538194e-04 -4.128248e-02  6.738710e-05
#>               [,11]         [,12]         [,13]         [,14]         [,15]
#>  [1,] -3.512132e-05  2.093716e-04  6.678148e-04 -1.501501e-04  3.466940e-05
#>  [2,] -3.154244e-05 -3.432091e-06 -2.657410e-07  5.039836e-06  1.064825e-05
#>  [3,] -3.452484e-04  2.207445e-05 -4.669232e-05  1.050776e-04 -2.476736e-05
#>  [4,]  2.062261e-05 -1.510017e-05 -3.204896e-05  1.080070e-05  6.229279e-06
#>  [5,] -3.509919e-05 -3.072496e-05 -5.027497e-04 -4.802927e-05 -2.605442e-05
#>  [6,]  3.326870e-04  4.473854e-05 -1.864809e-04 -1.262210e-05  2.700121e-05
#>  [7,]  5.141118e-05  1.212803e-05 -7.574758e-05  3.458621e-05 -4.434396e-04
#>  [8,] -6.502066e-07 -1.624229e-08  3.701026e-07 -2.631148e-07 -2.784404e-07
#>  [9,] -4.908609e-03 -1.719292e-03 -6.221111e-04 -2.382221e-03 -1.619448e-03
#> [10,]  1.606443e-04  1.716642e-05  2.720135e-05 -5.714482e-05 -6.834446e-05
#> [11,]  3.162279e-03  9.156810e-05 -6.173213e-04  4.292646e-04  4.166586e-04
#> [12,]  9.156810e-05  1.050257e-04  5.579628e-05  7.022270e-05  1.537829e-05
#> [13,] -6.173213e-04  5.579628e-05  3.327632e-03 -2.969169e-04 -9.295577e-05
#> [14,]  4.292646e-04  7.022270e-05 -2.969169e-04  1.275292e-03  2.862802e-04
#> [15,]  4.166586e-04  1.537829e-05 -9.295577e-05  2.862802e-04  2.546589e-03
#> [16,] -3.081356e-04 -4.241507e-05  1.861276e-04 -2.102128e-04 -6.424215e-05
#> [17,]  3.286068e-03  3.626051e-04 -2.025724e-03  2.211183e-03  1.107330e-03
#> [18,]  1.184755e-02  1.044868e-05 -7.938867e-03  7.300716e-03  1.000511e-02
#>               [,16]         [,17]         [,18]
#>  [1,]  3.389544e-04 -5.510202e-04  4.323498e-02
#>  [2,] -2.199003e-05 -1.613765e-04 -6.392716e-03
#>  [3,] -2.833154e-04  2.902144e-04 -4.021904e-02
#>  [4,] -1.840409e-05  5.860347e-05 -1.870223e-03
#>  [5,]  1.341094e-04 -1.458891e-04  1.900890e-02
#>  [6,] -4.730626e-04  9.441059e-04 -5.929615e-02
#>  [7,] -7.624058e-05  2.867290e-04 -6.893023e-03
#>  [8,] -7.863843e-07 -2.282407e-06 -1.538194e-04
#>  [9,]  1.378454e-03 -1.409395e-02 -4.128248e-02
#> [10,] -6.191654e-05  5.316861e-04  6.738710e-05
#> [11,] -3.081356e-04  3.286068e-03  1.184755e-02
#> [12,] -4.241507e-05  3.626051e-04  1.044868e-05
#> [13,]  1.861276e-04 -2.025724e-03 -7.938867e-03
#> [14,] -2.102128e-04  2.211183e-03  7.300716e-03
#> [15,] -6.424215e-05  1.107330e-03  1.000511e-02
#> [16,]  6.589598e-04 -1.082931e-03  4.013272e-02
#> [17,] -1.082931e-03  1.311913e-02  6.043407e-02
#> [18,]  4.013272e-02  6.043407e-02  8.151115e+00
#> 
#> $prop_sigmatS
#>  [1] 0.207686662 0.029744192 0.068540069 0.012804568 0.066523321 0.087203352
#>  [7] 0.068005646 0.001446544 0.208781917 0.022588764 0.056234142 0.010248202
#> [13] 0.057685628 0.035711227 0.050463743 0.025670212 0.114538789 2.855015718
#> 
#> $level
#> [1] "0" "1"
#> 
#> $nObs
#> [1] 3328
#> 
#> $nParam
#> [1] 18
#> 
#> $N0
#> [1] 526
#> 
#> $N1
#> [1] 2802
#> 
#> $NXS
#> [1] 8
#> 
#> $NXO
#> [1] 7
#> 
#> $df
#> [1] 3310
#> 
#> $aic
#> [1] 11680.15
#> 
#> $bic
#> [1] 11790.13
#> 
#> $initial.value
#>  [1] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [9] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 1.000000
#> [17] 0.000000 2.484907
#> 
#> attr(,"class")
#> [1] "HeckmantS" "list"