U.S. Women's Labor Force Participation

The Mroz87 data frame contains data about 753 married women. These data are collected within the "Panel Study of Income Dynamics" (PSID). Of the 753 observations, the first 428 are for women with positive hours worked in 1975, while the remaining 325 observations are for women who did not work for pay in 1975. A more complete discussion of the data is found in Mroz (1987) . It also includes the following explanatory variables:

lfp: Dummy variable for labor-force participation.

hours: Wife's hours of work in 1975.

kids5: Number of children 5 years old or younger.

kids618: Number of children 6 to 18 years old.

Age: Wife's age.

Educ: Wife's educational attainment, in years.

wage: Wife's average hourly earnings, in 1975 dollars.

repwage: Wife's wage reported at the time of the 1976 interview.

hushrs: Husband's hours worked in 1975.

husage: Husband's age.

huseduc: Husband's educational attainment, in years.

huswage: Husband's wage, in 1975 dollars.

faminc: Family income, in 1975 dollars.

mtr: Marginal tax rate facing the wife.

motheduc: Wife's mother's educational attainment, in years.

fatheduc: Wife's father's educational attainment, in years.

unem: Unemployment rate in county of residence, in percentage points.

city: Dummy variable = 1 if live in large city, else 0.

exper: Actual years of wife's previous labor market experience.

nwifeinc: Non-wife income.

wifecoll: Dummy variable for wife's college attendance.

huscoll: Dummy variable for husband's college attendance.

Mroz87

Format

An object of class data.frame with 753 rows and 22 columns.

Source

PSID Staff, The Panel Study of Income Dynamics, Institute for Social ResearchPanel Study of Income Dynamics, University of Michigan, http://psidonline.isr.umich.edu.

References

Thomas A Mroz (1987). “The sensitivity of an empirical model of married women's hours of work to economic and statistical assumptions.” Econometrica: Journal of the Econometric Society, 765--799. Mikhail Zhelonkin, Marc G. Genton, Elvezio Ronchetti (2019). ssmrob: Robust Estimation and Inference in Sample Selection Models. R package version 0.7, https://CRAN.R-project.org/package=ssmrob. Ott Toomet, Arne Henningsen (2008). “Sample Selection Models in R: Package sampleSelection.” Journal of Statistical Software, 27(7). http://www.jstatsoft.org/v27/i07/. Jeffrey M Wooldridge (2016). Introductory econometrics: A modern approach. Nelson Education.

Examples

# Wooldridge(2016): page 247
data(Mroz87)
attach(Mroz87)

#> The following objects are masked from MEPS2001 (pos = 3):
#> 
#>     age, educ

#> The following objects are masked from MEPS2001 (pos = 4):
#> 
#>     age, educ

#> The following objects are masked from MEPS2001 (pos = 5):
#> 
#>     age, educ

#> The following objects are masked from MEPS2001 (pos = 6):
#> 
#>     age, educ

#> The following objects are masked from MEPS2001 (pos = 7):
#> 
#>     age, educ

#> The following objects are masked from MEPS2001 (pos = 8):
#> 
#>     age, educ

#> The following objects are masked from MEPS2001 (pos = 9):
#> 
#>     age, educ

selectEq  <- lfp ~ nwifeinc + educ + exper + I(exper^2) + age + kids5 + kids618
outcomeEq <- log(wage) ~ educ + exper + I(exper^2)
outcomeS  <- cbind(educ, exper)
outcomeC  <- cbind(educ, exper)
outcomeBS <- wage ~ educ + exper + I(exper^2)
outcomeBS <- wage ~ educ + exper + I(exper^2)
HeckmanCL(selectEq, outcomeEq, data = Mroz87)

#> $coefficients
#>  (Intercept)     nwifeinc         educ        exper   I(exper^2)          age 
#>  0.270073573 -0.012023770  0.130903996  0.123347160 -0.001887289 -0.052852446 
#>        kids5      kids618  (Intercept)         educ        exper   I(exper^2) 
#> -0.868324675  0.036005607 -0.578102306  0.109065489  0.043887302 -0.000859097 
#>        sigma          rho 
#>  0.663799475  0.048601124 
#> 
#> $value
#> [1] -832.8978
#> 
#> $loglik
#> [1] 832.8978
#> 
#> $counts
#> gradient 
#>        1 
#> 
#> $hessian
#>               XS(Intercept)    XSnwifeinc        XSeduc       XSexper
#> XS(Intercept) -3.841431e+02 -7.801652e+03 -4.690202e+03 -3.930787e+03
#> XSnwifeinc    -7.801652e+03 -2.090808e+05 -9.834592e+04 -7.473540e+04
#> XSeduc        -4.690202e+03 -9.834592e+04 -5.911411e+04 -4.780152e+04
#> XSexper       -3.930787e+03 -7.473540e+04 -4.780152e+04 -6.333871e+04
#> XSI(exper^2)  -6.321051e+04 -1.123608e+06 -7.630328e+05 -1.317075e+06
#> XSage         -1.639732e+04 -3.355287e+05 -1.995606e+05 -1.774652e+05
#> XSkids5       -9.101045e+01 -1.755805e+03 -1.183796e+03 -7.214715e+02
#> XSkids618     -5.336924e+02 -1.094389e+04 -6.442923e+03 -4.130979e+03
#> XO(Intercept)  1.500566e+01  2.956320e+02  1.856870e+02  1.724166e+02
#> XOeduc         1.856871e+02  3.773567e+03  2.376064e+03  2.121908e+03
#> XOexper        1.724166e+02  3.187171e+03  2.121908e+03  2.927987e+03
#> XOI(exper^2)   2.927972e+03  5.072790e+04  3.568018e+04  6.264582e+04
#> sigma          6.426914e-02  1.290896e+01 -1.059846e+00  3.803792e-02
#> rho            6.757277e+00 -1.127708e+00  1.019434e+02  4.108881e+01
#>                XSI(exper^2)         XSage       XSkids5     XSkids618
#> XS(Intercept) -6.321051e+04 -1.639732e+04   -91.0104538 -5.336924e+02
#> XSnwifeinc    -1.123608e+06 -3.355287e+05 -1755.8048883 -1.094389e+04
#> XSeduc        -7.630328e+05 -1.995606e+05 -1183.7955517 -6.442923e+03
#> XSexper       -1.317075e+06 -1.774652e+05  -721.4714573 -4.130979e+03
#> XSI(exper^2)  -3.250851e+07 -3.003639e+06 -8126.2360442 -4.838512e+04
#> XSage         -3.003639e+06 -7.246758e+05 -3156.2793510 -2.110235e+04
#> XSkids5       -8.126236e+03 -3.156279e+03  -117.3955653 -1.450579e+02
#> XSkids618     -4.838512e+04 -2.110235e+04  -145.0579212 -1.425219e+03
#> XO(Intercept)  2.926868e+03  6.371433e+02     2.8586698  2.080913e+01
#> XOeduc         3.569400e+04  7.863720e+03    38.0828492  2.538479e+02
#> XOexper        6.258583e+04  7.751901e+03    23.9624977  1.761839e+02
#> XOI(exper^2)   1.560624e+06  1.384080e+05   281.4923278  2.179570e+03
#> sigma         -5.279017e+00  5.164078e+00    -0.2710216  3.907176e-01
#> rho            5.516245e+02  2.636675e+02     6.8164411  6.380234e+00
#>               XO(Intercept)        XOeduc       XOexper  XOI(exper^2)
#> XS(Intercept)      15.00566  1.856871e+02      172.4166  2.927972e+03
#> XSnwifeinc        295.63196  3.773567e+03     3187.1712  5.072790e+04
#> XSeduc            185.68703  2.376064e+03     2121.9076  3.568018e+04
#> XSexper           172.41662  2.121908e+03     2927.9868  6.264582e+04
#> XSI(exper^2)     2926.86768  3.569400e+04    62585.8346  1.560624e+06
#> XSage             637.14330  7.863720e+03     7751.9013  1.384080e+05
#> XSkids5             2.85867  3.808285e+01       23.9625  2.814923e+02
#> XSkids618          20.80913  2.538479e+02      176.1839  2.179570e+03
#> XO(Intercept)    -972.43572 -1.230963e+04   -12676.3172 -2.282062e+05
#> XOeduc         -12309.63323 -1.608894e+05  -160192.2177 -2.867601e+06
#> XOexper        -12676.31723 -1.601922e+05  -228206.2493 -4.985108e+06
#> XOI(exper^2)  -228206.24811 -2.867601e+06 -4985108.4243 -1.242963e+08
#> sigma             -25.01817 -3.038649e+02     -251.8162 -3.997888e+03
#> rho              -342.38510 -4.164895e+03    -3445.0338 -5.475566e+04
#>                       sigma           rho
#> XS(Intercept)  6.426914e-02      6.757277
#> XSnwifeinc     1.290896e+01     -1.127708
#> XSeduc        -1.059846e+00    101.943398
#> XSexper        3.803792e-02     41.088807
#> XSI(exper^2)  -5.279017e+00    551.624480
#> XSage          5.164078e+00    263.667498
#> XSkids5       -2.710216e-01      6.816441
#> XSkids618      3.907176e-01      6.380234
#> XO(Intercept) -2.501817e+01   -342.385099
#> XOeduc        -3.038649e+02  -4164.895260
#> XOexper       -2.518162e+02  -3445.033829
#> XOI(exper^2)  -3.997888e+03 -54755.656407
#> sigma         -1.942498e+03     20.715129
#> rho            2.071513e+01   -199.287133
#> 
#> $fisher_infoHC
#>               XS(Intercept)    XSnwifeinc        XSeduc       XSexper
#> XS(Intercept)  2.594991e-01  2.718850e-04 -7.789529e-03 -1.640174e-03
#> XSnwifeinc     2.718850e-04  2.373249e-05 -4.309163e-05  3.922394e-06
#> XSeduc        -7.789529e-03 -4.309163e-05  6.438223e-04 -1.945572e-05
#> XSexper       -1.640174e-03  3.922394e-06 -1.945572e-05  3.632435e-04
#> XSI(exper^2)   6.346992e-05  1.156980e-07  3.251721e-07 -1.079852e-05
#> XSage         -3.337445e-03 -6.513168e-06  2.068130e-05 -5.466997e-06
#> XSkids5       -1.899580e-02  2.033978e-05 -3.948289e-04 -1.808508e-05
#> XSkids618     -1.008835e-02 -1.000235e-05  1.158510e-04  6.871686e-05
#> XO(Intercept)  6.022574e-03  9.236679e-05 -4.871064e-04  7.546264e-05
#> XOeduc        -2.300538e-04 -2.629779e-06  1.916973e-05 -3.334309e-06
#> XOexper       -2.113263e-04 -3.814569e-06  1.562376e-05  1.508822e-06
#> XOI(exper^2)   4.446188e-06  7.898120e-08 -3.198640e-07 -8.261317e-08
#> sigma         -5.822507e-05 -1.128188e-06  4.711254e-06 -1.412136e-06
#> rho           -3.851831e-03 -7.967870e-05  3.344543e-04 -8.973288e-05
#>                XSI(exper^2)         XSage       XSkids5     XSkids618
#> XS(Intercept)  6.346992e-05 -3.337445e-03 -1.899580e-02 -1.008835e-02
#> XSnwifeinc     1.156980e-07 -6.513168e-06  2.033978e-05 -1.000235e-05
#> XSeduc         3.251721e-07  2.068130e-05 -3.948289e-04  1.158510e-04
#> XSexper       -1.079852e-05 -5.466997e-06 -1.808508e-05  6.871686e-05
#> XSI(exper^2)   3.798656e-07 -4.990624e-07 -2.743623e-07 -3.082357e-07
#> XSage         -4.990624e-07  7.192480e-05  4.549849e-04  1.276900e-04
#> XSkids5       -2.743623e-07  4.549849e-04  1.409431e-02  6.297921e-04
#> XSkids618     -3.082357e-07  1.276900e-04  6.297921e-04  1.889229e-03
#> XO(Intercept) -9.882218e-07 -5.538153e-05 -1.305286e-03  1.092054e-04
#> XOeduc         5.942579e-08  1.603208e-06  3.620852e-05 -2.860859e-06
#> XOexper       -1.009607e-07  2.186112e-06  5.295525e-05 -5.224609e-06
#> XOI(exper^2)   4.229821e-09 -4.419860e-08 -1.072319e-06  1.124489e-07
#> sigma          2.395838e-08  7.484541e-07  1.752961e-05 -1.270694e-06
#> rho            1.504750e-06  4.582844e-05  1.120394e-03 -8.701059e-05
#>               XO(Intercept)        XOeduc       XOexper  XOI(exper^2)
#> XS(Intercept)  6.022574e-03 -2.300538e-04 -2.113263e-04  4.446188e-06
#> XSnwifeinc     9.236679e-05 -2.629779e-06 -3.814569e-06  7.898120e-08
#> XSeduc        -4.871064e-04  1.916973e-05  1.562376e-05 -3.198640e-07
#> XSexper        7.546264e-05 -3.334309e-06  1.508822e-06 -8.261317e-08
#> XSI(exper^2)  -9.882218e-07  5.942579e-08 -1.009607e-07  4.229821e-09
#> XSage         -5.538153e-05  1.603208e-06  2.186112e-06 -4.419860e-08
#> XSkids5       -1.305286e-03  3.620852e-05  5.295525e-05 -1.072319e-06
#> XSkids618      1.092054e-04 -2.860859e-06 -5.224609e-06  1.124489e-07
#> XO(Intercept)  6.663591e-02 -3.220597e-03 -2.101249e-03  4.677659e-05
#> XOeduc        -3.220597e-03  2.197739e-04  2.074961e-05 -2.830950e-07
#> XOexper       -2.101249e-03  2.074961e-05  2.200181e-04 -5.885013e-06
#> XOI(exper^2)   4.677659e-05 -2.830950e-07 -5.885013e-06  1.738280e-07
#> sigma         -4.320090e-04  1.204950e-05  1.799837e-05 -3.720779e-07
#> rho           -2.389701e-02  6.652017e-04  9.972370e-04 -2.062457e-05
#>                       sigma           rho
#> XS(Intercept) -5.822507e-05 -3.851831e-03
#> XSnwifeinc    -1.128188e-06 -7.967870e-05
#> XSeduc         4.711254e-06  3.344543e-04
#> XSexper       -1.412136e-06 -8.973288e-05
#> XSI(exper^2)   2.395838e-08  1.504750e-06
#> XSage          7.484541e-07  4.582844e-05
#> XSkids5        1.752961e-05  1.120394e-03
#> XSkids618     -1.270694e-06 -8.701059e-05
#> XO(Intercept) -4.320090e-04 -2.389701e-02
#> XOeduc         1.204950e-05  6.652017e-04
#> XOexper        1.799837e-05  9.972370e-04
#> XOI(exper^2)  -3.720779e-07 -2.062457e-05
#> sigma          5.204977e-04  3.373580e-04
#> rho            3.373580e-04  2.075767e-02
#> 
#> $prop_sigmaHC
#> XS(Intercept)    XSnwifeinc        XSeduc       XSexper  XSI(exper^2) 
#>  0.5094105161  0.0048716003  0.0253736528  0.0190589484  0.0006163324 
#>         XSage       XSkids5     XSkids618 XO(Intercept)        XOeduc 
#>  0.0084808490  0.1187194404  0.0434652577  0.2581393215  0.0148247744 
#>       XOexper  XOI(exper^2)         sigma           rho 
#>  0.0148330060  0.0004169268  0.0228144176  0.1440752278 
#> 
#> $level
#> [1] "0" "1"
#> 
#> $nObs
#> [1] 753
#> 
#> $nParam
#> [1] 14
#> 
#> $N0
#> [1] 325
#> 
#> $N1
#> [1] 428
#> 
#> $NXS
#> [1] 8
#> 
#> $NXO
#> [1] 4
#> 
#> $df
#> [1] 739
#> 
#> $aic
#> [1] 1693.796
#> 
#> $bic
#> [1] 1758.532
#> 
#> $initial.value
#> XS(Intercept)    XSnwifeinc        XSeduc       XSexper  XSI(exper^2) 
#>  0.2700735726 -0.0120236370  0.1309039692  0.1233471675 -0.0018870674 
#>         XSage       XSkids5     XSkids618 XO(Intercept)        XOeduc 
#> -0.0528524416 -0.8683246795  0.0360056106 -0.5781023058  0.1090654909 
#>       XOexper  XOI(exper^2)         sigma           rho 
#>  0.0438873010 -0.0008591134  0.6637994802  0.0486011464 
#> 
#> attr(,"class")
#> [1] "HeckmanCL" "list"     

HeckmanBS(selectEq, outcomeBS, data = Mroz87)

#> $coefficients
#>   (Intercept)      nwifeinc          educ         exper    I(exper^2) 
#>  1.0572557181 -0.0147278022  0.1681025891  0.1593054823 -0.0024351869 
#>           age         kids5       kids618   (Intercept)          educ 
#> -0.0684422196 -1.1249025687  0.0458617934 -0.4607747704  0.1092064320 
#>         exper    I(exper^2)         sigma           rho 
#>  0.0469100554 -0.0008813633  3.4295872561  0.1973192668 
#> 
#> $value
#> [1] -1375.13
#> 
#> $loglik
#> [1] 1375.13
#> 
#> $counts
#> gradient 
#>       29 
#> 
#> $hessian
#>               XS(Intercept)    XSnwifeinc        XSeduc       XSexper
#> XS(Intercept)   -238.083749   -4865.42142   -2904.88466   -2509.54248
#> XSnwifeinc     -4865.421418 -135063.24692  -61206.28222  -46849.26031
#> XSeduc         -2904.884657  -61206.28222  -36643.63222  -30751.26907
#> XSexper        -2509.542480  -46849.26031  -30751.26907  -41957.75843
#> XSI(exper^2)  -42851.466743 -744867.68344 -524234.05860 -922233.93003
#> XSage         -10164.913362 -209321.43916 -123559.77600 -112841.25630
#> XSkids5          -56.132815   -1130.25127    -725.51697    -420.84153
#> XSkids618       -332.769995   -6893.76670   -4016.79185   -2677.41552
#> XO(Intercept)     45.203570     886.15188     560.37533     521.82236
#> XOeduc           560.377529   11306.46919    7178.28589    6447.60223
#> XOexper          521.829446    9512.68568    6447.65681    8949.98670
#> XOI(exper^2)    9024.120817  153929.16179  110554.85630  194413.71287
#> sigma              2.669478      47.09007      33.78183      31.99792
#> rho               12.370817     152.64164     153.43393      24.48918
#>                XSI(exper^2)         XSage       XSkids5     XSkids618
#> XS(Intercept)    -42851.467   -10164.9134   -56.1328147   -332.769995
#> XSnwifeinc      -744867.683  -209321.4392 -1130.2512675  -6893.766699
#> XSeduc          -524234.059  -123559.7760  -725.5169684  -4016.791846
#> XSexper         -922233.930  -112841.2563  -420.8415264  -2677.415517
#> XSI(exper^2)  -24415590.913 -2031228.7263 -4739.8363156 -32993.660516
#> XSage          -2031228.726  -449427.0826 -1957.8752339 -13143.220120
#> XSkids5           -4739.836    -1957.8752   -72.3171438    -94.919677
#> XSkids618        -32993.661   -13143.2201   -94.9196766   -903.074054
#> XO(Intercept)      8892.995     1918.1436     8.6099758     63.253575
#> XOeduc           108911.241    23729.4232   113.5922838    774.150771
#> XOexper          190480.877    23384.4720    68.3396231    543.418659
#> XOI(exper^2)    4805269.187   425408.2505   790.1306309   6877.654653
#> sigma               556.450      111.6992     0.5836707      3.714711
#> rho                -131.443      507.3594    10.7166028     19.309263
#>               XO(Intercept)        XOeduc       XOexper  XOI(exper^2)
#> XS(Intercept)  4.520357e+01      560.3775  5.218294e+02  9.024121e+03
#> XSnwifeinc     8.861519e+02    11306.4692  9.512686e+03  1.539292e+05
#> XSeduc         5.603753e+02     7178.2859  6.447657e+03  1.105549e+05
#> XSexper        5.218224e+02     6447.6022  8.949987e+03  1.944137e+05
#> XSI(exper^2)   8.892995e+03   108911.2408  1.904809e+05  4.805269e+06
#> XSage          1.918144e+03    23729.4232  2.338447e+04  4.254083e+05
#> XSkids5        8.609976e+00      113.5923  6.833962e+01  7.901306e+02
#> XSkids618      6.325358e+01      774.1508  5.434187e+02  6.877655e+03
#> XO(Intercept) -8.657102e+02   -10980.3842 -1.058246e+04 -1.933234e+05
#> XOeduc        -1.098038e+04  -143609.1500 -1.341087e+05 -2.433149e+06
#> XOexper       -1.058246e+04  -134108.7395 -1.851892e+05 -4.235504e+06
#> XOI(exper^2)  -1.933234e+05 -2433148.7038 -4.235504e+06 -1.147060e+08
#> sigma         -4.404924e+01     -563.8503 -5.632749e+02 -1.029695e+04
#> rho           -3.262976e+02    -3981.5125 -3.179036e+03 -5.063642e+04
#>                       sigma          rho
#> XS(Intercept)  2.669478e+00     12.37082
#> XSnwifeinc     4.709007e+01    152.64164
#> XSeduc         3.378183e+01    153.43393
#> XSexper        3.199792e+01     24.48918
#> XSI(exper^2)   5.564500e+02   -131.44297
#> XSage          1.116992e+02    507.35936
#> XSkids5        5.836707e-01     10.71660
#> XSkids618      3.714711e+00     19.30926
#> XO(Intercept) -4.404924e+01   -326.29756
#> XOeduc        -5.638503e+02  -3981.51251
#> XOexper       -5.632749e+02  -3179.03556
#> XOI(exper^2)  -1.029695e+04 -50636.41535
#> sigma         -2.041875e+01    -29.64021
#> rho           -2.964021e+01   -234.79415
#> 
#> $fisher_infoBS
#>               XS(Intercept)    XSnwifeinc        XSeduc       XSexper
#> XS(Intercept)  4.163452e-01  3.724328e-04 -1.231614e-02 -2.476946e-03
#> XSnwifeinc     3.724328e-04  3.341683e-05 -5.809189e-05  1.405027e-05
#> XSeduc        -1.231614e-02 -5.809189e-05  9.773869e-04 -6.114682e-05
#> XSexper       -2.476946e-03  1.405027e-05 -6.114682e-05  4.630791e-04
#> XSI(exper^2)   9.401491e-05 -1.166546e-08  7.235139e-07 -1.274078e-05
#> XSage         -5.419957e-03 -1.107658e-05  4.610968e-05  4.328911e-06
#> XSkids5       -3.061410e-02  3.455536e-06 -4.793473e-04  2.026045e-04
#> XSkids618     -1.538756e-02 -1.389451e-05  1.668090e-04  8.767279e-05
#> XO(Intercept)  1.276996e-02  5.690601e-05 -6.891985e-04  1.340770e-04
#> XOeduc        -7.463711e-04 -2.285162e-06  5.058765e-05 -1.053630e-05
#> XOexper       -2.059426e-04 -1.281458e-06  8.149283e-07  1.453030e-05
#> XOI(exper^2)   3.271104e-06  1.936811e-08  5.020224e-09 -4.197141e-07
#> sigma          3.104563e-03  2.791095e-05 -6.157169e-05  1.535370e-04
#> rho           -3.954513e-03 -5.074544e-05  1.323307e-04 -2.133469e-04
#>                XSI(exper^2)         XSage       XSkids5     XSkids618
#> XS(Intercept)  9.401491e-05 -5.419957e-03 -3.061410e-02 -1.538756e-02
#> XSnwifeinc    -1.166546e-08 -1.107658e-05  3.455536e-06 -1.389451e-05
#> XSeduc         7.235139e-07  4.610968e-05 -4.793473e-04  1.668090e-04
#> XSexper       -1.274078e-05  4.328911e-06  2.026045e-04  8.767279e-05
#> XSI(exper^2)   4.290347e-07 -1.038487e-06 -5.491239e-06 -3.334788e-08
#> XSage         -1.038487e-06  1.121030e-04  6.743485e-04  1.980049e-04
#> XSkids5       -5.491239e-06  6.743485e-04  2.243445e-02  7.860470e-04
#> XSkids618     -3.334788e-08  1.980049e-04  7.860470e-04  2.919259e-03
#> XO(Intercept) -2.003566e-06 -1.252026e-04 -2.432971e-03  6.279688e-05
#> XOeduc         1.811467e-07  4.956054e-06  8.239857e-05 -9.110411e-07
#> XOexper       -4.012678e-07  2.865076e-06  6.592801e-05 -5.557346e-06
#> XOI(exper^2)   1.342042e-08 -3.620876e-08 -9.315197e-07  1.205061e-07
#> sigma         -2.774219e-06 -7.855799e-05 -1.635099e-03  2.104829e-05
#> rho            3.953770e-06  9.658813e-05  2.144538e-03 -2.300693e-05
#>               XO(Intercept)        XOeduc       XOexper  XOI(exper^2)
#> XS(Intercept)  1.276996e-02 -7.463711e-04 -2.059426e-04  3.271104e-06
#> XSnwifeinc     5.690601e-05 -2.285162e-06 -1.281458e-06  1.936811e-08
#> XSeduc        -6.891985e-04  5.058765e-05  8.149283e-07  5.020224e-09
#> XSexper        1.340770e-04 -1.053630e-05  1.453030e-05 -4.197141e-07
#> XSI(exper^2)  -2.003566e-06  1.811467e-07 -4.012678e-07  1.342042e-08
#> XSage         -1.252026e-04  4.956054e-06  2.865076e-06 -3.620876e-08
#> XSkids5       -2.432971e-03  8.239857e-05  6.592801e-05 -9.315197e-07
#> XSkids618      6.279688e-05 -9.110411e-07 -5.557346e-06  1.205061e-07
#> XO(Intercept)  5.856782e-02 -3.397233e-03 -1.213138e-03  2.345059e-05
#> XOeduc        -3.397233e-03  2.511506e-04  1.705903e-06  1.639702e-07
#> XOexper       -1.213138e-03  1.705903e-06  1.614609e-04 -4.094126e-06
#> XOI(exper^2)   2.345059e-05  1.639702e-07 -4.094126e-06  1.191256e-07
#> sigma          8.997609e-03 -4.063090e-04 -4.068092e-04  7.184043e-06
#> rho           -1.365395e-02  4.605938e-04  4.045037e-04 -6.508920e-06
#>                       sigma           rho
#> XS(Intercept)  3.104563e-03 -3.954513e-03
#> XSnwifeinc     2.791095e-05 -5.074544e-05
#> XSeduc        -6.157169e-05  1.323307e-04
#> XSexper        1.535370e-04 -2.133469e-04
#> XSI(exper^2)  -2.774219e-06  3.953770e-06
#> XSage         -7.855799e-05  9.658813e-05
#> XSkids5       -1.635099e-03  2.144538e-03
#> XSkids618      2.104829e-05 -2.300693e-05
#> XO(Intercept)  8.997609e-03 -1.365395e-02
#> XOeduc        -4.063090e-04  4.605938e-04
#> XOexper       -4.068092e-04  4.045037e-04
#> XOI(exper^2)   7.184043e-06 -6.508920e-06
#> sigma          6.240435e-02 -9.616911e-03
#> rho           -9.616911e-03  1.268997e-02
#> 
#> $prop_sigmaBS
#> XS(Intercept)    XSnwifeinc        XSeduc       XSexper  XSI(exper^2) 
#>  0.6452482042  0.0057807292  0.0312631871  0.0215192725  0.0006550074 
#>         XSage       XSkids5     XSkids618 XO(Intercept)        XOeduc 
#>  0.0105878722  0.1497813481  0.0540301683  0.2420079012  0.0158477318 
#>       XOexper  XOI(exper^2)         sigma           rho 
#>  0.0127067251  0.0003451458  0.2498086234  0.1126497757 
#> 
#> $level
#> [1] "0" "1"
#> 
#> $nObs
#> [1] 753
#> 
#> $nParam
#> [1] 14
#> 
#> $N0
#> [1] 325
#> 
#> $N1
#> [1] 428
#> 
#> $NXS
#> [1] 8
#> 
#> $NXO
#> [1] 4
#> 
#> $df
#> [1] 739
#> 
#> $aic
#> [1] 2778.261
#> 
#> $bic
#> [1] 2842.997
#> 
#> $initial.value
#> XS(Intercept)    XSnwifeinc        XSeduc       XSexper  XSI(exper^2) 
#>  0.2700735726 -0.0120236370  0.1309039692  0.1233471675 -0.0018870674 
#>         XSage       XSkids5     XSkids618 XO(Intercept)        XOeduc 
#> -0.0528524416 -0.8683246795  0.0360056106 -0.5781023058  0.1090654909 
#>       XOexper  XOI(exper^2)         sigma           rho 
#>  0.0438873010 -0.0008591134  0.6637994802  0.0486011464 
#> 
#> attr(,"class")
#> [1] "HeckmanBS" "list"     

HeckmanSK(selectEq, outcomeEq, data = Mroz87, lambda = 1)

#> $coefficients
#>   (Intercept)      nwifeinc          educ         exper    I(exper^2) 
#>  0.2663350384 -0.0121320482  0.1313398741  0.1232814605 -0.0018862461 
#>           age         kids5       kids618   (Intercept)          educ 
#> -0.0528283462 -0.8673963062  0.0358745816 -0.5553913893  0.1083498518 
#>         exper    I(exper^2)         sigma           rho        lambda 
#>  0.0428353751 -0.0008373951  0.6633956023  0.0265869564  0.0051435224 
#> 
#> $value
#> [1] -832.8851
#> 
#> $loglik
#> [1] 832.8851
#> 
#> $counts
#> gradient 
#>       31 
#> 
#> $hessian
#>               XS(Intercept)    XSnwifeinc        XSeduc       XSexper
#> XS(Intercept) -3.839434e+02 -7.799189e+03 -4.687737e+03 -3.928680e+03
#> XSnwifeinc    -7.799189e+03 -2.090237e+05 -9.831712e+04 -7.473170e+04
#> XSeduc        -4.687737e+03 -9.831712e+04 -5.908392e+04 -4.777597e+04
#> XSexper       -3.928680e+03 -7.473170e+04 -4.777597e+04 -6.330552e+04
#> XSI(exper^2)  -6.317501e+04 -1.123750e+06 -7.626059e+05 -1.316495e+06
#> XSage         -1.638825e+04 -3.354107e+05 -1.994462e+05 -1.773594e+05
#> XSkids5       -9.096065e+01 -1.754743e+03 -1.183143e+03 -7.209231e+02
#> XSkids618     -5.333519e+02 -1.093638e+04 -6.438298e+03 -4.127602e+03
#> XO(Intercept)  8.205715e+00  1.617611e+02  1.015385e+02  9.429380e+01
#> XOeduc         1.015386e+02  2.064831e+03  1.299319e+03  1.160448e+03
#> XOexper        9.429380e+01  1.744692e+03  1.160447e+03  1.601411e+03
#> XOI(exper^2)   1.601409e+03  2.777785e+04  1.951460e+04  3.427190e+04
#> sigma          4.936320e-02  7.275883e+00 -3.085468e-01  4.978053e-01
#> rho            2.202930e+00 -1.010889e+02  4.657854e+01  3.858208e+00
#>               -3.801678e+00 -7.983124e+01 -4.570128e+01 -3.343441e+01
#>                XSI(exper^2)         XSage       XSkids5     XSkids618
#> XS(Intercept) -6.317501e+04 -1.638825e+04   -90.9606464 -5.333519e+02
#> XSnwifeinc    -1.123750e+06 -3.354107e+05 -1754.7434109 -1.093638e+04
#> XSeduc        -7.626059e+05 -1.994462e+05 -1183.1428466 -6.438298e+03
#> XSexper       -1.316495e+06 -1.773594e+05  -720.9230946 -4.127602e+03
#> XSI(exper^2)  -3.249887e+07 -3.001815e+06 -8118.4314413 -4.833259e+04
#> XSage         -3.001815e+06 -7.242440e+05 -3154.4414820 -2.108984e+04
#> XSkids5       -8.118431e+03 -3.154441e+03  -117.3355831 -1.449819e+02
#> XSkids618     -4.833259e+04 -2.108984e+04  -144.9819147 -1.424244e+03
#> XO(Intercept)  1.600706e+03  3.484004e+02     1.5615110  1.137736e+01
#> XOeduc         1.952090e+04  4.299773e+03    20.8026166  1.387688e+02
#> XOexper        3.423534e+04  4.239097e+03    13.0848409  9.630804e+01
#> XOI(exper^2)   8.540126e+05  7.569543e+04   153.6748441  1.190990e+03
#> sigma          6.820745e+00  3.456450e+00    -0.1504175  2.254802e-01
#> rho           -3.287902e+00  6.476978e+01     5.2320221 -2.378987e-01
#>               -4.929680e+02 -1.632514e+02    -1.1030797 -5.292520e+00
#>               XO(Intercept)        XOeduc       XOexper  XOI(exper^2)
#> XS(Intercept)  8.205715e+00  1.015386e+02  9.429380e+01  1.601409e+03
#> XSnwifeinc     1.617611e+02  2.064831e+03  1.744692e+03  2.777785e+04
#> XSeduc         1.015385e+02  1.299319e+03  1.160447e+03  1.951460e+04
#> XSexper        9.429380e+01  1.160448e+03  1.601411e+03  3.427190e+04
#> XSI(exper^2)   1.600706e+03  1.952090e+04  3.423534e+04  8.540126e+05
#> XSage          3.484004e+02  4.299773e+03  4.239097e+03  7.569543e+04
#> XSkids5        1.561511e+00  2.080262e+01  1.308484e+01  1.536748e+02
#> XSkids618      1.137736e+01  1.387688e+02  9.630804e+01  1.190990e+03
#> XO(Intercept) -9.728654e+02 -1.231529e+04 -1.268311e+04 -2.283376e+05
#> XOeduc        -1.231529e+04 -1.609660e+05 -1.602810e+05 -2.869308e+06
#> XOexper       -1.268311e+04 -1.602810e+05 -2.283376e+05 -4.988045e+06
#> XOI(exper^2)  -2.283376e+05 -2.869308e+06 -4.988045e+06 -1.243695e+08
#> sigma         -1.768870e+01 -2.169920e+02 -1.899467e+02 -3.128105e+03
#> rho           -3.421641e+02 -4.160352e+03 -3.444595e+03 -5.474556e+04
#>               -5.146964e+02 -6.515510e+03 -6.710456e+03 -1.208136e+05
#>                       sigma           rho              
#> XS(Intercept)     0.0493632  2.202930e+00 -3.801678e+00
#> XSnwifeinc        7.2758833 -1.010889e+02 -7.983124e+01
#> XSeduc           -0.3085468  4.657854e+01 -4.570128e+01
#> XSexper           0.4978053  3.858208e+00 -3.343441e+01
#> XSI(exper^2)      6.8207445 -3.287902e+00 -4.929680e+02
#> XSage             3.4564504  6.476978e+01 -1.632514e+02
#> XSkids5          -0.1504175  5.232022e+00 -1.103080e+00
#> XSkids618         0.2254802 -2.378987e-01 -5.292520e+00
#> XO(Intercept)   -17.6887048 -3.421641e+02 -5.146964e+02
#> XOeduc         -216.9920470 -4.160352e+03 -6.515510e+03
#> XOexper        -189.9466568 -3.444595e+03 -6.710456e+03
#> XOI(exper^2)  -3128.1051277 -5.474556e+04 -1.208136e+05
#> sigma         -1945.5691531  8.950718e+00 -5.135650e+00
#> rho               8.9507181 -1.966555e+02 -1.810612e+02
#>                  -5.1356500 -1.810612e+02 -2.728446e+02
#> 
#> $fisher_infoSK
#>               XS(Intercept)    XSnwifeinc        XSeduc       XSexper
#> XS(Intercept)  2.607742e-01  2.707460e-04 -7.787429e-03 -1.649631e-03
#> XSnwifeinc     2.707460e-04  2.378824e-05 -4.327614e-05  3.842568e-06
#> XSeduc        -7.787429e-03 -4.327614e-05  6.441054e-04 -1.916256e-05
#> XSexper       -1.649631e-03  3.842568e-06 -1.916256e-05  3.631697e-04
#> XSI(exper^2)   6.368457e-05  1.171386e-07  3.187663e-07 -1.079667e-05
#> XSage         -3.336645e-03 -6.448504e-06  2.056440e-05 -5.296766e-06
#> XSkids5       -1.894548e-02  2.122565e-05 -3.975518e-04 -1.555675e-05
#> XSkids618     -1.009446e-02 -9.874134e-06  1.161743e-04  6.906245e-05
#> XO(Intercept)  3.587688e-02  9.996349e-05 -4.635215e-04  7.042070e-05
#> XOeduc        -1.685233e-04 -2.824948e-06  1.602191e-05 -2.629876e-06
#> XOexper       -1.777005e-04 -4.114119e-06  1.663565e-05 -3.406032e-07
#> XOI(exper^2)   3.713389e-06  8.504490e-08 -3.418884e-07 -2.115870e-08
#> sigma         -1.537409e-04 -6.340256e-07  2.587376e-06 -5.844552e-07
#> rho           -3.419609e-03 -8.619460e-05  3.529992e-04 -7.367988e-05
#>               -5.870822e-02 -5.764860e-07  1.831933e-06 -1.422438e-06
#>                XSI(exper^2)         XSage       XSkids5     XSkids618
#> XS(Intercept)  6.368457e-05 -3.336645e-03 -1.894548e-02 -1.009446e-02
#> XSnwifeinc     1.171386e-07 -6.448504e-06  2.122565e-05 -9.874134e-06
#> XSeduc         3.187663e-07  2.056440e-05 -3.975518e-04  1.161743e-04
#> XSexper       -1.079667e-05 -5.296766e-06 -1.555675e-05  6.906245e-05
#> XSI(exper^2)   3.797807e-07 -5.027593e-07 -3.260366e-07 -3.141368e-07
#> XSage         -5.027593e-07  7.189071e-05  4.537903e-04  1.276206e-04
#> XSkids5       -3.260366e-07  4.537903e-04  1.407745e-02  6.287437e-04
#> XSkids618     -3.141368e-07  1.276206e-04  6.287437e-04  1.890101e-03
#> XO(Intercept) -9.800692e-07 -3.615505e-05 -1.063092e-03  1.222271e-04
#> XOeduc         4.464710e-08  1.037913e-06  2.953418e-05 -3.313035e-06
#> XOexper       -3.775544e-08  1.420247e-06  4.301682e-05 -5.416606e-06
#> XOI(exper^2)   1.954121e-09 -2.878579e-08 -8.761149e-07  1.143331e-07
#> sigma          9.565656e-09  2.488932e-07  7.180079e-06 -7.576980e-07
#> rho            1.192653e-06  2.989167e-05  9.097625e-04 -1.012726e-04
#>                2.437359e-08  5.591326e-07  1.248089e-05 -1.022437e-06
#>               XO(Intercept)        XOeduc       XOexper  XOI(exper^2)
#> XS(Intercept)  3.587688e-02 -1.685233e-04 -1.777005e-04  3.713389e-06
#> XSnwifeinc     9.996349e-05 -2.824948e-06 -4.114119e-06  8.504490e-08
#> XSeduc        -4.635215e-04  1.602191e-05  1.663565e-05 -3.418884e-07
#> XSexper        7.042070e-05 -2.629876e-06 -3.406032e-07 -2.115870e-08
#> XSI(exper^2)  -9.800692e-07  4.464710e-08 -3.775544e-08  1.954121e-09
#> XSage         -3.615505e-05  1.037913e-06  1.420247e-06 -2.878579e-08
#> XSkids5       -1.063092e-03  2.953418e-05  4.301682e-05 -8.761149e-07
#> XSkids618      1.222271e-04 -3.313035e-06 -5.416606e-06  1.143331e-07
#> XO(Intercept)  8.430479e-01 -3.260040e-03 -2.151335e-03  4.780054e-05
#> XOeduc        -3.260040e-03  2.208403e-04  2.220832e-05 -3.126330e-07
#> XOexper       -2.151335e-03  2.220832e-05  2.214467e-04 -5.912323e-06
#> XOI(exper^2)   4.780054e-05 -3.126330e-07 -5.912323e-06  1.743236e-07
#> sigma         -3.416738e-03  6.527188e-06  9.670545e-06 -1.998369e-07
#> rho           -2.514657e-02  6.987428e-04  1.034672e-03 -2.137560e-05
#>               -1.464421e+00  4.591159e-06  1.283234e-05 -2.977488e-07
#>                       sigma           rho              
#> XS(Intercept) -1.537409e-04 -3.419609e-03 -5.870822e-02
#> XSnwifeinc    -6.340256e-07 -8.619460e-05 -5.764860e-07
#> XSeduc         2.587376e-06  3.529992e-04  1.831933e-06
#> XSexper       -5.844552e-07 -7.367988e-05 -1.422438e-06
#> XSI(exper^2)   9.565656e-09  1.192653e-06  2.437359e-08
#> XSage          2.488932e-07  2.989167e-05  5.591327e-07
#> XSkids5        7.180079e-06  9.097625e-04  1.248090e-05
#> XSkids618     -7.576980e-07 -1.012726e-04 -1.022437e-06
#> XO(Intercept) -3.416738e-03 -2.514657e-02 -1.464421e+00
#> XOeduc         6.527188e-06  6.987428e-04  4.591159e-06
#> XOexper        9.670545e-06  1.034672e-03  1.283234e-05
#> XOI(exper^2)  -1.998369e-07 -2.137560e-05 -2.977488e-07
#> sigma          5.286648e-04  1.820265e-04  6.011182e-03
#> rho            1.820265e-04  2.164031e-02  4.055325e-04
#>                6.011182e-03  4.055325e-04  2.766304e+00
#> 
#> $prop_sigmaSK
#> XS(Intercept)    XSnwifeinc        XSeduc       XSexper  XSI(exper^2) 
#>  0.5106605347  0.0048773185  0.0253792318  0.0190570106  0.0006162635 
#>         XSage       XSkids5     XSkids618 XO(Intercept)        XOeduc 
#>  0.0084788387  0.1186484409  0.0434752885  0.9181764031  0.0148606968 
#>       XOexper  XOI(exper^2)         sigma           rho               
#>  0.0148810852  0.0004175207  0.0229927113  0.1471064574  1.6632210988 
#> 
#> $level
#> [1] "0" "1"
#> 
#> $nObs
#> [1] 753
#> 
#> $nParam
#> [1] 15
#> 
#> $N0
#> [1] 325
#> 
#> $N1
#> [1] 428
#> 
#> $NXS
#> [1] 8
#> 
#> $NXO
#> [1] 4
#> 
#> $df
#> [1] 738
#> 
#> $aic
#> [1] 1695.77
#> 
#> $bic
#> [1] 1765.131
#> 
#> $initial.value
#> XS(Intercept)    XSnwifeinc        XSeduc       XSexper  XSI(exper^2) 
#>  0.2700735726 -0.0120236370  0.1309039692  0.1233471675 -0.0018870674 
#>         XSage       XSkids5     XSkids618 XO(Intercept)        XOeduc 
#> -0.0528524416 -0.8683246795  0.0360056106 -0.5781023058  0.1090654909 
#>       XOexper  XOI(exper^2)         sigma           rho               
#>  0.0438873010 -0.0008591134  0.6637994802  0.0486011464  1.0000000000 
#> 
#> attr(,"class")
#> [1] "HeckmanSK" "list"     

HeckmantS(selectEq, outcomeEq, data = Mroz87, df=5)

#> Warning: NaNs produzidos

#> Warning: NaNs produzidos

#> Warning: NaNs produzidos

#> Warning: NaNs produzidos

#> Warning: NaNs produzidos

#> Warning: NaNs produzidos

#> $coefficients
#>   (Intercept)      nwifeinc          educ         exper    I(exper^2) 
#>  0.4978609360 -0.0119468890  0.1395750239  0.1496618759 -0.0022845408 
#>           age         kids5       kids618   (Intercept)          educ 
#> -0.0642790009 -1.0438382682  0.0344075726 -0.2524442781  0.1084650789 
#>         exper    I(exper^2)         sigma           rho            df 
#>  0.0251406451 -0.0004567222  0.4437009922 -0.3825537027  3.0611213085 
#> 
#> $value
#> [1] -788.0048
#> 
#> $loglik
#> [1] 788.0048
#> 
#> $counts
#> gradient 
#>       23 
#> 
#> $hessian
#>               XS(Intercept)    XSnwifeinc        XSeduc       XSexper
#> XS(Intercept)   -276.279889   -5676.89800   -3370.17261   -2745.78741
#> XSnwifeinc     -5676.897996 -152129.18872  -71581.70592  -54076.69840
#> XSeduc         -3370.172614  -71581.70592  -42424.72414  -33188.52498
#> XSexper        -2745.787406  -54076.69840  -33188.52498  -42778.40650
#> XSI(exper^2)  -42409.625635 -783271.72339 -506697.63501 -866851.38998
#> XSage         -11808.440930 -244825.97584 -143597.72247 -124432.56229
#> XSkids5          -64.189661   -1198.67593    -838.69899    -549.19385
#> XSkids618       -377.064256   -7769.45217   -4533.27928   -2790.41094
#> XO(Intercept)   -123.631966   -2455.25564   -1535.38653   -1428.43246
#> XOeduc         -1535.329286  -31841.27646  -19621.99579  -17644.23223
#> XOexper        -1428.306982  -27455.03826  -17643.51185  -24243.46636
#> XOI(exper^2)  -22514.180044 -413077.85500 -275652.74970 -464145.01948
#> sigma             41.564271     848.93668     477.45835     145.17510
#> rho              -33.248443    -858.65335    -387.63244    -327.09601
#>                   -5.291253     -90.14072     -66.73986     -86.32669
#>                XSI(exper^2)         XSage       XSkids5     XSkids618
#> XS(Intercept) -4.240963e+04   -11808.4409   -64.1896612   -377.064256
#> XSnwifeinc    -7.832717e+05  -244825.9758 -1198.6759299  -7769.452170
#> XSeduc        -5.066976e+05  -143597.7225  -838.6989926  -4533.279283
#> XSexper       -8.668514e+05  -124432.5623  -549.1938456  -2790.410941
#> XSI(exper^2)  -2.065509e+07 -2021786.9256 -6419.7179662 -31083.726758
#> XSage         -2.021787e+06  -522088.1862 -2216.3303243 -14985.014873
#> XSkids5       -6.419718e+03    -2216.3303   -81.8098089    -91.095572
#> XSkids618     -3.108373e+04   -14985.0149   -91.0955722   -980.878858
#> XO(Intercept) -2.383508e+04    -5243.9968   -22.7479482   -169.136936
#> XOeduc        -2.910894e+05   -64987.7790  -298.3629265  -2059.997034
#> XOexper       -5.029150e+05   -64592.8359  -219.2311093  -1387.282345
#> XOI(exper^2)  -1.062248e+07 -1058417.2988 -2846.1763689 -16543.637657
#> sigma         -1.021381e+02     1794.6487    21.3015373     69.263258
#> rho           -5.601655e+03    -1466.8375    -7.4992960    -48.007271
#>               -1.691839e+03     -223.1172     0.1008317     -7.400591
#>               XO(Intercept)        XOeduc       XOexper  XOI(exper^2)
#> XS(Intercept) -1.236320e+02    -1535.3293    -1428.3070 -2.251418e+04
#> XSnwifeinc    -2.455256e+03   -31841.2765   -27455.0383 -4.130779e+05
#> XSeduc        -1.535387e+03   -19621.9958   -17643.5118 -2.756527e+05
#> XSexper       -1.428432e+03   -17644.2322   -24243.4664 -4.641450e+05
#> XSI(exper^2)  -2.383508e+04  -291089.3860  -502915.0381 -1.062248e+07
#> XSage         -5.243997e+03   -64987.7790   -64592.8359 -1.058417e+06
#> XSkids5       -2.274795e+01     -298.3629     -219.2311 -2.846176e+03
#> XSkids618     -1.691369e+02    -2059.9970    -1387.2823 -1.654364e+04
#> XO(Intercept) -1.603049e+03   -20225.7308   -21948.5227 -3.656782e+05
#> XOeduc        -2.022573e+04  -263234.3212  -276185.6292 -4.578532e+06
#> XOexper       -2.194852e+04  -276185.6292  -399817.8712 -7.711990e+06
#> XOI(exper^2)  -3.656782e+05 -4578531.7428 -7711990.1335 -1.549952e+08
#> sigma          3.640734e+01      320.6567     -175.8686 -4.252300e+03
#> rho           -3.733093e+02    -4533.1794    -3949.9510 -6.345921e+04
#>               -8.901774e+00     -118.6308     -124.2653 -1.970930e+03
#>                     sigma           rho              
#> XS(Intercept)    41.56427    -33.248443    -5.2912530
#> XSnwifeinc      848.93668   -858.653347   -90.1407190
#> XSeduc          477.45835   -387.632441   -66.7398624
#> XSexper         145.17510   -327.096006   -86.3266925
#> XSI(exper^2)   -102.13813  -5601.655257 -1691.8385443
#> XSage          1794.64866  -1466.837458  -223.1172273
#> XSkids5          21.30154     -7.499296     0.1008317
#> XSkids618        69.26326    -48.007271    -7.4005910
#> XO(Intercept)    36.40734   -373.309330    -8.9017744
#> XOeduc          320.65666  -4533.179426  -118.6308043
#> XOexper        -175.86861  -3949.951008  -124.2652960
#> XOI(exper^2)  -4252.30003 -63459.211467 -1970.9298784
#> sigma         -1024.21124    -65.572109    52.6782735
#> rho             -65.57211   -160.708893     4.7831234
#>                  52.67827      4.783123    -7.0565100
#> 
#> $fisher_infotS
#>               XS(Intercept)    XSnwifeinc        XSeduc       XSexper
#> XS(Intercept)  3.788323e-01  4.300820e-04 -1.126833e-02 -2.446954e-03
#> XSnwifeinc     4.300820e-04  3.465533e-05 -6.662903e-05 -1.401856e-05
#> XSeduc        -1.126833e-02 -6.662903e-05  9.595737e-04  4.867902e-05
#> XSexper       -2.446954e-03 -1.401856e-05  4.867902e-05  7.451617e-04
#> XSI(exper^2)   9.607302e-05  7.033272e-07 -1.193874e-06 -2.260065e-05
#> XSage         -4.879588e-03 -6.887225e-06  1.304668e-05 -4.248193e-05
#> XSkids5       -2.973245e-02  8.714184e-05 -8.552571e-04 -6.644081e-04
#> XSkids618     -1.557904e-02 -1.679540e-05  2.020970e-04  1.386587e-04
#> XO(Intercept) -6.771346e-03  3.006379e-05  6.049990e-04 -3.415546e-04
#> XOeduc         6.723054e-04 -4.547400e-06 -3.944619e-05  1.305610e-05
#> XOexper       -4.036150e-04  1.120199e-05 -4.116268e-05  2.708583e-07
#> XOI(exper^2)   1.710373e-05 -4.566285e-07  1.649845e-06  1.444797e-07
#> sigma         -6.749056e-04  1.217992e-05 -1.072031e-04 -1.997086e-04
#> rho           -1.443353e-03 -8.395560e-05  2.517242e-04  4.972278e-04
#>               -1.232842e-02  1.072962e-04 -1.186627e-03 -1.943165e-03
#>                XSI(exper^2)         XSage       XSkids5     XSkids618
#> XS(Intercept)  9.607302e-05 -4.879588e-03 -2.973245e-02 -1.557904e-02
#> XSnwifeinc     7.033272e-07 -6.887225e-06  8.714184e-05 -1.679540e-05
#> XSeduc        -1.193874e-06  1.304668e-05 -8.552571e-04  2.020970e-04
#> XSexper       -2.260065e-05 -4.248193e-05 -6.644081e-04  1.386587e-04
#> XSI(exper^2)   7.842362e-07  1.522063e-07  1.544261e-05 -1.465862e-06
#> XSage          1.522063e-07  1.123224e-04  8.307862e-04  1.910053e-04
#> XSkids5        1.544261e-05  8.307862e-04  2.268361e-02  1.261896e-03
#> XSkids618     -1.465862e-06  1.910053e-04  1.261896e-03  2.830404e-03
#> XO(Intercept)  1.381655e-05 -2.836660e-06  4.597487e-05 -2.839333e-05
#> XOeduc        -4.068438e-07 -3.571677e-06 -5.706934e-05  3.151376e-06
#> XOexper       -2.667300e-07  1.527107e-05  2.322106e-04 -4.131314e-07
#> XOI(exper^2)  -1.302426e-09 -6.204294e-07 -9.467217e-06  4.312753e-08
#> sigma          4.218231e-06  6.619023e-05  1.133724e-03  6.176752e-06
#> rho           -1.815825e-05 -5.140153e-05 -9.770105e-04 -9.749775e-05
#>                2.932867e-05  7.940525e-04  1.368609e-02 -3.915545e-04
#>               XO(Intercept)        XOeduc       XOexper  XOI(exper^2)
#> XS(Intercept) -6.771346e-03  6.723054e-04 -4.036150e-04  1.710373e-05
#> XSnwifeinc     3.006379e-05 -4.547400e-06  1.120199e-05 -4.566285e-07
#> XSeduc         6.049990e-04 -3.944619e-05 -4.116268e-05  1.649845e-06
#> XSexper       -3.415546e-04  1.305610e-05  2.708583e-07  1.444797e-07
#> XSI(exper^2)   1.381655e-05 -4.068438e-07 -2.667300e-07 -1.302426e-09
#> XSage         -2.836660e-06 -3.571677e-06  1.527107e-05 -6.204294e-07
#> XSkids5        4.597487e-05 -5.706934e-05  2.322106e-04 -9.467217e-06
#> XSkids618     -2.839333e-05  3.151376e-06 -4.131314e-07  4.312753e-08
#> XO(Intercept)  1.686422e-02 -1.787067e-03  1.551650e-03 -6.348620e-05
#> XOeduc        -1.787067e-03  1.386079e-04 -1.695025e-05  8.238775e-07
#> XOexper        1.551650e-03 -1.695025e-05 -1.884593e-04  6.722046e-06
#> XOI(exper^2)  -6.348620e-05  8.238775e-07  6.722046e-06 -2.230080e-07
#> sigma          3.856769e-05 -5.328380e-05  1.462318e-04 -5.809670e-06
#> rho           -1.877314e-03  3.509341e-04 -1.221758e-03  5.009474e-05
#>               -1.850548e-03 -1.936741e-04  1.627219e-04 -3.880025e-06
#>                       sigma           rho              
#> XS(Intercept) -6.749056e-04 -1.443353e-03 -1.232842e-02
#> XSnwifeinc     1.217992e-05 -8.395560e-05  1.072962e-04
#> XSeduc        -1.072031e-04  2.517242e-04 -1.186627e-03
#> XSexper       -1.997086e-04  4.972278e-04 -1.943165e-03
#> XSI(exper^2)   4.218231e-06 -1.815825e-05  2.932867e-05
#> XSage          6.619023e-05 -5.140153e-05  7.940525e-04
#> XSkids5        1.133724e-03 -9.770105e-04  1.368609e-02
#> XSkids618      6.176752e-06 -9.749775e-05 -3.915545e-04
#> XO(Intercept)  3.856769e-05 -1.877314e-03 -1.850548e-03
#> XOeduc        -5.328380e-05  3.509341e-04 -1.936741e-04
#> XOexper        1.462318e-04 -1.221758e-03  1.627219e-04
#> XOI(exper^2)  -5.809670e-06  5.009474e-05 -3.880025e-06
#> sigma          1.692617e-03 -2.545606e-04  1.307093e-02
#> rho           -2.545606e-04  1.163042e-02  9.733140e-03
#>                1.307093e-02  9.733140e-03  2.610326e-01
#> 
#> $prop_sigmatS
#> XS(Intercept)    XSnwifeinc        XSeduc       XSexper  XSI(exper^2) 
#>  0.6154935149  0.0058868774  0.0309769866  0.0272976502  0.0008855711 
#>         XSage       XSkids5     XSkids618 XO(Intercept)        XOeduc 
#>  0.0105982265  0.1506108063  0.0532015371  0.1298623109  0.0117731861 
#>       XOexper  XOI(exper^2)         sigma           rho               
#>           NaN           NaN  0.0411414228  0.1078444373  0.5109135270 
#> 
#> $level
#> [1] "0" "1"
#> 
#> $nObs
#> [1] 753
#> 
#> $nParam
#> [1] 15
#> 
#> $N0
#> [1] 325
#> 
#> $N1
#> [1] 428
#> 
#> $NXS
#> [1] 8
#> 
#> $NXO
#> [1] 4
#> 
#> $df
#> [1] 738
#> 
#> $aic
#> [1] 1606.01
#> 
#> $bic
#> [1] 1675.371
#> 
#> $initial.value
#> XS(Intercept)    XSnwifeinc        XSeduc       XSexper  XSI(exper^2) 
#>  0.2700735726 -0.0120236370  0.1309039692  0.1233471675 -0.0018870674 
#>         XSage       XSkids5     XSkids618 XO(Intercept)        XOeduc 
#> -0.0528524416 -0.8683246795  0.0360056106 -0.5781023058  0.1090654909 
#>       XOexper  XOI(exper^2)         sigma           rho               
#>  0.0438873010 -0.0008591134  0.6637994802  0.0486011464  5.0000000000 
#> 
#> attr(,"class")
#> [1] "HeckmantS" "list"     

HeckmanGe(selectEq, outcomeEq, outcomeS, outcomeC, data = Mroz87)

#> $coefficients
#>   (Intercept)      nwifeinc          educ         exper    I(exper^2) 
#>  0.3083446512 -0.0108256215  0.1288475059  0.1213859436 -0.0018564658 
#>           age         kids5       kids618   (Intercept)          educ 
#> -0.0531769095 -0.8485347564  0.0262167847 -0.8001492248  0.1461329093 
#>         exper    I(exper^2)    interceptS          educ         exper 
#>  0.0191510501 -0.0002680912 -0.3911809963  0.0180533864 -0.0199367566 
#>    interceptC          educ         exper 
#>  1.4413106688 -0.1333918941  0.0064362287 
#> 
#> $value
#> [1] -818.5952
#> 
#> $loglik
#> [1] 818.5952
#> 
#> $counts
#> gradient 
#>       33 
#> 
#> $hessian
#>               (Intercept)      nwifeinc          educ         exper
#> (Intercept)   -398.551080 -8.204773e+03   -4890.50447 -4.066634e+03
#> nwifeinc     -8204.773050 -2.259989e+05 -104444.34984 -7.796831e+04
#> educ         -4890.504470 -1.044443e+05  -62149.08415 -4.951590e+04
#> exper        -4066.633838 -7.796831e+04  -49515.89890 -6.534474e+04
#> I(exper^2)  -65274.034940 -1.166368e+06 -785755.22235 -1.359202e+06
#> age         -16985.839303 -3.517924e+05 -207425.72688 -1.833351e+05
#> kids5          -97.233757 -1.930903e+03   -1283.49167 -7.581414e+02
#> kids618       -547.060304 -1.144307e+04   -6637.77093 -4.197408e+03
#> (Intercept)    -35.797611 -1.113305e+03    -683.32312 -2.304140e+02
#> educ          -683.322940 -1.921839e+04  -11563.55198 -5.989302e+03
#> exper         -230.419589 -9.306608e+03   -5989.34015 -4.491184e+02
#> I(exper^2)    -621.027325 -1.028417e+05  -67999.02688  5.337552e+04
#> interceptS       1.341774  3.226102e+01     -10.44589 -3.650596e+00
#> educ           -10.455900 -1.528122e+02    -451.48188 -4.151298e+02
#> exper           -3.666178 -5.648848e+02    -415.20649 -1.035252e+01
#> interceptC     -11.996483 -6.425779e+02    -229.86996 -5.468326e+01
#> educ          -229.855324 -1.024875e+04   -3843.54775 -1.214391e+03
#> exper          -54.674544 -4.461255e+03   -1214.44571 -5.547303e-01
#>                I(exper^2)           age        kids5       kids618
#> (Intercept) -6.527403e+04 -1.698584e+04   -97.233757   -547.060304
#> nwifeinc    -1.166368e+06 -3.517924e+05 -1930.902865 -11443.067948
#> educ        -7.857552e+05 -2.074257e+05 -1283.491668  -6637.770925
#> exper       -1.359202e+06 -1.833351e+05  -758.141396  -4197.407696
#> I(exper^2)  -3.364361e+07 -3.097877e+06 -8442.699915 -48912.865482
#> age         -3.097877e+06 -7.495459e+05 -3362.538412 -21617.569209
#> kids5       -8.442700e+03 -3.362538e+03  -125.705555   -151.913491
#> kids618     -4.891287e+04 -2.161757e+04  -151.913491  -1458.294931
#> (Intercept) -6.247448e+02 -1.355571e+03   -16.922933    -49.785181
#> educ        -6.793196e+04 -2.729628e+04  -270.673631   -849.002371
#> exper        5.324676e+04 -7.827096e+03  -135.797924   -389.972049
#> I(exper^2)   2.580724e+06  1.299966e+04 -1455.206897  -4356.606907
#> interceptS   3.447092e+01  7.014813e+01     1.313179      7.361349
#> educ        -6.286345e+03 -1.658559e+02    12.926206     77.140279
#> exper        7.440769e+03  1.761447e+02    17.745501     45.822980
#> interceptC   3.089743e+02 -4.852040e+02    -9.548804    -19.574461
#> educ        -5.982054e+03 -9.388902e+03  -169.400404   -306.692928
#> exper        6.193264e+03 -9.140686e+02   -61.064351   -107.881597
#>               (Intercept)          educ         exper    I(exper^2)
#> (Intercept)     -35.79761     -683.3229 -2.304196e+02 -6.210273e+02
#> nwifeinc      -1113.30525   -19218.3894 -9.306608e+03 -1.028417e+05
#> educ           -683.32312   -11563.5520 -5.989340e+03 -6.799903e+04
#> exper          -230.41397    -5989.3019 -4.491184e+02  5.337552e+04
#> I(exper^2)     -624.74478   -67931.9582  5.324676e+04  2.580724e+06
#> age           -1355.57113   -27296.2827 -7.827096e+03  1.299966e+04
#> kids5           -16.92293     -270.6736 -1.357979e+02 -1.455207e+03
#> kids618         -49.78518     -849.0024 -3.899720e+02 -4.356607e+03
#> (Intercept)   -1108.51126   -13823.4203 -1.758825e+04 -3.686964e+05
#> educ         -13823.42025  -178878.6520 -2.173520e+05 -4.511726e+06
#> exper        -17588.25447  -217352.0356 -3.686497e+05 -9.090340e+06
#> I(exper^2)  -368696.35171 -4511725.6079 -9.090340e+06 -2.492372e+08
#> interceptS       23.47509      508.0138  4.770057e+01 -4.000514e+02
#> educ            508.10304     8519.1350  4.522912e+03  6.898470e+04
#> exper            47.52839     4521.4703 -2.567228e+03 -1.678541e+05
#> interceptC     -319.65106    -3822.9037 -3.565225e+03 -6.682114e+04
#> educ          -3822.93903   -47392.4321 -4.218923e+04 -7.675729e+05
#> exper         -3565.20571   -42189.4617 -6.179399e+04 -1.538446e+06
#>                interceptS          educ         exper    interceptC
#> (Intercept)      1.341774     -10.45590 -3.666178e+00    -11.996483
#> nwifeinc        32.261016    -152.81216 -5.648848e+02   -642.577852
#> educ           -10.445889    -451.48188 -4.152065e+02   -229.869963
#> exper           -3.650596    -415.12978 -1.035252e+01    -54.683258
#> I(exper^2)      34.470923   -6286.34465  7.440769e+03    308.974261
#> age             70.148126    -165.85591  1.761447e+02   -485.204008
#> kids5            1.313179      12.92621  1.774550e+01     -9.548804
#> kids618          7.361349      77.14028  4.582298e+01    -19.574461
#> (Intercept)     23.475094     508.10304  4.752839e+01   -319.651057
#> educ           508.013844    8519.13499  4.521470e+03  -3822.903725
#> exper           47.700571    4522.91228 -2.567228e+03  -3565.224912
#> I(exper^2)    -400.051421   68984.69771 -1.678541e+05 -66821.139518
#> interceptS    -868.818800  -10986.88215 -1.129482e+04    -28.317283
#> educ        -10986.882154 -143119.43838 -1.423573e+05   -452.919732
#> exper       -11294.815498 -142357.28120 -2.161494e+05   -235.501586
#> interceptC     -28.317283    -452.91973 -2.355016e+02   -169.712551
#> educ          -452.850421   -7060.81448 -4.304506e+03  -1973.286726
#> exper         -235.352522   -4303.37553 -1.838144e+03  -1234.054328
#>                     educ         exper
#> (Intercept)    -229.8553 -5.467454e+01
#> nwifeinc     -10248.7472 -4.461255e+03
#> educ          -3843.5478 -1.214446e+03
#> exper         -1214.3908 -5.547303e-01
#> I(exper^2)    -5982.0536  6.193264e+03
#> age           -9388.9022 -9.140686e+02
#> kids5          -169.4004 -6.106435e+01
#> kids618        -306.6929 -1.078816e+02
#> (Intercept)   -3822.9390 -3.565206e+03
#> educ         -47392.4321 -4.218946e+04
#> exper        -42189.2324 -6.179399e+04
#> I(exper^2)  -767572.9156 -1.538446e+06
#> interceptS     -452.8504 -2.353525e+02
#> educ          -7060.8145 -4.303376e+03
#> exper         -4304.5059 -1.838144e+03
#> interceptC    -1973.2867 -1.234054e+03
#> educ         -23807.0100 -1.413797e+04
#> exper        -14137.9709 -1.991912e+04
#> 
#> $fisher_infoHG
#>               (Intercept)      nwifeinc          educ         exper
#> (Intercept)  2.603676e-01  4.344265e-04 -7.935599e-03 -1.747633e-03
#> nwifeinc     4.344265e-04  3.180715e-05 -7.433471e-05  3.544360e-06
#> educ        -7.935599e-03 -7.433471e-05  6.958936e-04 -1.585502e-05
#> exper       -1.747633e-03  3.544360e-06 -1.585502e-05  3.646826e-04
#> I(exper^2)   6.802001e-05  2.557944e-07 -2.308316e-07 -1.066377e-05
#> age         -3.392467e-03 -6.153114e-06  2.633900e-05 -5.427745e-06
#> kids5       -1.735109e-02  1.295733e-04 -7.182041e-04 -4.650714e-05
#> kids618     -1.014705e-02 -1.380988e-05  1.329252e-04  9.090650e-05
#> (Intercept)  2.315337e-02  1.993647e-03 -6.457762e-03  1.238171e-03
#> educ        -3.237517e-04 -4.211509e-05  1.250750e-04 -6.410923e-05
#> exper       -1.488057e-03 -1.160957e-04  3.835710e-04 -4.169231e-05
#> I(exper^2)   4.009754e-05  2.767688e-06 -9.174747e-06  9.661756e-07
#> interceptS  -7.707261e-03 -6.150549e-04  1.971320e-03 -3.873692e-05
#> educ         8.633239e-04  6.585313e-05 -2.129236e-04  5.108854e-06
#> exper       -1.430005e-04 -1.027442e-05  3.330627e-05 -3.991563e-06
#> interceptC   1.186882e-02 -4.880488e-04  1.524963e-03 -2.648500e-03
#> educ        -1.537696e-03 -5.174707e-05  1.571378e-04  1.980368e-04
#> exper       -1.708242e-03 -6.402566e-05  2.194319e-04 -6.321165e-06
#>                I(exper^2)           age         kids5       kids618
#> (Intercept)  6.802001e-05 -3.392467e-03 -1.735109e-02 -1.014705e-02
#> nwifeinc     2.557944e-07 -6.153114e-06  1.295733e-04 -1.380988e-05
#> educ        -2.308316e-07  2.633900e-05 -7.182041e-04  1.329252e-04
#> exper       -1.066377e-05 -5.427745e-06 -4.650714e-05  9.090650e-05
#> I(exper^2)   3.721558e-07 -4.997887e-07  1.643267e-06 -9.428348e-07
#> age         -4.997887e-07  7.158041e-05  4.553776e-04  1.250069e-04
#> kids5        1.643267e-06  4.553776e-04  1.475623e-02  5.743327e-04
#> kids618     -9.428348e-07  1.250069e-04  5.743327e-04  1.848135e-03
#> (Intercept)  5.139493e-06 -6.578000e-05  1.581651e-02  8.984588e-04
#> educ         5.822582e-07  6.127358e-06 -1.759993e-04 -4.502597e-05
#> exper       -8.686285e-07  9.034131e-07 -1.047899e-03 -3.011847e-05
#> I(exper^2)   2.354161e-08 -1.147500e-07  2.422392e-05  5.887100e-07
#> interceptS  -9.033616e-06 -9.475953e-06 -6.006745e-03 -1.243908e-04
#> educ         9.209892e-07  7.386249e-07  6.408295e-04  1.106387e-05
#> exper       -6.598222e-08  4.462253e-07 -8.649722e-05 -2.299114e-06
#> interceptC   4.985871e-05  7.525703e-05  3.803715e-03 -2.811501e-03
#> educ        -4.857417e-06 -1.660003e-05 -1.265148e-03  2.031383e-04
#> exper       -1.094985e-06  1.545757e-05 -4.817376e-04  1.685328e-05
#>               (Intercept)          educ         exper    I(exper^2)
#> (Intercept)  2.315337e-02 -3.237517e-04 -1.488057e-03  4.009754e-05
#> nwifeinc     1.993647e-03 -4.211509e-05 -1.160957e-04  2.767688e-06
#> educ        -6.457762e-03  1.250750e-04  3.835710e-04 -9.174747e-06
#> exper        1.238171e-03 -6.410923e-05 -4.169231e-05  9.661756e-07
#> I(exper^2)   5.139493e-06  5.822582e-07 -8.686285e-07  2.354161e-08
#> age         -6.578000e-05  6.127358e-06  9.034131e-07 -1.147500e-07
#> kids5        1.581651e-02 -1.759993e-04 -1.047899e-03  2.422392e-05
#> kids618      8.984588e-04 -4.502597e-05 -3.011847e-05  5.887100e-07
#> (Intercept)  5.088791e-01 -1.503322e-02 -2.610571e-02  6.215133e-04
#> educ        -1.503322e-02  6.028516e-04  6.114560e-04 -1.469930e-05
#> exper       -2.610571e-02  6.114560e-04  1.539787e-03 -3.690323e-05
#> I(exper^2)   6.215133e-04 -1.469930e-05 -3.690323e-05  8.974179e-07
#> interceptS  -1.270050e-01  3.238028e-03  6.721062e-03 -1.609833e-04
#> educ         1.347412e-02 -3.312055e-04 -7.327796e-04  1.753561e-05
#> exper       -2.287959e-03  5.816112e-05  1.276389e-04 -3.067672e-06
#> interceptC  -3.150998e-01  1.321302e-02  1.270930e-02 -3.038808e-04
#> educ         6.903637e-03 -6.585675e-04  3.720231e-05 -1.768544e-07
#> exper       -1.299927e-02  3.324039e-04  7.139309e-04 -1.782301e-05
#>                interceptS          educ         exper    interceptC
#> (Intercept) -7.707261e-03  8.633239e-04 -1.430005e-04  1.186882e-02
#> nwifeinc    -6.150549e-04  6.585313e-05 -1.027442e-05 -4.880488e-04
#> educ         1.971320e-03 -2.129236e-04  3.330627e-05  1.524963e-03
#> exper       -3.873692e-05  5.108854e-06 -3.991563e-06 -2.648500e-03
#> I(exper^2)  -9.033616e-06  9.209892e-07 -6.598222e-08  4.985871e-05
#> age         -9.475953e-06  7.386249e-07  4.462253e-07  7.525703e-05
#> kids5       -6.006745e-03  6.408295e-04 -8.649722e-05  3.803715e-03
#> kids618     -1.243908e-04  1.106387e-05 -2.299114e-06 -2.811501e-03
#> (Intercept) -1.270050e-01  1.347412e-02 -2.287959e-03 -3.150998e-01
#> educ         3.238028e-03 -3.312055e-04  5.816112e-05  1.321302e-02
#> exper        6.721062e-03 -7.327796e-04  1.276389e-04  1.270930e-02
#> I(exper^2)  -1.609833e-04  1.753561e-05 -3.067672e-06 -3.038808e-04
#> interceptS   8.157542e-02 -7.135222e-03  4.050405e-04  4.977729e-02
#> educ        -7.135222e-03  6.688778e-04 -6.405069e-05 -5.200294e-03
#> exper        4.050405e-04 -6.405069e-05  2.542208e-05  1.171812e-03
#> interceptC   4.977729e-02 -5.200294e-03  1.171812e-03  4.928443e-01
#> educ         1.238237e-03 -1.463538e-04 -3.415627e-06 -3.191357e-02
#> exper        4.089712e-03 -4.253878e-04  6.471656e-05  4.999540e-03
#>                      educ         exper
#> (Intercept) -1.537696e-03 -1.708242e-03
#> nwifeinc    -5.174707e-05 -6.402566e-05
#> educ         1.571378e-04  2.194319e-04
#> exper        1.980368e-04 -6.321165e-06
#> I(exper^2)  -4.857417e-06 -1.094985e-06
#> age         -1.660003e-05  1.545757e-05
#> kids5       -1.265148e-03 -4.817376e-04
#> kids618      2.031383e-04  1.685328e-05
#> (Intercept)  6.903637e-03 -1.299927e-02
#> educ        -6.585675e-04  3.324039e-04
#> exper        3.720231e-05  7.139309e-04
#> I(exper^2)  -1.768544e-07 -1.782301e-05
#> interceptS   1.238237e-03  4.089712e-03
#> educ        -1.463538e-04 -4.253878e-04
#> exper       -3.415627e-06  6.471656e-05
#> interceptC  -3.191357e-02  4.999540e-03
#> educ         2.838775e-03  4.531140e-05
#> exper        4.531140e-05  5.362827e-04
#> 
#> $prop_sigmaHG
#>  (Intercept)     nwifeinc         educ        exper   I(exper^2)          age 
#> 0.5102623139 0.0056397828 0.0263797951 0.0190966653 0.0006100457 0.0084605205 
#>        kids5      kids618  (Intercept)         educ        exper   I(exper^2) 
#> 0.1214752278 0.0429899420 0.7133576442 0.0245530362 0.0392401172 0.0009473214 
#>   interceptS         educ        exper   interceptC         educ        exper 
#> 0.2856141132 0.0258626711 0.0050420311 0.7020287094 0.0532801600 0.0231577786 
#> 
#> $level
#> [1] "0" "1"
#> 
#> $nObs
#> [1] 753
#> 
#> $nParam
#> [1] 18
#> 
#> $N0
#> [1] 325
#> 
#> $N1
#> [1] 428
#> 
#> $NXS
#> [1] 8
#> 
#> $NXO
#> [1] 4
#> 
#> $df
#> [1] 735
#> 
#> $aic
#> [1] 1673.19
#> 
#> $bic
#> [1] 1756.424
#> 
#> $initial.value
#>   (Intercept)      nwifeinc          educ         exper    I(exper^2) 
#>  0.2700735726 -0.0120236370  0.1309039692  0.1233471675 -0.0018870674 
#>           age         kids5       kids618   (Intercept)          educ 
#> -0.0528524416 -0.8683246795  0.0360056106 -0.5781023058  0.1090654909 
#>         exper    I(exper^2)    interceptS          educ         exper 
#>  0.0438873010 -0.0008591134  0.6637994802  1.0000000000  1.0000000000 
#>    interceptC          educ         exper 
#>  0.0486011464  0.0000000000  0.0000000000 
#> 
#> $NE
#> [1] 3
#> 
#> $NV
#> [1] 3
#> 
#> attr(,"class")
#> [1] "HeckmanGe" "list"