RAND Health Insurance Experiment

'The RAND Health Insurance Experiment (RAND HIE) was a comprehensive study of health care cost, utilization and outcome in the United States. It is the only randomized study of health insurance, and the only study which can give definitive evidence as to the causal effects of different health insurance plans. For more information about the database visit: https://en.wikipedia.org/w/index.php?title=RAND_Health_Insurance_Experiment&oldid=110166949 accessed september 09, 2019). This data frame contains the following columns:

plan: HIE plan number.

site: Participant's place of residence when the participant was initially enrolled.

coins: Coinsurance rate.

tookphys: Took baseline physical.

year: Study year.

zper: Person identifier.

black: 1 if race of household head is black.

income: Family income.

xage: Age in years.

female: 1 if person is female.

educdec: Education of household head in years.

time: Time eligible during the year.

outpdol: Outpatient expenses: all covered outpatient medical services excluding dental care, outpatient psychotherapy, outpatient drugs or supplies.

drugdol: Drug expenses: all covered outpatient and dental drugs.

suppdol: Supply expenses: all covered outpatient supplies including dental.

mentdol: Psychotherapy expenses: all covered outpatient psychotherapy services including injections excluding charges for visits in excess of 52 per year, prescription drugs, and inpatient care.

inpdol: Inpatient expenses: all covered inpatient expenses in a hospital, mental hospital, or nursing home, excluding outpatient care and renal dialysis.

meddol: Medical expenses: all covered inpatient and outpatient services, including drugs, supplies, and inpatient costs of newborns excluding dental care and outpatient psychotherapy.

totadm: Hospital admissions: annual number of covered hospitalizations.

inpmis: Incomplete Hospital Records: missing inpatient records.

mentvis: Psychotherapy visits: indicates the annual number of outpatient visits for psychotherapy. It includes billed visits only. The limit was 52 covered visits per person per year. The count includes an initial visit to a psychiatrist or psychologist.

mdvis: Face-to-Face visits to physicians: annual covered outpatient visits with physician providers (excludes dental, psychotherapy, and radiology/anesthesiology/pathology-only visits).

notmdvis: Face-to-Face visits to nonphysicians: annual covered outpatient visits with nonphysician providers such as speech and physical therapists, chiropractors, podiatrists, acupuncturists, Christian Science etc. (excludes dental, healers, psychotherapy, and radiology/anesthesiology/pathology-only visits).

num: Family size.

mhi: Mental health index.

disea: Number of chronic diseases.

physlm: Physical limitations.

ghindx: General health index.

mdeoff: Maximum expenditure offer.

pioff: Participation incentive payment.

child: 1 if age is less than 18 years.

fchild: female * child.

lfam: log of num (family size).

lpi: log of pioff (participation incentive payment).

idp: 1 if individual deductible plan.

logc: log(coins+1).

fmde: 0 if idp=1, ln(max(1,mdeoff/(0.01*coins))) otherwise.

hlthg: 1 if self-rated health is good -- baseline is excellent self-rated health.

hlthf: 1 if self-rated health is fair -- baseline is excellent self-rated health.

hlthp: 1 if self-rated health is poor -- baseline is excellent self-rated health.

xghindx: ghindx (general healt index) with imputations of missing values.

linc: log of income (family income).

lnum: log of num (family size).

lnmeddol: log of meddol (medical expenses).

binexp: 1 if meddol > 0.

RandHIE

Format

An object of class data.frame with 20190 rows and 45 columns.

Source

http://cameron.econ.ucdavis.edu/mmabook/mmadata.html

References

A Colin Cameron, Pravin K Trivedi (2005). Microeconometrics: methods and applications. Cambridge university press. Mikhail Zhelonkin, Marc G. Genton, Elvezio Ronchetti (2019). ssmrob: Robust Estimation and Inference in Sample Selection Models. R package version 0.7, https://CRAN.R-project.org/package=ssmrob. Ott Toomet, Arne Henningsen (2008). “Sample Selection Models in R: Package sampleSelection.” Journal of Statistical Software, 27(7). https://www.jstatsoft.org/article/view/v027i07. Wikipedia contributors (2019). “RAND Health Insurance Experiment --- Wikipedia, The Free Encyclopedia.” https://en.wikipedia.org/w/index.php?title=RAND_Health_Insurance_Experiment&oldid=909771077. [Online; accessed 9-September-2019].

Examples

##Cameron and Trivedi (2005): Section 16.6
data(RandHIE)
subsample <- RandHIE$year == 2 & !is.na( RandHIE$educdec )
selectEq <- binexp ~ logc + idp + lpi + fmde + physlm + disea +
  hlthg + hlthf + hlthp + linc + lfam + educdec + xage + female +
  child + fchild + black
  outcomeEq <- lnmeddol ~ logc + idp + lpi + fmde + physlm + disea +
  hlthg + hlthf + hlthp + linc + lfam + educdec + xage + female +
  child + fchild + black
  cameron <- HeckmanCL(selectEq, outcomeEq, data = RandHIE[subsample, ])
  summary(cameron)
#> 
#> --------------------------------------------------------------
#>           Classic Heckman Model (Package: ssmodels)           
#> --------------------------------------------------------------
#> --------------------------------------------------------------
#> Maximum Likelihood estimation 
#> optim function with method BFGS-iterations numbers: 48 
#> Log-Likelihood: -10170.11 
#> AIC: 20416.22 BIC: 20668 
#> Number of observations: ( 1293 censored and 4281 observed ) 
#> 38 free parameters ( df= 5536 ) 
#> --------------------------------------------------------------
#> Probit selection equation:
#>               Estimate Std. Error t value Pr(>|t|)    
#> (Intercept) -0.2141574  0.1842053  -1.163 0.245041    
#> logc        -0.1068216  0.0264766  -4.035 5.54e-05 ***
#> idp         -0.1087302  0.0509938  -2.132 0.033032 *  
#> lpi          0.0294717  0.0086215   3.418 0.000634 ***
#> fmde         0.0007550  0.0158738   0.048 0.962067    
#> physlm       0.2847934  0.0722657   3.941 8.22e-05 ***
#> disea        0.0210810  0.0034966   6.029 1.76e-09 ***
#> hlthg        0.0576813  0.0427990   1.348 0.177802    
#> hlthf        0.2237060  0.0814546   2.746 0.006045 ** 
#> hlthp        0.7984070  0.2048093   3.898 9.80e-05 ***
#> linc         0.0553151  0.0166180   3.329 0.000878 ***
#> lfam        -0.0312118  0.0402989  -0.775 0.438663    
#> educdec      0.0314992  0.0074982   4.201 2.70e-05 ***
#> xage        -0.0006073  0.0021059  -0.288 0.773082    
#> female       0.4093225  0.0532548   7.686 1.78e-14 ***
#> child        0.0530837  0.0786201   0.675 0.499582    
#> fchild      -0.3953677  0.0783813  -5.044 4.70e-07 ***
#> black       -0.5831094  0.0520541 -11.202  < 2e-16 ***
#> --------------------------------------------------------------
#> Outcome equation:
#>              Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  2.107659   0.244233   8.630  < 2e-16 ***
#> logc        -0.076039   0.033745  -2.253  0.02428 *  
#> idp         -0.149670   0.066137  -2.263  0.02367 *  
#> lpi          0.014921   0.010501   1.421  0.15543    
#> fmde        -0.023506   0.019474  -1.207  0.22747    
#> physlm       0.354818   0.075541   4.697 2.70e-06 ***
#> disea        0.028648   0.003797   7.545 5.27e-14 ***
#> hlthg        0.155915   0.052177   2.988  0.00282 ** 
#> hlthf        0.445133   0.095525   4.660 3.24e-06 ***
#> hlthp        0.998558   0.187876   5.315 1.11e-07 ***
#> linc         0.121412   0.023084   5.260 1.50e-07 ***
#> lfam        -0.158296   0.049746  -3.182  0.00147 ** 
#> educdec      0.017595   0.009018   1.951  0.05110 .  
#> xage         0.005738   0.002443   2.349  0.01885 *  
#> female       0.550347   0.063330   8.690  < 2e-16 ***
#> child       -0.197684   0.097398  -2.030  0.04244 *  
#> fchild      -0.565327   0.097528  -5.797 7.14e-09 ***
#> black       -0.535856   0.074917  -7.153 9.61e-13 ***
#> --------------------------------------------------------------
#> Error terms:
#>       Estimate Std. Error t value Pr(>|t|)    
#> sigma  1.57003    0.02782   56.43   <2e-16 ***
#> rho    0.73558    0.03378   21.77   <2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> --------------------------------------------------------------